start toward a modular approach for JavaScript; see #37
