+++ .Rbuildignore
@@ -4,3 +4,4 @@
 ^CONDUCT\.md$
 ^\.travis\.yml$
 ^docs$
+^\./javascript$
