+++ inst/htmlwidgets/sunburst.js
@@ -1,604 +1,1 @@
-HTMLWidgets.widget({
-
-  name: 'sunburst',
-
-  type: 'output',
-
-  factory: function(el, width, height) {
-
-    var instance = {};
-
-    instance.chart = {};
-
-    var dispatch = d3.dispatch("mouseover","mouseleave","click");
-
-    // Copies a variable number of methods from source to target.
-    d3.rebind = function(target, source) {
-      var i = 1, n = arguments.length, method;
-      while (++i < n) target[method = arguments[i]] = d3_rebind(target, source, source[method]);
-      return target;
-    };
-
-    // Method is assumed to be a standard D3 getter-setter:
-    // If passed with no arguments, gets the value.
-    // If passed with arguments, sets the value and returns the target.
-    function d3_rebind(target, source, method) {
-      return function() {
-        var value = method.apply(source, arguments);
-        return value === source ? target : value;
-      };
-    }
-
-    d3.rebind(instance.chart, dispatch, 'on');
-
-    var draw = function(el, instance) {
-
-      // would be much nicer to implement transitions/animation
-      // remove previous in case of Shiny/dynamic
-      d3.select(el).select(".sunburst-chart svg").remove();
-
-      var x = instance.x;
-      var json = instance.json;
-      var chart = instance.chart;
-
-      // Dimensions of sunburst
-      var width = el.getBoundingClientRect().width - (x.options.legend.w ? x.options.legend.w : 75);
-      var height = el.getBoundingClientRect().height - 70;
-      var radius = Math.min(width, height) / 2;
-
-      d3.select(el).select(".sunburst-chart").append("svg")
-        .style("width", width + "px") // shouldnt have to do this
-        .style("height", height + "px"); // shouldnt have to do this
-
-      // Breadcrumb dimensions: width, height, spacing, width of tip/tail.
-      //  these will be the defaults
-      var b = {
-        w: 0, h: 30, s: 3, t: 10
-      };
-      //  if breadcrumb is provided in the option, we will overwrite
-      //   with what is provided
-      Object.keys(x.options.breadcrumb).map(function(ky){
-        b[ky] = x.options.breadcrumb[ky];
-      });
-  /*
-      // Mapping of step names to colors.
-      var colors = {
-        "home": "#5687d1",
-        "product": "#7b615c",
-        "search": "#de783b",
-        "account": "#6ab975",
-        "other": "#a173d1",
-        "end": "#bbbbbb"
-      };
-  */
-
-      var colors = d3.scaleOrdinal(d3.schemeCategory20);
-
-      if(x.options.colors !== null){
-        // if an array then we assume the colors
-        //  represent an array of hexadecimal colors to be used
-        if(Array.isArray(x.options.colors)) {
-          try{
-            colors.range(x.options.colors)
-          } catch(e) {
-
-          }
-        }
-
-        // if an object with range then we assume
-        //  that this is an array of colors to be used as range
-        if(x.options.colors.range){
-          try{
-            colors.range(x.options.colors.range)
-          } catch(e) {
-
-          }
-        }
-
-        // if an object with domain then we assume
-        //  that this is an array of colors to be used as domain
-        //  for more precise control of the colors assigned
-        if(x.options.colors.domain){
-          try{
-            colors.domain(x.options.colors.domain);
-          } catch(e) {
-
-          }
-        }
-
-        // if a function then set to the function
-        if(typeof(x.options.colors) === "function") {
-          colors = x.options.colors;
-        }
-      }
-      // Total size of all segments; we set this later, after loading the data.
-      var totalSize = 0;
-
-      var vis = d3.select(el).select(".sunburst-chart").select("svg")
-          .append("g")
-          .attr("id", el.id + "-container")
-          .attr("transform", "translate(" + width / 2 + "," + height / 2 + ")");
-
-      var partition = d3.partition()
-          .size([2 * Math.PI, radius * radius]);
-
-      var arc = d3.arc()
-          .startAngle(function(d) { return d.x0; })
-          .endAngle(function(d) { return d.x1; })
-          .innerRadius(function(d) { return Math.sqrt(d.y0); })
-          .outerRadius(function(d) { return Math.sqrt(d.y1); });
-
-      createVisualization(json);
-
-      // set up a container for tasks to perform after completion
-      //  one example would be add callbacks for event handling
-      //  styling
-      if (!(typeof x.tasks === "undefined") ){
-        if ( (typeof x.tasks.length === "undefined") ||
-         (typeof x.tasks === "function" ) ) {
-           // handle a function not enclosed in array
-           // should be able to remove once using jsonlite
-           x.tasks = [x.tasks];
-        }
-        x.tasks.map(function(t){
-          // for each tasks call the task with el supplied as `this`
-          t.call({el:el,x:x,instance:instance});
-        });
-      }
-
-      // Main function to draw and set up the visualization, once we have the data.
-      function createVisualization(json) {
-
-        // Basic setup of page elements.
-        initializeBreadcrumbTrail();
-
-        // Bounding circle underneath the sunburst, to make it easier to detect
-        // when the mouse leaves the parent g.
-        vis.append("circle")
-            .attr("r", radius)
-            .style("opacity", 0);
-
-        // Turn the data into a d3 hierarchy and calculate the sums.
-        var root = d3.hierarchy(json)
-            .sum(function(d) { return d[x.options.valueField || "size"]; });
-
-        // check for sort function
-        if(x.options.sortFunction){
-          root.sort(x.options.sortFunction);
-        }
-
-        // For efficiency, filter nodes to keep only those large enough to see.
-        var nodes = partition(root).descendants()
-            .filter(function(d) {
-                return (d.x1 - d.x0 > 0.005); // 0.005 radians = 0.29 degrees
-            });
-
-        var path = vis.data([json]).selectAll("path")
-            .data(nodes)
-            .enter().append("path")
-            .attr("display", function(d) { return d.depth ? null : "none"; })
-            .attr("d", arc)
-            .attr("fill-rule", "evenodd")
-            .style("fill", function(d) { return colors.call(this, d.data.name); })
-            .style("opacity", 1)
-            .on("mouseover", mouseover)
-            .on("click", click);
-
-        // Add the mouseleave handler to the bounding circle.
-        d3.select(el).select("#"+ el.id + "-container").on("mouseleave", mouseleave);
-
-        // Get total size of the tree = value of root node from partition.
-        totalSize = path.datum().value;
-
-        drawLegend(nodes);
-        d3.select(el).select(".sunburst-togglelegend").on("click", toggleLegend);
-
-       }
-
-      // Fade all but the current sequence, and show it in the breadcrumb trail.
-      function mouseover(d) {
-
-        var percentage = (100 * d.value / totalSize).toPrecision(3);
-        var percentageString = percentage + "%";
-        if (percentage < 0.1) {
-          percentageString = "< 0.1%";
-        }
-
-        var countString = [
-            '<span style = "font-size:.7em">',
-            d3.format("1.2s")(d.value) + ' of ' + d3.format("1.2s")(totalSize),
-            '</span>'
-          ].join('')
-
-        var explanationString = "";
-        if(x.options.percent && x.options.count){
-          explanationString = percentageString + '<br/>' + countString;
-        } else if(x.options.percent){
-          explanationString = percentageString;
-        } else if(x.options.count){
-          explanationString = countString;
-        }
-
-        //if explanation defined in R then use this instead
-        if(x.options.explanation !== null){
-          explanationString = x.options.explanation.bind(totalSize)(d);
-        }
-
-
-        d3.select(el).selectAll(".sunburst-explanation")
-            .style("visibility", "")
-            .style("top",((height - 70)/2) + "px")
-            .style("width",width + "px")
-            .html(explanationString);
-
-        var sequenceArray = d.ancestors().reverse();
-        sequenceArray.shift(); // remove root node from the array
-
-        chart._selection = sequenceArray.map(
-          function(d){return d.data.name}
-        );
-        dispatch.call("mouseover", chart._selection);
-
-        updateBreadcrumbs(sequenceArray, percentageString);
-
-        // Fade all the segments.
-        d3.select(el).selectAll("path")
-            .style("opacity", 0.3);
-
-        // Then highlight only those that are an ancestor of the current segment.
-        vis.selectAll("path")
-            .filter(function(node) {
-                      return (sequenceArray.indexOf(node) >= 0);
-                    })
-            .style("opacity", 1);
-      }
-
-      // Restore everything to full opacity when moving off the visualization.
-      function mouseleave(d) {
-
-        dispatch.call("mouseleave", chart._selection);
-        chart._selection = [];
-
-        // Hide the breadcrumb trail
-        d3.select(el).select("#" + el.id + "-trail")
-            .style("visibility", "hidden");
-
-        // Deactivate all segments during transition.
-        d3.select(el).selectAll("path").on("mouseover", null);
-
-        // Transition each segment to full opacity and then reactivate it.
-        d3.select(el).selectAll("path")
-            .transition()
-            .duration(1000)
-            .style("opacity", 1)
-            .on("end", function() {
-              d3.select(this).on("mouseover", mouseover);
-            });
-
-        d3.select(el).selectAll(".sunburst-explanation")
-            .style("visibility", "hidden");
-      }
-
-      function click(d,i) {
-        var sequenceArray = d.ancestors().reverse();
-        sequenceArray.shift(); // remove root node from the array
-
-        dispatch.call("click", sequenceArray.map(
-          function(d){return d.data.name}
-        ));
-      }
-
-      function initializeBreadcrumbTrail() {
-        // Add the svg area.
-        var trail = d3.select(el).select(".sunburst-sequence").append("svg")
-            .attr("width", width)
-            //.attr("height", 50)
-            .attr("id", el.id + "-trail");
-        // Add the label at the end, for the percentage.
-        trail.append("text")
-          .attr("id", el.id + "-endlabel")
-          .style("fill", "#000");
-      }
-
-      // Generate a string that describes the points of a breadcrumb polygon.
-      function breadcrumbPoints(d, i) {
-        var points = [];
-        points.push("0,0");
-        if (b.w <= 0) {
-          // calculate breadcrumb width based on string length
-          points.push(d.string_length + ",0");
-          points.push(d.string_length + b.t + "," + (b.h / 2));
-          points.push(d.string_length + "," + b.h);
-        } else {
-          points.push(b.w + ",0");
-          points.push(b.w + b.t + "," + (b.h / 2));
-          points.push(b.w + "," + b.h);
-        }
-        points.push("0," + b.h);
-
-        if (i > 0) { // Leftmost breadcrumb; don't include 6th vertex.
-          points.push(b.t + "," + (b.h / 2));
-        }
-        return points.join(" ");
-      }
-
-      // Update the breadcrumb trail to show the current sequence and percentage.
-      function updateBreadcrumbs(nodeArray, percentageString) {
-
-        // Data join; key function combines name and depth (= position in sequence).
-        var g = d3.select(el).select("#" + el.id + "-trail")
-            .selectAll("g")
-            .data(nodeArray, function(d) { return d.data.name + d.depth; });
-
-        // Add breadcrumb and label for entering nodes.
-        var entering = g.enter().append("g");
-
-
-        if (b.w <= 0) {
-          // Create a node array that contains all the breadcrumb widths
-          // Calculate positions of breadcrumbs based on string lengths
-          var curr_breadcrumb_x = 0;
-          nodeArray[0].breadcrumb_x = 0;
-          nodeArray[0].breadcrumb_h = 0;
-
-          entering.append("polygon")
-              .style("z-index",function(d,i) { return(999-i); })
-              .style("fill", function(d) { return colors.call(this, d.data.name); });
-
-          entering.append("text")
-              .attr("x", b.t + 2)
-              .attr("y", b.h / 2)
-              .attr("dy", "0.35em")
-              .attr("text-anchor", "left")
-              .text(function(d) { return d.data.name; });
-
-          // Remove exiting nodes.
-          g.exit().remove();
-
-          // loop through each g element
-          //  calculate string length
-          //  draw the breadcrumb polygon
-          //  and determine if breadcrumb should be wrapped to next row
-          entering.merge(g).each(function(d,k){
-            var crumbg = d3.select(this);
-            var my_string_length = crumbg.select("text").node().getBoundingClientRect().width;
-            nodeArray[k].string_length = my_string_length + 12;
-            crumbg.select("polygon").attr("points", function(d){
-              return breadcrumbPoints(d, k);
-            });
-            var my_g_length = crumbg.node().getBoundingClientRect().width;
-            curr_breadcrumb_x += k===0 ? 0 : nodeArray[k-1].string_length + b.s;
-            nodeArray[k].breadcrumb_h = k===0 ? 0 : nodeArray[k-1].breadcrumb_h;
-
-            if (curr_breadcrumb_x + my_g_length > width*0.99) {
-              nodeArray[k].breadcrumb_h += b.h;  // got to next line
-              curr_breadcrumb_x = b.t + b.s;     // restart counter
-            }
-            nodeArray[k].breadcrumb_x = curr_breadcrumb_x;
-          });
-
-
-          // Set position for entering and updating nodes.
-          entering.merge(g).attr("transform", function(d, i) {
-            return "translate(" + d.breadcrumb_x + ", "+d.breadcrumb_h+")";
-          });
-
-
-          // Now move and update the percentage at the end.
-          d3.select(el).select("#" + el.id + "-trail").select("#" + el.id + "-endlabel")
-              .attr("x", function(d){
-                var bend = d3.select(this);
-                var curr_breadcrumb_x = nodeArray[nodeArray.length-1].breadcrumb_x +  nodeArray[nodeArray.length-1].string_length + b.t + b.s;
-                var my_g_length = bend.node().getBoundingClientRect().width;
-
-                var curr_breadcrumb_h = nodeArray[nodeArray.length-1].breadcrumb_h + b.h/2;
-                if (curr_breadcrumb_x + my_g_length > width*0.99) {
-                  curr_breadcrumb_h += b.h + b.h/2;
-                  curr_breadcrumb_x = b.t + b.s;     // restart counter
-                }
-                bend.datum({
-                  "breadcrumb_x": curr_breadcrumb_x,
-                  "breadcrumb_h": curr_breadcrumb_h
-                });
-                return curr_breadcrumb_x;
-              })
-              .attr("y", function(d){return d.breadcrumb_h})
-              .attr("dy", "0.35em")
-              .attr("text-anchor", "start")
-              .text(percentageString);
-
-
-        } else {
-          entering.append("polygon")
-              .attr("points", breadcrumbPoints)
-              .style("fill", function(d) { return colors.call(this, d.data.name); });
-
-          entering.append("text")
-              .attr("x", (b.w + b.t) / 2)
-              .attr("y", b.h / 2)
-              .attr("dy", "0.35em")
-              .attr("text-anchor", "middle")
-              .text(function(d) { return d.data.name; });
-
-          // Set position for entering and updating nodes.
-          entering.merge(g).attr("transform", function(d, i) {
-            return "translate(" + i * (b.w + b.s) + ", 0)";
-          });
-
-          // Remove exiting nodes.
-          g.exit().remove();
-
-          // Now move and update the percentage at the end.
-          d3.select(el).select("#" + el.id + "-trail").select("#" + el.id + "-endlabel")
-              .attr("x", (nodeArray.length + 0.5) * (b.w + b.s))
-              .attr("y", b.h / 2)
-              .attr("dy", "0.35em")
-              .attr("text-anchor", "middle")
-              .text(percentageString);
-
-        }
-
-        // Make the breadcrumb trail visible, if it's hidden.
-        d3.select(el).select("#" + el.id + "-trail")
-            .style("visibility", "");
-
-      }
-
-      function drawLegend(nodes) {
-
-        // Dimensions of legend item: width, height, spacing, radius of rounded rect.
-        var li = {
-          w: 75, h: 30, s: 3, r: 3
-        };
-
-        //  if legend is provided in the option, we will overwrite
-        //   with what is provided
-        Object.keys(x.options.legend).map(function(ky){
-          li[ky] = x.options.legend[ky];
-        });
-
-        // remove if already drawn
-        d3.select(el).select(".sunburst-legend svg").remove();
-
-        // get labels from node names
-        var labels = d3.nest()
-          .key(function(d) {return d.data.name})
-          .entries(
-            nodes.sort(
-              function(a,b) {return d3.ascending(a.depth,b.depth)}
-            )
-          )
-          .map(function(d) {
-            return d.values[0];
-          })
-          .filter(function(d) {
-            return d.data.name !== "root";
-          });
-
-        var legend = d3.select(el).select(".sunburst-legend").append("svg")
-            .attr("width", li.w)
-            .attr("height", labels.length * (li.h + li.s));
-
-        var g = legend.selectAll("g")
-            .data( function(){
-              if(x.options.legendOrder !== null){
-                return x.options.legendOrder.map(function(d) {
-                  return labels.filter(function(dd) {return dd.data.name === d })[0];
-                });
-              } else {
-                // get sorted by top level
-                return labels;
-              }
-            })
-            .enter().append("g")
-            .attr("transform", function(d, i) {
-                    return "translate(0," + i * (li.h + li.s) + ")";
-                 });
-
-        g.append("rect")
-            .attr("rx", li.r)
-            .attr("ry", li.r)
-            .attr("width", li.w)
-            .attr("height", li.h)
-            .style("fill", function(d) { return colors.call(this, d.data.name); });
-
-        g.append("text")
-            .attr("x", li.w / 2)
-            .attr("y", li.h / 2)
-            .attr("dy", "0.35em")
-            .attr("text-anchor", "middle")
-            .text(function(d) { return d.data.name; });
-      }
-
-      function toggleLegend() {
-        var legend = d3.select(el).select(".sunburst-legend")
-        if (legend.style("visibility") == "hidden") {
-          legend.style("visibility", "");
-        } else {
-          legend.style("visibility", "hidden");
-        }
-      }
-    };
-
-    // Take a 2-column CSV and transform it into a hierarchical structure suitable
-    // for a partition layout. The first column is a sequence of step names, from
-    // root to leaf, separated by hyphens. The second column is a count of how
-    // often that sequence occurred.
-    function buildHierarchy(csv) {
-      var root = {"name": "root", "children": []};
-      for (var i = 0; i < csv.length; i++) {
-        var sequence = csv[i][0];
-        var size = +csv[i][1];
-        if (isNaN(size)) { // e.g. if this is a header row
-          continue;
-        }
-        var parts = sequence.split("-");
-        var currentNode = root;
-        for (var j = 0; j < parts.length; j++) {
-          var children = currentNode["children"];
-          var nodeName = parts[j];
-          var childNode;
-          if (j + 1 < parts.length) {
-       // Not yet at the end of the sequence; move down the tree.
-     	var foundChild = false;
-     	for (var k = 0; k < children.length; k++) {
-     	  if (children[k]["name"] == nodeName) {
-     	    childNode = children[k];
-     	    foundChild = true;
-     	    break;
-     	  }
-     	}
-      // If we don't already have a child node for this branch, create it.
-     	if (!foundChild) {
-     	  childNode = {"name": nodeName, "children": []};
-     	  children.push(childNode);
-     	}
-     	currentNode = childNode;
-          } else {
-     	// Reached the end of the sequence; create a leaf node.
-     	childNode = {"name": nodeName, "size": size};
-     	children.push(childNode);
-          }
-        }
-      }
-      return root;
-
-    };
-
-    return {
-
-      renderValue: function(x) {
-
-        instance.x = x;
-
-        // x.data should be a data.frame in R so an Javascript Object of Objects
-        //     but buildHierarchy expects an Array of Arrays
-        //     so use d3.zip and apply to do this
-        var json = [];
-        if(x.csvdata !== null){
-          json = buildHierarchy(
-            d3.zip.apply(
-              null,
-              Object.keys(x.csvdata).map(function(ky){return x.csvdata[ky]})
-            )
-          );
-        } else {
-          json = x.jsondata
-        }
-        instance.json = json;
-
-        draw(el, instance);
-
-      },
-
-      resize: function(width, height) {
-
-        draw(el, instance);
-
-      },
-
-      instance: instance
-
-    };
-  }
-});
+!function t(n,e,r){function i(a,u){if(!e[a]){if(!n[a]){var c="function"==typeof require&&require;if(!u&&c)return c(a,!0);if(o)return o(a,!0);var s=new Error("Cannot find module '"+a+"'");throw s.code="MODULE_NOT_FOUND",s}var f=e[a]={exports:{}};n[a][0].call(f.exports,function(t){var e=n[a][1][t];return i(e||t)},f,f.exports,t,n,e,r)}return e[a].exports}for(var o="function"==typeof require&&require,a=0;a<r.length;a++)i(r[a]);return i}({1:[function(t,n,e){!function(r,i){"object"==typeof e&&void 0!==n?i(t("d3-dispatch"),t("d3-array"),t("d3-format"),t("d3-selection"),t("d3-scale"),t("d3-shape"),t("d3-hierarchy"),t("d3-collection"),t("d3-transition")):"function"==typeof define&&define.amd?define(["d3-dispatch","d3-array","d3-format","d3-selection","d3-scale","d3-shape","d3-hierarchy","d3-collection","d3-transition"],i):i(r.d3,r.d3,r.d3,r.d3,r.d3,r.d3,r.d3,r.d3,r.d3)}(this,function(t,n,e,r,i,o,a,u,c){"use strict";var s=function(t,c,s){function f(n){var i=(100*n.value/A).toPrecision(3),o=i+"%";i<.1&&(o="< 0.1%");var a=['<span style = "font-size:.7em">',e.format("1.2s")(n.value)+" of "+e.format("1.2s")(A),"</span>"].join(""),u="";g.options.percent&&g.options.count?u=o+"<br/>"+a:g.options.percent?u=o:g.options.count&&(u=a),null!==g.options.explanation&&(u=g.options.explanation.bind(A)(n)),r.select(t).selectAll(".sunburst-explanation").style("visibility","").style("top",(w-70)/2+"px").style("width",x+"px").html(u);var c=n.ancestors().reverse();c.shift(),b._selection=c.map(function(t){return t.data.name}),s.call("mouseover",b._selection),_(c,o),r.select(t).selectAll("path").style("opacity",.3),N.selectAll("path").filter(function(t){return c.indexOf(t)>=0}).style("opacity",1)}function l(n){s.call("mouseleave",b._selection),b._selection=[],r.select(t).select("#"+t.id+"-trail").style("visibility","hidden"),r.select(t).selectAll("path").on("mouseover",null),r.select(t).selectAll("path").transition().duration(1e3).style("opacity",1).on("end",function(){r.select(this).on("mouseover",f)}),r.select(t).selectAll(".sunburst-explanation").style("visibility","hidden")}function h(t,n){var e=t.ancestors().reverse();e.shift(),s.call("click",e.map(function(t){return t.data.name}))}function d(){r.select(t).select(".sunburst-sequence").append("svg").attr("width",x).attr("id",t.id+"-trail").append("text").attr("id",t.id+"-endlabel").style("fill","#000")}function p(t,n){var e=[];return e.push("0,0"),T.w<=0?(e.push(t.string_length+",0"),e.push(t.string_length+T.t+","+T.h/2),e.push(t.string_length+","+T.h)):(e.push(T.w+",0"),e.push(T.w+T.t+","+T.h/2),e.push(T.w+","+T.h)),e.push("0,"+T.h),n>0&&e.push(T.t+","+T.h/2),e.join(" ")}function _(n,e){var i=r.select(t).select("#"+t.id+"-trail").selectAll("g").data(n,function(t){return t.data.name+t.depth}),o=i.enter().append("g");if(T.w<=0){var a=0;n[0].breadcrumb_x=0,n[0].breadcrumb_h=0,o.append("polygon").style("z-index",function(t,n){return 999-n}).style("fill",function(t){return k.call(this,t.data.name)}),o.append("text").attr("x",T.t+2).attr("y",T.h/2).attr("dy","0.35em").attr("text-anchor","left").text(function(t){return t.data.name}),i.exit().remove(),o.merge(i).each(function(t,e){var i=r.select(this),o=i.select("text").node().getBoundingClientRect().width;n[e].string_length=o+12,i.select("polygon").attr("points",function(t){return p(t,e)});var u=i.node().getBoundingClientRect().width;a+=0===e?0:n[e-1].string_length+T.s,n[e].breadcrumb_h=0===e?0:n[e-1].breadcrumb_h,a+u>.99*x&&(n[e].breadcrumb_h+=T.h,a=T.t+T.s),n[e].breadcrumb_x=a}),o.merge(i).attr("transform",function(t,n){return"translate("+t.breadcrumb_x+", "+t.breadcrumb_h+")"}),r.select(t).select("#"+t.id+"-trail").select("#"+t.id+"-endlabel").attr("x",function(t){var e=r.select(this),i=n[n.length-1].breadcrumb_x+n[n.length-1].string_length+T.t+T.s,o=e.node().getBoundingClientRect().width,a=n[n.length-1].breadcrumb_h+T.h/2;return i+o>.99*x&&(a+=T.h+T.h/2,i=T.t+T.s),e.datum({breadcrumb_x:i,breadcrumb_h:a}),i}).attr("y",function(t){return t.breadcrumb_h}).attr("dy","0.35em").attr("text-anchor","start").text(e)}else o.append("polygon").attr("points",p).style("fill",function(t){return k.call(this,t.data.name)}),o.append("text").attr("x",(T.w+T.t)/2).attr("y",T.h/2).attr("dy","0.35em").attr("text-anchor","middle").text(function(t){return t.data.name}),o.merge(i).attr("transform",function(t,n){return"translate("+n*(T.w+T.s)+", 0)"}),i.exit().remove(),r.select(t).select("#"+t.id+"-trail").select("#"+t.id+"-endlabel").attr("x",(n.length+.5)*(T.w+T.s)).attr("y",T.h/2).attr("dy","0.35em").attr("text-anchor","middle").text(e);r.select(t).select("#"+t.id+"-trail").style("visibility","")}function y(e){var i={w:75,h:30,s:3,r:3};Object.keys(g.options.legend).map(function(t){i[t]=g.options.legend[t]}),r.select(t).select(".sunburst-legend svg").remove();var o=u.nest().key(function(t){return t.data.name}).entries(e.sort(function(t,e){return n.ascending(t.depth,e.depth)})).map(function(t){return t.values[0]}).filter(function(t){return"root"!==t.data.name}),a=r.select(t).select(".sunburst-legend").append("svg").attr("width",i.w).attr("height",o.length*(i.h+i.s)).selectAll("g").data(function(){return null!==g.options.legendOrder?g.options.legendOrder.map(function(t){return o.filter(function(n){return n.data.name===t})[0]}):o}).enter().append("g").attr("transform",function(t,n){return"translate(0,"+n*(i.h+i.s)+")"});a.append("rect").attr("rx",i.r).attr("ry",i.r).attr("width",i.w).attr("height",i.h).style("fill",function(t){return k.call(this,t.data.name)}),a.append("text").attr("x",i.w/2).attr("y",i.h/2).attr("dy","0.35em").attr("text-anchor","middle").text(function(t){return t.data.name})}function v(){var n=r.select(t).select(".sunburst-legend");"hidden"==n.style("visibility")?n.style("visibility",""):n.style("visibility","hidden")}r.select(t).select(".sunburst-chart svg").remove();var g=c.x,m=c.json,b=c.chart,x=t.getBoundingClientRect().width-(g.options.legend.w?g.options.legend.w:75),w=t.getBoundingClientRect().height-70,M=Math.min(x,w)/2;r.select(t).select(".sunburst-chart").append("svg").style("width",x+"px").style("height",w+"px");var T={w:0,h:30,s:3,t:10};Object.keys(g.options.breadcrumb).map(function(t){T[t]=g.options.breadcrumb[t]});var k=i.scaleOrdinal(i.schemeCategory20);if(null!==g.options.colors){if(Array.isArray(g.options.colors))try{k.range(g.options.colors)}catch(t){}if(g.options.colors.range)try{k.range(g.options.colors.range)}catch(t){}if(g.options.colors.domain)try{k.domain(g.options.colors.domain)}catch(t){}"function"==typeof g.options.colors&&(k=g.options.colors)}var A=0,N=r.select(t).select(".sunburst-chart").select("svg").append("g").attr("id",t.id+"-container").attr("transform","translate("+x/2+","+w/2+")"),S=a.partition().size([2*Math.PI,M*M]),C=o.arc().startAngle(function(t){return t.x0}).endAngle(function(t){return t.x1}).innerRadius(function(t){return Math.sqrt(t.y0)}).outerRadius(function(t){return Math.sqrt(t.y1)});!function(n){d(),N.append("circle").attr("r",M).style("opacity",0);var e=a.hierarchy(n).sum(function(t){return t[g.options.valueField||"size"]});g.options.sortFunction&&e.sort(g.options.sortFunction);var i=S(e).descendants().filter(function(t){return t.x1-t.x0>.005}),o=N.data([n]).selectAll("path").data(i).enter().append("path").attr("display",function(t){return t.depth?null:"none"}).attr("d",C).attr("fill-rule","evenodd").style("fill",function(t){return k.call(this,t.data.name)}).style("opacity",1).on("mouseover",f).on("click",h);r.select(t).select("#"+t.id+"-container").on("mouseleave",l),A=o.datum().value,y(i),r.select(t).select(".sunburst-togglelegend").on("click",v)}(m),void 0!==g.tasks&&(void 0!==g.tasks.length&&"function"!=typeof g.tasks||(g.tasks=[g.tasks]),g.tasks.map(function(n){n.call({el:t,x:g,instance:c})}))};window.HTMLWidgets.widget({name:"sunburst",type:"output",factory:function(e,r,i){function o(t){for(var n={name:"root",children:[]},e=0;e<t.length;e++){var r=t[e][0],i=+t[e][1];if(!isNaN(i))for(var o=r.split("-"),a=n,u=0;u<o.length;u++){var c,s=a.children,f=o[u];if(u+1<o.length){for(var l=!1,h=0;h<s.length;h++)if(s[h].name==f){c=s[h],l=!0;break}l||(c={name:f,children:[]},s.push(c)),a=c}else c={name:f,size:i},s.push(c)}}return n}var a={};a.chart={};var u=t.dispatch("mouseover","mouseleave","click");return a.chart.on=u.on,{renderValue:function(t){a.x=t;var r=[];r=null!==t.csvdata?o(n.zip.apply(null,Object.keys(t.csvdata).map(function(n){return t.csvdata[n]}))):t.jsondata,a.json=r,s(e,a,u)},resize:function(t,n){s(e,a,u)},instance:a}}})})},{"d3-array":2,"d3-collection":3,"d3-dispatch":5,"d3-format":7,"d3-hierarchy":8,"d3-scale":11,"d3-selection":12,"d3-shape":13,"d3-transition":17}],2:[function(t,n,e){!function(t,r){"object"==typeof e&&void 0!==n?r(e):"function"==typeof define&&define.amd?define(["exports"],r):r(t.d3=t.d3||{})}(this,function(t){"use strict";function n(t){return function(n,e){return a(t(n),e)}}function e(t,n){return[t,n]}function r(t,n,e){var r=(n-t)/Math.max(0,e),i=Math.floor(Math.log(r)/Math.LN10),o=r/Math.pow(10,i);return i>=0?(o>=T?10:o>=k?5:o>=A?2:1)*Math.pow(10,i):-Math.pow(10,-i)/(o>=T?10:o>=k?5:o>=A?2:1)}function i(t,n,e){var r=Math.abs(n-t)/Math.max(0,e),i=Math.pow(10,Math.floor(Math.log(r)/Math.LN10)),o=r/i;return o>=T?i*=10:o>=k?i*=5:o>=A&&(i*=2),n<t?-i:i}function o(t){return t.length}var a=function(t,n){return t<n?-1:t>n?1:t>=n?0:NaN},u=function(t){return 1===t.length&&(t=n(t)),{left:function(n,e,r,i){for(null==r&&(r=0),null==i&&(i=n.length);r<i;){var o=r+i>>>1;t(n[o],e)<0?r=o+1:i=o}return r},right:function(n,e,r,i){for(null==r&&(r=0),null==i&&(i=n.length);r<i;){var o=r+i>>>1;t(n[o],e)>0?i=o:r=o+1}return r}}},c=u(a),s=c.right,f=c.left,l=function(t,n){null==n&&(n=e);for(var r=0,i=t.length-1,o=t[0],a=new Array(i<0?0:i);r<i;)a[r]=n(o,o=t[++r]);return a},h=function(t,n,r){var i,o,a,u,c=t.length,s=n.length,f=new Array(c*s);for(null==r&&(r=e),i=a=0;i<c;++i)for(u=t[i],o=0;o<s;++o,++a)f[a]=r(u,n[o]);return f},d=function(t,n){return n<t?-1:n>t?1:n>=t?0:NaN},p=function(t){return null===t?NaN:+t},_=function(t,n){var e,r,i=t.length,o=0,a=-1,u=0,c=0;if(null==n)for(;++a<i;)isNaN(e=p(t[a]))||(c+=(r=e-u)*(e-(u+=r/++o)));else for(;++a<i;)isNaN(e=p(n(t[a],a,t)))||(c+=(r=e-u)*(e-(u+=r/++o)));if(o>1)return c/(o-1)},y=function(t,n){var e=_(t,n);return e?Math.sqrt(e):e},v=function(t,n){var e,r,i,o=t.length,a=-1;if(null==n){for(;++a<o;)if(null!=(e=t[a])&&e>=e)for(r=i=e;++a<o;)null!=(e=t[a])&&(r>e&&(r=e),i<e&&(i=e))}else for(;++a<o;)if(null!=(e=n(t[a],a,t))&&e>=e)for(r=i=e;++a<o;)null!=(e=n(t[a],a,t))&&(r>e&&(r=e),i<e&&(i=e));return[r,i]},g=Array.prototype,m=g.slice,b=g.map,x=function(t){return function(){return t}},w=function(t){return t},M=function(t,n,e){t=+t,n=+n,e=(i=arguments.length)<2?(n=t,t=0,1):i<3?1:+e;for(var r=-1,i=0|Math.max(0,Math.ceil((n-t)/e)),o=new Array(i);++r<i;)o[r]=t+r*e;return o},T=Math.sqrt(50),k=Math.sqrt(10),A=Math.sqrt(2),N=function(t,n,e){var i,o,a,u=n<t,c=-1;if(u&&(i=t,t=n,n=i),0===(a=r(t,n,e))||!isFinite(a))return[];if(a>0)for(t=Math.ceil(t/a),n=Math.floor(n/a),o=new Array(i=Math.ceil(n-t+1));++c<i;)o[c]=(t+c)*a;else for(t=Math.floor(t*a),n=Math.ceil(n*a),o=new Array(i=Math.ceil(t-n+1));++c<i;)o[c]=(t-c)/a;return u&&o.reverse(),o},S=function(t){return Math.ceil(Math.log(t.length)/Math.LN2)+1},C=function(){function t(t){var o,a,u=t.length,c=new Array(u);for(o=0;o<u;++o)c[o]=n(t[o],o,t);var f=e(c),l=f[0],h=f[1],d=r(c,l,h);Array.isArray(d)||(d=i(l,h,d),d=M(Math.ceil(l/d)*d,Math.floor(h/d)*d,d));for(var p=d.length;d[0]<=l;)d.shift(),--p;for(;d[p-1]>h;)d.pop(),--p;var _,y=new Array(p+1);for(o=0;o<=p;++o)(_=y[o]=[]).x0=o>0?d[o-1]:l,_.x1=o<p?d[o]:h;for(o=0;o<u;++o)l<=(a=c[o])&&a<=h&&y[s(d,a,0,p)].push(t[o]);return y}var n=w,e=v,r=S;return t.value=function(e){return arguments.length?(n="function"==typeof e?e:x(e),t):n},t.domain=function(n){return arguments.length?(e="function"==typeof n?n:x([n[0],n[1]]),t):e},t.thresholds=function(n){return arguments.length?(r="function"==typeof n?n:x(Array.isArray(n)?m.call(n):n),t):r},t},E=function(t,n,e){if(null==e&&(e=p),r=t.length){if((n=+n)<=0||r<2)return+e(t[0],0,t);if(n>=1)return+e(t[r-1],r-1,t);var r,i=(r-1)*n,o=Math.floor(i),a=+e(t[o],o,t);return a+(+e(t[o+1],o+1,t)-a)*(i-o)}},O=function(t,n,e){return t=b.call(t,p).sort(a),Math.ceil((e-n)/(2*(E(t,.75)-E(t,.25))*Math.pow(t.length,-1/3)))},P=function(t,n,e){return Math.ceil((e-n)/(3.5*y(t)*Math.pow(t.length,-1/3)))},D=function(t,n){var e,r,i=t.length,o=-1;if(null==n){for(;++o<i;)if(null!=(e=t[o])&&e>=e)for(r=e;++o<i;)null!=(e=t[o])&&e>r&&(r=e)}else for(;++o<i;)if(null!=(e=n(t[o],o,t))&&e>=e)for(r=e;++o<i;)null!=(e=n(t[o],o,t))&&e>r&&(r=e);return r},q=function(t,n){var e,r=t.length,i=r,o=-1,a=0;if(null==n)for(;++o<r;)isNaN(e=p(t[o]))?--i:a+=e;else for(;++o<r;)isNaN(e=p(n(t[o],o,t)))?--i:a+=e;if(i)return a/i},U=function(t,n){var e,r=t.length,i=-1,o=[];if(null==n)for(;++i<r;)isNaN(e=p(t[i]))||o.push(e);else for(;++i<r;)isNaN(e=p(n(t[i],i,t)))||o.push(e);return E(o.sort(a),.5)},j=function(t){for(var n,e,r,i=t.length,o=-1,a=0;++o<i;)a+=t[o].length;for(e=new Array(a);--i>=0;)for(n=(r=t[i]).length;--n>=0;)e[--a]=r[n];return e},F=function(t,n){var e,r,i=t.length,o=-1;if(null==n){for(;++o<i;)if(null!=(e=t[o])&&e>=e)for(r=e;++o<i;)null!=(e=t[o])&&r>e&&(r=e)}else for(;++o<i;)if(null!=(e=n(t[o],o,t))&&e>=e)for(r=e;++o<i;)null!=(e=n(t[o],o,t))&&r>e&&(r=e);return r},I=function(t,n){for(var e=n.length,r=new Array(e);e--;)r[e]=t[n[e]];return r},z=function(t,n){if(e=t.length){var e,r,i=0,o=0,u=t[o];for(null==n&&(n=a);++i<e;)(n(r=t[i],u)<0||0!==n(u,u))&&(u=r,o=i);return 0===n(u,u)?o:void 0}},Y=function(t,n,e){for(var r,i,o=(null==e?t.length:e)-(n=null==n?0:+n);o;)i=Math.random()*o--|0,r=t[o+n],t[o+n]=t[i+n],t[i+n]=r;return t},L=function(t,n){var e,r=t.length,i=-1,o=0;if(null==n)for(;++i<r;)(e=+t[i])&&(o+=e);else for(;++i<r;)(e=+n(t[i],i,t))&&(o+=e);return o},B=function(t){if(!(i=t.length))return[];for(var n=-1,e=F(t,o),r=new Array(e);++n<e;)for(var i,a=-1,u=r[n]=new Array(i);++a<i;)u[a]=t[a][n];return r},R=function(){return B(arguments)};t.bisect=s,t.bisectRight=s,t.bisectLeft=f,t.ascending=a,t.bisector=u,t.cross=h,t.descending=d,t.deviation=y,t.extent=v,t.histogram=C,t.thresholdFreedmanDiaconis=O,t.thresholdScott=P,t.thresholdSturges=S,t.max=D,t.mean=q,t.median=U,t.merge=j,t.min=F,t.pairs=l,t.permute=I,t.quantile=E,t.range=M,t.scan=z,t.shuffle=Y,t.sum=L,t.ticks=N,t.tickIncrement=r,t.tickStep=i,t.transpose=B,t.variance=_,t.zip=R,Object.defineProperty(t,"__esModule",{value:!0})})},{}],3:[function(t,n,e){!function(t,r){"object"==typeof e&&void 0!==n?r(e):"function"==typeof define&&define.amd?define(["exports"],r):r(t.d3=t.d3||{})}(this,function(t){"use strict";function n(){}function e(t,e){var r=new n;if(t instanceof n)t.each(function(t,n){r.set(n,t)});else if(Array.isArray(t)){var i,o=-1,a=t.length;if(null==e)for(;++o<a;)r.set(o,t[o]);else for(;++o<a;)r.set(e(i=t[o],o,t),i)}else if(t)for(var u in t)r.set(u,t[u]);return r}function r(){return{}}function i(t,n,e){t[n]=e}function o(){return e()}function a(t,n,e){t.set(n,e)}function u(){}function c(t,n){var e=new u;if(t instanceof u)t.each(function(t){e.add(t)});else if(t){var r=-1,i=t.length;if(null==n)for(;++r<i;)e.add(t[r]);else for(;++r<i;)e.add(n(t[r],r,t))}return e}n.prototype=e.prototype={constructor:n,has:function(t){return"$"+t in this},get:function(t){return this["$"+t]},set:function(t,n){return this["$"+t]=n,this},remove:function(t){var n="$"+t;return n in this&&delete this[n]},clear:function(){for(var t in this)"$"===t[0]&&delete this[t]},keys:function(){var t=[];for(var n in this)"$"===n[0]&&t.push(n.slice(1));return t},values:function(){var t=[];for(var n in this)"$"===n[0]&&t.push(this[n]);return t},entries:function(){var t=[];for(var n in this)"$"===n[0]&&t.push({key:n.slice(1),value:this[n]});return t},size:function(){var t=0;for(var n in this)"$"===n[0]&&++t;return t},empty:function(){for(var t in this)if("$"===t[0])return!1;return!0},each:function(t){for(var n in this)"$"===n[0]&&t(this[n],n.slice(1),this)}};var s=function(){function t(n,r,i,o){if(r>=f.length)return null!=c?c(n):null!=u?n.sort(u):n;for(var a,s,l,h=-1,d=n.length,p=f[r++],_=e(),y=i();++h<d;)(l=_.get(a=p(s=n[h])+""))?l.push(s):_.set(a,[s]);return _.each(function(n,e){o(y,e,t(n,r,i,o))}),y}function n(t,e){if(++e>f.length)return t;var r,i=l[e-1];return null!=c&&e>=f.length?r=t.entries():(r=[],t.each(function(t,i){r.push({key:i,values:n(t,e)})})),null!=i?r.sort(function(t,n){return i(t.key,n.key)}):r}var u,c,s,f=[],l=[];return s={object:function(n){return t(n,0,r,i)},map:function(n){return t(n,0,o,a)},entries:function(e){return n(t(e,0,o,a),0)},key:function(t){return f.push(t),s},sortKeys:function(t){return l[f.length-1]=t,s},sortValues:function(t){return u=t,s},rollup:function(t){return c=t,s}}},f=e.prototype;u.prototype=c.prototype={constructor:u,has:f.has,add:function(t){return t+="",this["$"+t]=t,this},remove:f.remove,clear:f.clear,values:f.keys,size:f.size,empty:f.empty,each:f.each};var l=function(t){var n=[];for(var e in t)n.push(e);return n},h=function(t){var n=[];for(var e in t)n.push(t[e]);return n},d=function(t){var n=[];for(var e in t)n.push({key:e,value:t[e]});return n};t.nest=s,t.set=c,t.map=e,t.keys=l,t.values=h,t.entries=d,Object.defineProperty(t,"__esModule",{value:!0})})},{}],4:[function(t,n,e){!function(t,r){"object"==typeof e&&void 0!==n?r(e):"function"==typeof define&&define.amd?define(["exports"],r):r(t.d3=t.d3||{})}(this,function(t){"use strict";function n(t,n){var e=Object.create(t.prototype);for(var r in n)e[r]=n[r];return e}function e(){}function r(t){var n;return t=(t+"").trim().toLowerCase(),(n=O.exec(t))?(n=parseInt(n[1],16),new c(n>>8&15|n>>4&240,n>>4&15|240&n,(15&n)<<4|15&n,1)):(n=P.exec(t))?i(parseInt(n[1],16)):(n=D.exec(t))?new c(n[1],n[2],n[3],1):(n=q.exec(t))?new c(255*n[1]/100,255*n[2]/100,255*n[3]/100,1):(n=U.exec(t))?o(n[1],n[2],n[3],n[4]):(n=j.exec(t))?o(255*n[1]/100,255*n[2]/100,255*n[3]/100,n[4]):(n=F.exec(t))?s(n[1],n[2]/100,n[3]/100,1):(n=I.exec(t))?s(n[1],n[2]/100,n[3]/100,n[4]):z.hasOwnProperty(t)?i(z[t]):"transparent"===t?new c(NaN,NaN,NaN,0):null}function i(t){return new c(t>>16&255,t>>8&255,255&t,1)}function o(t,n,e,r){return r<=0&&(t=n=e=NaN),new c(t,n,e,r)}function a(t){return t instanceof e||(t=r(t)),t?(t=t.rgb(),new c(t.r,t.g,t.b,t.opacity)):new c}function u(t,n,e,r){return 1===arguments.length?a(t):new c(t,n,e,null==r?1:r)}function c(t,n,e,r){this.r=+t,this.g=+n,this.b=+e,this.opacity=+r}function s(t,n,e,r){return r<=0?t=n=e=NaN:e<=0||e>=1?t=n=NaN:n<=0&&(t=NaN),new h(t,n,e,r)}function f(t){if(t instanceof h)return new h(t.h,t.s,t.l,t.opacity);if(t instanceof e||(t=r(t)),!t)return new h;if(t instanceof h)return t;var n=(t=t.rgb()).r/255,i=t.g/255,o=t.b/255,a=Math.min(n,i,o),u=Math.max(n,i,o),c=NaN,s=u-a,f=(u+a)/2;return s?(c=n===u?(i-o)/s+6*(i<o):i===u?(o-n)/s+2:(n-i)/s+4,s/=f<.5?u+a:2-u-a,c*=60):s=f>0&&f<1?0:c,new h(c,s,f,t.opacity)}function l(t,n,e,r){return 1===arguments.length?f(t):new h(t,n,e,null==r?1:r)}function h(t,n,e,r){this.h=+t,this.s=+n,this.l=+e,this.opacity=+r}function d(t,n,e){return 255*(t<60?n+(e-n)*t/60:t<180?e:t<240?n+(e-n)*(240-t)/60:n)}function p(t){if(t instanceof y)return new y(t.l,t.a,t.b,t.opacity);if(t instanceof M){var n=t.h*Y;return new y(t.l,Math.cos(n)*t.c,Math.sin(n)*t.c,t.opacity)}t instanceof c||(t=a(t));var e=b(t.r),r=b(t.g),i=b(t.b),o=v((.4124564*e+.3575761*r+.1804375*i)/B),u=v((.2126729*e+.7151522*r+.072175*i)/R);return new y(116*u-16,500*(o-u),200*(u-v((.0193339*e+.119192*r+.9503041*i)/H)),t.opacity)}function _(t,n,e,r){return 1===arguments.length?p(t):new y(t,n,e,null==r?1:r)}function y(t,n,e,r){this.l=+t,this.a=+n,this.b=+e,this.opacity=+r}function v(t){return t>V?Math.pow(t,1/3):t/W+X}function g(t){return t>$?t*t*t:W*(t-X)}function m(t){return 255*(t<=.0031308?12.92*t:1.055*Math.pow(t,1/2.4)-.055)}function b(t){return(t/=255)<=.04045?t/12.92:Math.pow((t+.055)/1.055,2.4)}function x(t){if(t instanceof M)return new M(t.h,t.c,t.l,t.opacity);t instanceof y||(t=p(t));var n=Math.atan2(t.b,t.a)*L;return new M(n<0?n+360:n,Math.sqrt(t.a*t.a+t.b*t.b),t.l,t.opacity)}function w(t,n,e,r){return 1===arguments.length?x(t):new M(t,n,e,null==r?1:r)}function M(t,n,e,r){this.h=+t,this.c=+n,this.l=+e,this.opacity=+r}function T(t){if(t instanceof A)return new A(t.h,t.s,t.l,t.opacity);t instanceof c||(t=a(t));var n=t.r/255,e=t.g/255,r=t.b/255,i=(et*r+tt*n-nt*e)/(et+tt-nt),o=r-i,u=(K*(e-i)-J*o)/G,s=Math.sqrt(u*u+o*o)/(K*i*(1-i)),f=s?Math.atan2(u,o)*L-120:NaN;return new A(f<0?f+360:f,s,i,t.opacity)}function k(t,n,e,r){return 1===arguments.length?T(t):new A(t,n,e,null==r?1:r)}function A(t,n,e,r){this.h=+t,this.s=+n,this.l=+e,this.opacity=+r}var N=function(t,n,e){t.prototype=n.prototype=e,e.constructor=t},S="\\s*([+-]?\\d+)\\s*",C="\\s*([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)\\s*",E="\\s*([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)%\\s*",O=/^#([0-9a-f]{3})$/,P=/^#([0-9a-f]{6})$/,D=new RegExp("^rgb\\("+[S,S,S]+"\\)$"),q=new RegExp("^rgb\\("+[E,E,E]+"\\)$"),U=new RegExp("^rgba\\("+[S,S,S,C]+"\\)$"),j=new RegExp("^rgba\\("+[E,E,E,C]+"\\)$"),F=new RegExp("^hsl\\("+[C,E,E]+"\\)$"),I=new RegExp("^hsla\\("+[C,E,E,C]+"\\)$"),z={aliceblue:15792383,antiquewhite:16444375,aqua:65535,aquamarine:8388564,azure:15794175,beige:16119260,bisque:16770244,black:0,blanchedalmond:16772045,blue:255,blueviolet:9055202,brown:10824234,burlywood:14596231,cadetblue:6266528,chartreuse:8388352,chocolate:13789470,coral:16744272,cornflowerblue:6591981,cornsilk:16775388,crimson:14423100,cyan:65535,darkblue:139,darkcyan:35723,darkgoldenrod:12092939,darkgray:11119017,darkgreen:25600,darkgrey:11119017,darkkhaki:12433259,darkmagenta:9109643,darkolivegreen:5597999,darkorange:16747520,darkorchid:10040012,darkred:9109504,darksalmon:15308410,darkseagreen:9419919,darkslateblue:4734347,darkslategray:3100495,darkslategrey:3100495,darkturquoise:52945,darkviolet:9699539,deeppink:16716947,deepskyblue:49151,dimgray:6908265,dimgrey:6908265,dodgerblue:2003199,firebrick:11674146,floralwhite:16775920,forestgreen:2263842,fuchsia:16711935,gainsboro:14474460,ghostwhite:16316671,gold:16766720,goldenrod:14329120,gray:8421504,green:32768,greenyellow:11403055,grey:8421504,honeydew:15794160,hotpink:16738740,indianred:13458524,indigo:4915330,ivory:16777200,khaki:15787660,lavender:15132410,lavenderblush:16773365,lawngreen:8190976,lemonchiffon:16775885,lightblue:11393254,lightcoral:15761536,lightcyan:14745599,lightgoldenrodyellow:16448210,lightgray:13882323,lightgreen:9498256,lightgrey:13882323,lightpink:16758465,lightsalmon:16752762,lightseagreen:2142890,lightskyblue:8900346,lightslategray:7833753,lightslategrey:7833753,lightsteelblue:11584734,lightyellow:16777184,lime:65280,limegreen:3329330,linen:16445670,magenta:16711935,maroon:8388608,mediumaquamarine:6737322,mediumblue:205,mediumorchid:12211667,mediumpurple:9662683,mediumseagreen:3978097,mediumslateblue:8087790,mediumspringgreen:64154,mediumturquoise:4772300,mediumvioletred:13047173,midnightblue:1644912,mintcream:16121850,mistyrose:16770273,moccasin:16770229,navajowhite:16768685,navy:128,oldlace:16643558,olive:8421376,olivedrab:7048739,orange:16753920,orangered:16729344,orchid:14315734,palegoldenrod:15657130,palegreen:10025880,paleturquoise:11529966,palevioletred:14381203,papayawhip:16773077,peachpuff:16767673,peru:13468991,pink:16761035,plum:14524637,powderblue:11591910,purple:8388736,rebeccapurple:6697881,red:16711680,rosybrown:12357519,royalblue:4286945,saddlebrown:9127187,salmon:16416882,sandybrown:16032864,seagreen:3050327,seashell:16774638,sienna:10506797,silver:12632256,skyblue:8900331,slateblue:6970061,slategray:7372944,slategrey:7372944,snow:16775930,springgreen:65407,steelblue:4620980,tan:13808780,teal:32896,thistle:14204888,tomato:16737095,turquoise:4251856,violet:15631086,wheat:16113331,white:16777215,whitesmoke:16119285,yellow:16776960,yellowgreen:10145074};N(e,r,{displayable:function(){return this.rgb().displayable()},toString:function(){return this.rgb()+""}}),N(c,u,n(e,{brighter:function(t){return t=null==t?1/.7:Math.pow(1/.7,t),new c(this.r*t,this.g*t,this.b*t,this.opacity)},darker:function(t){return t=null==t?.7:Math.pow(.7,t),new c(this.r*t,this.g*t,this.b*t,this.opacity)},rgb:function(){return this},displayable:function(){return 0<=this.r&&this.r<=255&&0<=this.g&&this.g<=255&&0<=this.b&&this.b<=255&&0<=this.opacity&&this.opacity<=1},toString:function(){var t=this.opacity;return t=isNaN(t)?1:Math.max(0,Math.min(1,t)),(1===t?"rgb(":"rgba(")+Math.max(0,Math.min(255,Math.round(this.r)||0))+", "+Math.max(0,Math.min(255,Math.round(this.g)||0))+", "+Math.max(0,Math.min(255,Math.round(this.b)||0))+(1===t?")":", "+t+")")}})),N(h,l,n(e,{brighter:function(t){return t=null==t?1/.7:Math.pow(1/.7,t),new h(this.h,this.s,this.l*t,this.opacity)},darker:function(t){return t=null==t?.7:Math.pow(.7,t),new h(this.h,this.s,this.l*t,this.opacity)},rgb:function(){var t=this.h%360+360*(this.h<0),n=isNaN(t)||isNaN(this.s)?0:this.s,e=this.l,r=e+(e<.5?e:1-e)*n,i=2*e-r;return new c(d(t>=240?t-240:t+120,i,r),d(t,i,r),d(t<120?t+240:t-120,i,r),this.opacity)},displayable:function(){return(0<=this.s&&this.s<=1||isNaN(this.s))&&0<=this.l&&this.l<=1&&0<=this.opacity&&this.opacity<=1}}));var Y=Math.PI/180,L=180/Math.PI,B=.95047,R=1,H=1.08883,X=4/29,$=6/29,W=3*$*$,V=$*$*$;N(y,_,n(e,{brighter:function(t){return new y(this.l+18*(null==t?1:t),this.a,this.b,this.opacity)},darker:function(t){return new y(this.l-18*(null==t?1:t),this.a,this.b,this.opacity)},rgb:function(){var t=(this.l+16)/116,n=isNaN(this.a)?t:t+this.a/500,e=isNaN(this.b)?t:t-this.b/200;return t=R*g(t),n=B*g(n),e=H*g(e),new c(m(3.2404542*n-1.5371385*t-.4985314*e),m(-.969266*n+1.8760108*t+.041556*e),m(.0556434*n-.2040259*t+1.0572252*e),this.opacity)}})),N(M,w,n(e,{brighter:function(t){return new M(this.h,this.c,this.l+18*(null==t?1:t),this.opacity)},darker:function(t){return new M(this.h,this.c,this.l-18*(null==t?1:t),this.opacity)},rgb:function(){return p(this).rgb()}}));var Z=-.14861,Q=1.78277,J=-.29227,G=-.90649,K=1.97294,tt=K*G,nt=K*Q,et=Q*J-G*Z;N(A,k,n(e,{brighter:function(t){return t=null==t?1/.7:Math.pow(1/.7,t),new A(this.h,this.s,this.l*t,this.opacity)},darker:function(t){return t=null==t?.7:Math.pow(.7,t),new A(this.h,this.s,this.l*t,this.opacity)},rgb:function(){var t=isNaN(this.h)?0:(this.h+120)*Y,n=+this.l,e=isNaN(this.s)?0:this.s*n*(1-n),r=Math.cos(t),i=Math.sin(t);return new c(255*(n+e*(Z*r+Q*i)),255*(n+e*(J*r+G*i)),255*(n+e*(K*r)),this.opacity)}})),t.color=r,t.rgb=u,t.hsl=l,t.lab=_,t.hcl=w,t.cubehelix=k,Object.defineProperty(t,"__esModule",{value:!0})})},{}],5:[function(t,n,e){!function(t,r){"object"==typeof e&&void 0!==n?r(e):"function"==typeof define&&define.amd?define(["exports"],r):r(t.d3=t.d3||{})}(this,function(t){"use strict";function n(){for(var t,n=0,r=arguments.length,i={};n<r;++n){if(!(t=arguments[n]+"")||t in i)throw new Error("illegal type: "+t);i[t]=[]}return new e(i)}function e(t){this._=t}function r(t,n){return t.trim().split(/^|\s+/).map(function(t){var e="",r=t.indexOf(".");if(r>=0&&(e=t.slice(r+1),t=t.slice(0,r)),t&&!n.hasOwnProperty(t))throw new Error("unknown type: "+t);return{type:t,name:e}})}function i(t,n){for(var e,r=0,i=t.length;r<i;++r)if((e=t[r]).name===n)return e.value}function o(t,n,e){for(var r=0,i=t.length;r<i;++r)if(t[r].name===n){t[r]=a,t=t.slice(0,r).concat(t.slice(r+1));break}return null!=e&&t.push({name:n,value:e}),t}var a={value:function(){}};e.prototype=n.prototype={constructor:e,on:function(t,n){var e,a=this._,u=r(t+"",a),c=-1,s=u.length;{if(!(arguments.length<2)){if(null!=n&&"function"!=typeof n)throw new Error("invalid callback: "+n);for(;++c<s;)if(e=(t=u[c]).type)a[e]=o(a[e],t.name,n);else if(null==n)for(e in a)a[e]=o(a[e],t.name,null);return this}for(;++c<s;)if((e=(t=u[c]).type)&&(e=i(a[e],t.name)))return e}},copy:function(){var t={},n=this._;for(var r in n)t[r]=n[r].slice();return new e(t)},call:function(t,n){if((e=arguments.length-2)>0)for(var e,r,i=new Array(e),o=0;o<e;++o)i[o]=arguments[o+2];if(!this._.hasOwnProperty(t))throw new Error("unknown type: "+t);for(o=0,e=(r=this._[t]).length;o<e;++o)r[o].value.apply(n,i)},apply:function(t,n,e){if(!this._.hasOwnProperty(t))throw new Error("unknown type: "+t);for(var r=this._[t],i=0,o=r.length;i<o;++i)r[i].value.apply(n,e)}},t.dispatch=n,Object.defineProperty(t,"__esModule",{value:!0})})},{}],6:[function(t,n,e){!function(t,r){"object"==typeof e&&void 0!==n?r(e):"function"==typeof define&&define.amd?define(["exports"],r):r(t.d3=t.d3||{})}(this,function(t){"use strict";function n(t){return+t}function e(t){return t*t}function r(t){return t*(2-t)}function i(t){return((t*=2)<=1?t*t:--t*(2-t)+1)/2}function o(t){return t*t*t}function a(t){return--t*t*t+1}function u(t){return((t*=2)<=1?t*t*t:(t-=2)*t*t+2)/2}function c(t){return 1-Math.cos(t*T)}function s(t){return Math.sin(t*T)}function f(t){return(1-Math.cos(M*t))/2}function l(t){return Math.pow(2,10*t-10)}function h(t){return 1-Math.pow(2,-10*t)}function d(t){return((t*=2)<=1?Math.pow(2,10*t-10):2-Math.pow(2,10-10*t))/2}function p(t){return 1-Math.sqrt(1-t*t)}function _(t){return Math.sqrt(1- --t*t)}function y(t){return((t*=2)<=1?1-Math.sqrt(1-t*t):Math.sqrt(1-(t-=2)*t)+1)/2}function v(t){return 1-g(1-t)}function g(t){return(t=+t)<k?q*t*t:t<N?q*(t-=A)*t+S:t<E?q*(t-=C)*t+O:q*(t-=P)*t+D}function m(t){return((t*=2)<=1?1-g(1-t):g(t-1)+1)/2}var b=function t(n){function e(t){return Math.pow(t,n)}return n=+n,e.exponent=t,e}(3),x=function t(n){function e(t){return 1-Math.pow(1-t,n)}return n=+n,e.exponent=t,e}(3),w=function t(n){function e(t){return((t*=2)<=1?Math.pow(t,n):2-Math.pow(2-t,n))/2}return n=+n,e.exponent=t,e}(3),M=Math.PI,T=M/2,k=4/11,A=6/11,N=8/11,S=.75,C=9/11,E=10/11,O=.9375,P=21/22,D=63/64,q=1/k/k,U=function t(n){function e(t){return t*t*((n+1)*t-n)}return n=+n,e.overshoot=t,e}(1.70158),j=function t(n){function e(t){return--t*t*((n+1)*t+n)+1}return n=+n,e.overshoot=t,e}(1.70158),F=function t(n){function e(t){return((t*=2)<1?t*t*((n+1)*t-n):(t-=2)*t*((n+1)*t+n)+2)/2}return n=+n,e.overshoot=t,e}(1.70158),I=2*Math.PI,z=function t(n,e){function r(t){return n*Math.pow(2,10*--t)*Math.sin((i-t)/e)}var i=Math.asin(1/(n=Math.max(1,n)))*(e/=I);return r.amplitude=function(n){return t(n,e*I)},r.period=function(e){return t(n,e)},r}(1,.3),Y=function t(n,e){function r(t){return 1-n*Math.pow(2,-10*(t=+t))*Math.sin((t+i)/e)}var i=Math.asin(1/(n=Math.max(1,n)))*(e/=I);return r.amplitude=function(n){return t(n,e*I)},r.period=function(e){return t(n,e)},r}(1,.3),L=function t(n,e){function r(t){return((t=2*t-1)<0?n*Math.pow(2,10*t)*Math.sin((i-t)/e):2-n*Math.pow(2,-10*t)*Math.sin((i+t)/e))/2}var i=Math.asin(1/(n=Math.max(1,n)))*(e/=I);return r.amplitude=function(n){return t(n,e*I)},r.period=function(e){return t(n,e)},r}(1,.3);t.easeLinear=n,t.easeQuad=i,t.easeQuadIn=e,t.easeQuadOut=r,t.easeQuadInOut=i,t.easeCubic=u,t.easeCubicIn=o,t.easeCubicOut=a,t.easeCubicInOut=u,t.easePoly=w,t.easePolyIn=b,t.easePolyOut=x,t.easePolyInOut=w,t.easeSin=f,t.easeSinIn=c,t.easeSinOut=s,t.easeSinInOut=f,t.easeExp=d,t.easeExpIn=l,t.easeExpOut=h,t.easeExpInOut=d,t.easeCircle=y,t.easeCircleIn=p,t.easeCircleOut=_,t.easeCircleInOut=y,t.easeBounce=g,t.easeBounceIn=v,t.easeBounceOut=g,t.easeBounceInOut=m,t.easeBack=F,t.easeBackIn=U,t.easeBackOut=j,t.easeBackInOut=F,t.easeElastic=Y,t.easeElasticIn=z,t.easeElasticOut=Y,t.easeElasticInOut=L,Object.defineProperty(t,"__esModule",{value:!0})})},{}],7:[function(t,n,e){!function(t,r){"object"==typeof e&&void 0!==n?r(e):"function"==typeof define&&define.amd?define(["exports"],r):r(t.d3=t.d3||{})}(this,function(t){"use strict";function n(t){return new e(t)}function e(t){if(!(n=l.exec(t)))throw new Error("invalid format: "+t);var n,e=n[1]||" ",r=n[2]||">",i=n[3]||"-",o=n[4]||"",a=!!n[5],u=n[6]&&+n[6],c=!!n[7],s=n[8]&&+n[8].slice(1),h=n[9]||"";"n"===h?(c=!0,h="g"):f[h]||(h=""),(a||"0"===e&&"="===r)&&(a=!0,e="0",r="="),this.fill=e,this.align=r,this.sign=i,this.symbol=o,this.zero=a,this.width=u,this.comma=c,this.precision=s,this.type=h}function r(n){return h=_(n),t.format=h.format,t.formatPrefix=h.formatPrefix,h}var i,o=function(t,n){if((e=(t=n?t.toExponential(n-1):t.toExponential()).indexOf("e"))<0)return null;var e,r=t.slice(0,e);return[r.length>1?r[0]+r.slice(2):r,+t.slice(e+1)]},a=function(t){return t=o(Math.abs(t)),t?t[1]:NaN},u=function(t,n){return function(e,r){for(var i=e.length,o=[],a=0,u=t[0],c=0;i>0&&u>0&&(c+u+1>r&&(u=Math.max(1,r-c)),o.push(e.substring(i-=u,i+u)),!((c+=u+1)>r));)u=t[a=(a+1)%t.length];return o.reverse().join(n)}},c=function(t){return function(n){return n.replace(/[0-9]/g,function(n){return t[+n]})}},s=function(t,n){var e=o(t,n);if(!e)return t+"";var r=e[0],i=e[1];return i<0?"0."+new Array(-i).join("0")+r:r.length>i+1?r.slice(0,i+1)+"."+r.slice(i+1):r+new Array(i-r.length+2).join("0")},f={"":function(t,n){t:for(var e,r=(t=t.toPrecision(n)).length,i=1,o=-1;i<r;++i)switch(t[i]){case".":o=e=i;break;case"0":0===o&&(o=i),e=i;break;case"e":break t;default:o>0&&(o=0)}return o>0?t.slice(0,o)+t.slice(e+1):t},"%":function(t,n){return(100*t).toFixed(n)},b:function(t){return Math.round(t).toString(2)},c:function(t){return t+""},d:function(t){return Math.round(t).toString(10)},e:function(t,n){return t.toExponential(n)},f:function(t,n){return t.toFixed(n)},g:function(t,n){return t.toPrecision(n)},o:function(t){return Math.round(t).toString(8)},p:function(t,n){return s(100*t,n)},r:s,s:function(t,n){var e=o(t,n);if(!e)return t+"";var r=e[0],a=e[1],u=a-(i=3*Math.max(-8,Math.min(8,Math.floor(a/3))))+1,c=r.length;return u===c?r:u>c?r+new Array(u-c+1).join("0"):u>0?r.slice(0,u)+"."+r.slice(u):"0."+new Array(1-u).join("0")+o(t,Math.max(0,n+u-1))[0]},X:function(t){return Math.round(t).toString(16).toUpperCase()},x:function(t){return Math.round(t).toString(16)}},l=/^(?:(.)?([<>=^]))?([+\-\( ])?([$#])?(0)?(\d+)?(,)?(\.\d+)?([a-z%])?$/i;n.prototype=e.prototype,e.prototype.toString=function(){return this.fill+this.align+this.sign+this.symbol+(this.zero?"0":"")+(null==this.width?"":Math.max(1,0|this.width))+(this.comma?",":"")+(null==this.precision?"":"."+Math.max(0,0|this.precision))+this.type};var h,d=function(t){return t},p=["y","z","a","f","p","n","µ","m","","k","M","G","T","P","E","Z","Y"],_=function(t){function e(t){function e(t){var n,e,c,s=b,f=x;if("c"===m)f=w(t)+f,t="";else{var _=(t=+t)<0;if(t=w(Math.abs(t),g),_&&0==+t&&(_=!1),s=(_?"("===u?u:"-":"-"===u||"("===u?"":u)+s,f=f+("s"===m?p[8+i/3]:"")+(_&&"("===u?")":""),M)for(n=-1,e=t.length;++n<e;)if(48>(c=t.charCodeAt(n))||c>57){f=(46===c?l+t.slice(n+1):t.slice(n))+f,t=t.slice(0,n);break}}v&&!d&&(t=o(t,1/0));var T=s.length+t.length+f.length,k=T<y?new Array(y-T+1).join(r):"";switch(v&&d&&(t=o(k+t,k.length?y-f.length:1/0),k=""),a){case"<":t=s+t+f+k;break;case"=":t=s+k+t+f;break;case"^":t=k.slice(0,T=k.length>>1)+s+t+f+k.slice(T);break;default:t=k+s+t+f}return h(t)}var r=(t=n(t)).fill,a=t.align,u=t.sign,c=t.symbol,d=t.zero,y=t.width,v=t.comma,g=t.precision,m=t.type,b="$"===c?s[0]:"#"===c&&/[boxX]/.test(m)?"0"+m.toLowerCase():"",x="$"===c?s[1]:/[%p]/.test(m)?_:"",w=f[m],M=!m||/[defgprs%]/.test(m);return g=null==g?m?6:12:/[gprs]/.test(m)?Math.max(1,Math.min(21,g)):Math.max(0,Math.min(20,g)),e.toString=function(){return t+""},e}function r(t,r){var i=e((t=n(t),t.type="f",t)),o=3*Math.max(-8,Math.min(8,Math.floor(a(r)/3))),u=Math.pow(10,-o),c=p[8+o/3];return function(t){return i(u*t)+c}}var o=t.grouping&&t.thousands?u(t.grouping,t.thousands):d,s=t.currency,l=t.decimal,h=t.numerals?c(t.numerals):d,_=t.percent||"%";return{format:e,formatPrefix:r}};r({decimal:".",thousands:",",grouping:[3],currency:["$",""]});var y=function(t){return Math.max(0,-a(Math.abs(t)))},v=function(t,n){return Math.max(0,3*Math.max(-8,Math.min(8,Math.floor(a(n)/3)))-a(Math.abs(t)))},g=function(t,n){return t=Math.abs(t),n=Math.abs(n)-t,Math.max(0,a(n)-a(t))+1};t.formatDefaultLocale=r,t.formatLocale=_,t.formatSpecifier=n,t.precisionFixed=y,t.precisionPrefix=v,t.precisionRound=g,Object.defineProperty(t,"__esModule",{value:!0})})},{}],8:[function(t,n,e){!function(t,r){"object"==typeof e&&void 0!==n?r(e):"function"==typeof define&&define.amd?define(["exports"],r):r(t.d3=t.d3||{})}(this,function(t){"use strict";function n(t,n){return t.parent===n.parent?1:2}function e(t){return t.reduce(r,0)/t.length}function r(t,n){return t+n.x}function i(t){return 1+t.reduce(o,0)}function o(t,n){return Math.max(t,n.y)}function a(t){for(var n;n=t.children;)t=n[0];return t}function u(t){for(var n;n=t.children;)t=n[n.length-1];return t}function c(t){var n=0,e=t.children,r=e&&e.length;if(r)for(;--r>=0;)n+=e[r].value;else n=1;t.value=n}function s(t,n){if(t===n)return t;var e=t.ancestors(),r=n.ancestors(),i=null;for(t=e.pop(),n=r.pop();t===n;)i=t,t=e.pop(),n=r.pop();return i}function f(t,n){var e,r,i,o,a,u=new _(t),c=+t.value&&(u.value=t.value),s=[u];for(null==n&&(n=h);e=s.pop();)if(c&&(e.value=+e.data.value),(i=n(e.data))&&(a=i.length))for(e.children=new Array(a),o=a-1;o>=0;--o)s.push(r=e.children[o]=new _(i[o])),r.parent=e,r.depth=e.depth+1;return u.eachBefore(p)}function l(){return f(this).eachBefore(d)}function h(t){return t.children}function d(t){t.data=t.data.data}function p(t){var n=0;do{t.height=n}while((t=t.parent)&&t.height<++n)}function _(t){this.data=t,this.depth=this.height=0,this.parent=null}function y(t){for(var n,e,r=t.length;r;)e=Math.random()*r--|0,n=t[r],t[r]=t[e],t[e]=n;return t}function v(t,n){var e,r;if(b(n,t))return[n];for(e=0;e<t.length;++e)if(g(n,t[e])&&b(M(t[e],n),t))return[t[e],n];for(e=0;e<t.length-1;++e)for(r=e+1;r<t.length;++r)if(g(M(t[e],t[r]),n)&&g(M(t[e],n),t[r])&&g(M(t[r],n),t[e])&&b(T(t[e],t[r],n),t))return[t[e],t[r],n];throw new Error}function g(t,n){var e=t.r-n.r,r=n.x-t.x,i=n.y-t.y;return e<0||e*e<r*r+i*i}function m(t,n){var e=t.r-n.r+1e-6,r=n.x-t.x,i=n.y-t.y;return e>0&&e*e>r*r+i*i}function b(t,n){for(var e=0;e<n.length;++e)if(!m(t,n[e]))return!1;return!0}function x(t){switch(t.length){case 1:return w(t[0]);case 2:return M(t[0],t[1]);case 3:return T(t[0],t[1],t[2])}}function w(t){return{x:t.x,y:t.y,r:t.r}}function M(t,n){var e=t.x,r=t.y,i=t.r,o=n.x,a=n.y,u=n.r,c=o-e,s=a-r,f=u-i,l=Math.sqrt(c*c+s*s);return{x:(e+o+c/l*f)/2,y:(r+a+s/l*f)/2,r:(l+i+u)/2}}function T(t,n,e){var r=t.x,i=t.y,o=t.r,a=n.x,u=n.y,c=n.r,s=e.x,f=e.y,l=e.r,h=r-a,d=r-s,p=i-u,_=i-f,y=c-o,v=l-o,g=r*r+i*i-o*o,m=g-a*a-u*u+c*c,b=g-s*s-f*f+l*l,x=d*p-h*_,w=(p*b-_*m)/(2*x)-r,M=(_*y-p*v)/x,T=(d*m-h*b)/(2*x)-i,k=(h*v-d*y)/x,A=M*M+k*k-1,N=2*(o+w*M+T*k),S=w*w+T*T-o*o,C=-(A?(N+Math.sqrt(N*N-4*A*S))/(2*A):S/N);return{x:r+w+M*C,y:i+T+k*C,r:C}}function k(t,n,e){var r=t.x,i=t.y,o=n.r+e.r,a=t.r+e.r,u=n.x-r,c=n.y-i,s=u*u+c*c;if(s){var f=.5+((a*=a)-(o*=o))/(2*s),l=Math.sqrt(Math.max(0,2*o*(a+s)-(a-=s)*a-o*o))/(2*s);e.x=r+f*u+l*c,e.y=i+f*c-l*u}else e.x=r+a,e.y=i}function A(t,n){var e=n.x-t.x,r=n.y-t.y,i=t.r+n.r;return i*i-1e-6>e*e+r*r}function N(t){var n=t._,e=t.next._,r=n.r+e.r,i=(n.x*e.r+e.x*n.r)/r,o=(n.y*e.r+e.y*n.r)/r;return i*i+o*o}function S(t){this._=t,this.next=null,this.previous=null}function C(t){if(!(i=t.length))return 0;var n,e,r,i,o,a,u,c,s,f,l;if(n=t[0],n.x=0,n.y=0,!(i>1))return n.r;if(e=t[1],n.x=-e.r,e.x=n.r,e.y=0,!(i>2))return n.r+e.r;k(e,n,r=t[2]),n=new S(n),e=new S(e),r=new S(r),n.next=r.previous=e,e.next=n.previous=r,r.next=e.previous=n;t:for(u=3;u<i;++u){k(n._,e._,r=t[u]),r=new S(r),c=e.next,s=n.previous,f=e._.r,l=n._.r;do{if(f<=l){if(A(c._,r._)){e=c,n.next=e,e.previous=n,--u;continue t}f+=c._.r,c=c.next}else{if(A(s._,r._)){(n=s).next=e,e.previous=n,--u;continue t}l+=s._.r,s=s.previous}}while(c!==s.next);for(r.previous=n,r.next=e,n.next=e.previous=e=r,o=N(n);(r=r.next)!==e;)(a=N(r))<o&&(n=r,o=a);e=n.next}for(n=[e._],r=e;(r=r.next)!==e;)n.push(r._);for(r=ut(n),u=0;u<i;++u)n=t[u],n.x-=r.x,n.y-=r.y;return r.r}function E(t){return null==t?null:O(t)}function O(t){if("function"!=typeof t)throw new Error;return t}function P(){return 0}function D(t){return Math.sqrt(t.value)}function q(t){return function(n){n.children||(n.r=Math.max(0,+t(n)||0))}}function U(t,n){return function(e){if(r=e.children){var r,i,o,a=r.length,u=t(e)*n||0;if(u)for(i=0;i<a;++i)r[i].r+=u;if(o=C(r),u)for(i=0;i<a;++i)r[i].r-=u;e.r=o+u}}}function j(t){return function(n){var e=n.parent;n.r*=t,e&&(n.x=e.x+t*n.x,n.y=e.y+t*n.y)}}function F(t){return t.id}function I(t){return t.parentId}function z(t,n){return t.parent===n.parent?1:2}function Y(t){var n=t.children;return n?n[0]:t.t}function L(t){var n=t.children;return n?n[n.length-1]:t.t}function B(t,n,e){var r=e/(n.i-t.i);n.c-=r,n.s+=e,t.c+=r,n.z+=e,n.m+=e}function R(t){for(var n,e=0,r=0,i=t.children,o=i.length;--o>=0;)(n=i[o]).z+=e,n.m+=e,e+=n.s+(r+=n.c)}function H(t,n,e){return t.a.parent===n.parent?t.a:e}function X(t,n){this._=t,this.parent=null,this.children=null,this.A=null,this.a=this,this.z=0,this.m=0,this.c=0,this.s=0,this.t=null,this.i=n}function $(t){for(var n,e,r,i,o,a=new X(t,0),u=[a];n=u.pop();)if(r=n._.children)for(n.children=new Array(o=r.length),i=o-1;i>=0;--i)u.push(e=n.children[i]=new X(r[i],i)),e.parent=n;return(a.parent=new X(null,0)).children=[a],a}function W(t,n,e,r,i,o){for(var a,u,c,s,f,l,h,d,p,_,y,v=[],g=n.children,m=0,b=0,x=g.length,w=n.value;m<x;){c=i-e,s=o-r;do{f=g[b++].value}while(!f&&b<x);for(l=h=f,y=f*f*(_=Math.max(s/c,c/s)/(w*t)),p=Math.max(h/y,y/l);b<x;++b){if(f+=u=g[b].value,u<l&&(l=u),u>h&&(h=u),y=f*f*_,(d=Math.max(h/y,y/l))>p){f-=u;break}p=d}v.push(a={value:f,dice:c<s,children:g.slice(m,b)}),a.dice?ht(a,e,r,i,w?r+=s*f/w:o):mt(a,e,r,w?e+=c*f/w:i,o),w-=f,m=b}return v}var V=function(){function t(t){var n,f=0;t.eachAfter(function(t){var o=t.children;o?(t.x=e(o),t.y=i(o)):(t.x=n?f+=r(t,n):0,t.y=0,n=t)});var l=a(t),h=u(t),d=l.x-r(l,h)/2,p=h.x+r(h,l)/2;return t.eachAfter(s?function(n){n.x=(n.x-t.x)*o,n.y=(t.y-n.y)*c}:function(n){n.x=(n.x-d)/(p-d)*o,n.y=(1-(t.y?n.y/t.y:1))*c})}var r=n,o=1,c=1,s=!1;return t.separation=function(n){return arguments.length?(r=n,t):r},t.size=function(n){return arguments.length?(s=!1,o=+n[0],c=+n[1],t):s?null:[o,c]},t.nodeSize=function(n){return arguments.length?(s=!0,o=+n[0],c=+n[1],t):s?[o,c]:null},t},Z=function(){return this.eachAfter(c)},Q=function(t){var n,e,r,i,o=this,a=[o];do{for(n=a.reverse(),a=[];o=n.pop();)if(t(o),e=o.children)for(r=0,i=e.length;r<i;++r)a.push(e[r])}while(a.length);return this},J=function(t){for(var n,e,r=this,i=[r];r=i.pop();)if(t(r),n=r.children)for(e=n.length-1;e>=0;--e)i.push(n[e]);return this},G=function(t){for(var n,e,r,i=this,o=[i],a=[];i=o.pop();)if(a.push(i),n=i.children)for(e=0,r=n.length;e<r;++e)o.push(n[e]);for(;i=a.pop();)t(i);return this},K=function(t){return this.eachAfter(function(n){for(var e=+t(n.data)||0,r=n.children,i=r&&r.length;--i>=0;)e+=r[i].value;n.value=e})},tt=function(t){return this.eachBefore(function(n){n.children&&n.children.sort(t)})},nt=function(t){for(var n=this,e=s(n,t),r=[n];n!==e;)n=n.parent,r.push(n);for(var i=r.length;t!==e;)r.splice(i,0,t),t=t.parent;return r},et=function(){for(var t=this,n=[t];t=t.parent;)n.push(t);return n},rt=function(){var t=[];return this.each(function(n){t.push(n)}),t},it=function(){var t=[];return this.eachBefore(function(n){n.children||t.push(n)}),t},ot=function(){var t=this,n=[];return t.each(function(e){e!==t&&n.push({source:e.parent,target:e})}),n};_.prototype=f.prototype={constructor:_,count:Z,each:Q,eachAfter:G,eachBefore:J,sum:K,sort:tt,path:nt,ancestors:et,descendants:rt,leaves:it,links:ot,copy:l};var at=Array.prototype.slice,ut=function(t){for(var n,e,r=0,i=(t=y(at.call(t))).length,o=[];r<i;)n=t[r],e&&m(e,n)?++r:(e=x(o=v(o,n)),r=0);return e},ct=function(t){return C(t),t},st=function(t){return function(){return t}},ft=function(){function t(t){return t.x=e/2,t.y=r/2,n?t.eachBefore(q(n)).eachAfter(U(i,.5)).eachBefore(j(1)):t.eachBefore(q(D)).eachAfter(U(P,1)).eachAfter(U(i,t.r/Math.min(e,r))).eachBefore(j(Math.min(e,r)/(2*t.r))),t}var n=null,e=1,r=1,i=P;return t.radius=function(e){return arguments.length?(n=E(e),t):n},t.size=function(n){return arguments.length?(e=+n[0],r=+n[1],t):[e,r]},t.padding=function(n){return arguments.length?(i="function"==typeof n?n:st(+n),t):i},t},lt=function(t){t.x0=Math.round(t.x0),t.y0=Math.round(t.y0),t.x1=Math.round(t.x1),t.y1=Math.round(t.y1)},ht=function(t,n,e,r,i){for(var o,a=t.children,u=-1,c=a.length,s=t.value&&(r-n)/t.value;++u<c;)(o=a[u]).y0=e,o.y1=i,o.x0=n,o.x1=n+=o.value*s},dt=function(){function t(t){var a=t.height+1;return t.x0=t.y0=i,t.x1=e,t.y1=r/a,t.eachBefore(n(r,a)),o&&t.eachBefore(lt),t}function n(t,n){return function(e){e.children&&ht(e,e.x0,t*(e.depth+1)/n,e.x1,t*(e.depth+2)/n);var r=e.x0,o=e.y0,a=e.x1-i,u=e.y1-i;a<r&&(r=a=(r+a)/2),u<o&&(o=u=(o+u)/2),e.x0=r,e.y0=o,e.x1=a,e.y1=u}}var e=1,r=1,i=0,o=!1;return t.round=function(n){return arguments.length?(o=!!n,t):o},t.size=function(n){return arguments.length?(e=+n[0],r=+n[1],t):[e,r]},t.padding=function(n){return arguments.length?(i=+n,t):i},t},pt="$",_t={depth:-1},yt={},vt=function(){function t(t){var r,i,o,a,u,c,s,f=t.length,l=new Array(f),h={};for(i=0;i<f;++i)r=t[i],u=l[i]=new _(r),null!=(c=n(r,i,t))&&(c+="")&&(h[s=pt+(u.id=c)]=s in h?yt:u);for(i=0;i<f;++i)if(u=l[i],null!=(c=e(t[i],i,t))&&(c+="")){if(!(a=h[pt+c]))throw new Error("missing: "+c);if(a===yt)throw new Error("ambiguous: "+c);a.children?a.children.push(u):a.children=[u],u.parent=a}else{if(o)throw new Error("multiple roots");o=u}if(!o)throw new Error("no root");if(o.parent=_t,o.eachBefore(function(t){t.depth=t.parent.depth+1,--f}).eachBefore(p),o.parent=null,f>0)throw new Error("cycle");return o}var n=F,e=I;return t.id=function(e){return arguments.length?(n=O(e),t):n},t.parentId=function(n){return arguments.length?(e=O(n),t):e},t};X.prototype=Object.create(_.prototype);var gt=function(){function t(t){var r=$(t);if(r.eachAfter(n),r.parent.m=-r.z,r.eachBefore(e),c)t.eachBefore(i);else{var s=t,f=t,l=t;t.eachBefore(function(t){t.x<s.x&&(s=t),t.x>f.x&&(f=t),t.depth>l.depth&&(l=t)});var h=s===f?1:o(s,f)/2,d=h-s.x,p=a/(f.x+h+d),_=u/(l.depth||1);t.eachBefore(function(t){t.x=(t.x+d)*p,t.y=t.depth*_})}return t}function n(t){var n=t.children,e=t.parent.children,i=t.i?e[t.i-1]:null;if(n){R(t);var a=(n[0].z+n[n.length-1].z)/2;i?(t.z=i.z+o(t._,i._),t.m=t.z-a):t.z=a}else i&&(t.z=i.z+o(t._,i._));t.parent.A=r(t,i,t.parent.A||e[0])}function e(t){t._.x=t.z+t.parent.m,t.m+=t.parent.m}function r(t,n,e){if(n){for(var r,i=t,a=t,u=n,c=i.parent.children[0],s=i.m,f=a.m,l=u.m,h=c.m;u=L(u),i=Y(i),u&&i;)c=Y(c),(a=L(a)).a=t,(r=u.z+l-i.z-s+o(u._,i._))>0&&(B(H(u,t,e),t,r),s+=r,f+=r),l+=u.m,s+=i.m,h+=c.m,f+=a.m;u&&!L(a)&&(a.t=u,a.m+=l-f),i&&!Y(c)&&(c.t=i,c.m+=s-h,e=t)}return e}function i(t){t.x*=a,t.y=t.depth*u}var o=z,a=1,u=1,c=null;return t.separation=function(n){return arguments.length?(o=n,t):o},t.size=function(n){return arguments.length?(c=!1,a=+n[0],u=+n[1],t):c?null:[a,u]},t.nodeSize=function(n){return arguments.length?(c=!0,a=+n[0],u=+n[1],t):c?[a,u]:null},t},mt=function(t,n,e,r,i){for(var o,a=t.children,u=-1,c=a.length,s=t.value&&(i-e)/t.value;++u<c;)(o=a[u]).x0=n,o.x1=r,o.y0=e,o.y1=e+=o.value*s},bt=(1+Math.sqrt(5))/2,xt=function t(n){function e(t,e,r,i,o){W(n,t,e,r,i,o)}return e.ratio=function(n){return t((n=+n)>1?n:1)},e}(bt),wt=function(){function t(t){return t.x0=t.y0=0,t.x1=i,t.y1=o,t.eachBefore(n),a=[0],r&&t.eachBefore(lt),t}function n(t){var n=a[t.depth],r=t.x0+n,i=t.y0+n,o=t.x1-n,h=t.y1-n;o<r&&(r=o=(r+o)/2),h<i&&(i=h=(i+h)/2),t.x0=r,t.y0=i,t.x1=o,t.y1=h,t.children&&(n=a[t.depth+1]=u(t)/2,r+=l(t)-n,i+=c(t)-n,o-=s(t)-n,h-=f(t)-n,o<r&&(r=o=(r+o)/2),h<i&&(i=h=(i+h)/2),e(t,r,i,o,h))}var e=xt,r=!1,i=1,o=1,a=[0],u=P,c=P,s=P,f=P,l=P;return t.round=function(n){return arguments.length?(r=!!n,t):r},t.size=function(n){return arguments.length?(i=+n[0],o=+n[1],t):[i,o]},t.tile=function(n){return arguments.length?(e=O(n),t):e},t.padding=function(n){return arguments.length?t.paddingInner(n).paddingOuter(n):t.paddingInner()},t.paddingInner=function(n){return arguments.length?(u="function"==typeof n?n:st(+n),t):u},t.paddingOuter=function(n){return arguments.length?t.paddingTop(n).paddingRight(n).paddingBottom(n).paddingLeft(n):t.paddingTop()},t.paddingTop=function(n){return arguments.length?(c="function"==typeof n?n:st(+n),t):c},t.paddingRight=function(n){return arguments.length?(s="function"==typeof n?n:st(+n),t):s},t.paddingBottom=function(n){return arguments.length?(f="function"==typeof n?n:st(+n),t):f},t.paddingLeft=function(n){return arguments.length?(l="function"==typeof n?n:st(+n),t):l},t},Mt=function(t,n,e,r,i){function o(t,n,e,r,i,a,u){if(t>=n-1){var s=c[t];return s.x0=r,s.y0=i,s.x1=a,void(s.y1=u)}for(var l=f[t],h=e/2+l,d=t+1,p=n-1;d<p;){var _=d+p>>>1;f[_]<h?d=_+1:p=_}h-f[d-1]<f[d]-h&&t+1<d&&--d;var y=f[d]-l,v=e-y;if(a-r>u-i){var g=(r*v+a*y)/e;o(t,d,y,r,i,g,u),o(d,n,v,g,i,a,u)}else{var m=(i*v+u*y)/e;o(t,d,y,r,i,a,m),o(d,n,v,r,m,a,u)}}var a,u,c=t.children,s=c.length,f=new Array(s+1);for(f[0]=u=a=0;a<s;++a)f[a+1]=u+=c[a].value;o(0,s,t.value,n,e,r,i)},Tt=function(t,n,e,r,i){(1&t.depth?mt:ht)(t,n,e,r,i)},kt=function t(n){function e(t,e,r,i,o){if((a=t._squarify)&&a.ratio===n)for(var a,u,c,s,f,l=-1,h=a.length,d=t.value;++l<h;){for(c=(u=a[l]).children,s=u.value=0,f=c.length;s<f;++s)u.value+=c[s].value;u.dice?ht(u,e,r,i,r+=(o-r)*u.value/d):mt(u,e,r,e+=(i-e)*u.value/d,o),d-=u.value}else t._squarify=a=W(n,t,e,r,i,o),a.ratio=n}return e.ratio=function(n){return t((n=+n)>1?n:1)},e}(bt);t.cluster=V,t.hierarchy=f,t.pack=ft,t.packSiblings=ct,t.packEnclose=ut,t.partition=dt,t.stratify=vt,t.tree=gt,t.treemap=wt,t.treemapBinary=Mt,t.treemapDice=ht,t.treemapSlice=mt,t.treemapSliceDice=Tt,t.treemapSquarify=xt,t.treemapResquarify=kt,Object.defineProperty(t,"__esModule",{value:!0})})},{}],9:[function(t,n,e){!function(r,i){"object"==typeof e&&void 0!==n?i(e,t("d3-color")):"function"==typeof define&&define.amd?define(["exports","d3-color"],i):i(r.d3=r.d3||{},r.d3)}(this,function(t,n){"use strict";function e(t,n,e,r,i){var o=t*t,a=o*t;return((1-3*t+3*o-a)*n+(4-6*o+3*a)*e+(1+3*t+3*o-3*a)*r+a*i)/6}function r(t,n){return function(e){return t+e*n}}function i(t,n,e){return t=Math.pow(t,e),n=Math.pow(n,e)-t,e=1/e,function(r){return Math.pow(t+r*n,e)}}function o(t,n){var e=n-t;return e?r(t,e>180||e<-180?e-360*Math.round(e/360):e):N(isNaN(t)?n:t)}function a(t){return 1==(t=+t)?u:function(n,e){return e-n?i(n,e,t):N(isNaN(n)?e:n)}}function u(t,n){var e=n-t;return e?r(t,e):N(isNaN(t)?n:t)}function c(t){return function(e){var r,i,o=e.length,a=new Array(o),u=new Array(o),c=new Array(o);for(r=0;r<o;++r)i=n.rgb(e[r]),a[r]=i.r||0,u[r]=i.g||0,c[r]=i.b||0;return a=t(a),u=t(u),c=t(c),i.opacity=1,function(t){return i.r=a(t),i.g=u(t),i.b=c(t),i+""}}}function s(t){return function(){return t}}function f(t){return function(n){return t(n)+""}}function l(t){return"none"===t?L:(x||(x=document.createElement("DIV"),w=document.documentElement,M=document.defaultView),x.style.transform=t,t=M.getComputedStyle(w.appendChild(x),null).getPropertyValue("transform"),w.removeChild(x),t=t.slice(7,-1).split(","),B(+t[0],+t[1],+t[2],+t[3],+t[4],+t[5]))}function h(t){return null==t?L:(T||(T=document.createElementNS("http://www.w3.org/2000/svg","g")),T.setAttribute("transform",t),(t=T.transform.baseVal.consolidate())?(t=t.matrix,B(t.a,t.b,t.c,t.d,t.e,t.f)):L)}function d(t,n,e,r){function i(t){return t.length?t.pop()+" ":""}function o(t,r,i,o,a,u){if(t!==i||r!==o){var c=a.push("translate(",null,n,null,e);u.push({i:c-4,x:D(t,i)},{i:c-2,x:D(r,o)})}else(i||o)&&a.push("translate("+i+n+o+e)}function a(t,n,e,o){t!==n?(t-n>180?n+=360:n-t>180&&(t+=360),o.push({i:e.push(i(e)+"rotate(",null,r)-2,x:D(t,n)})):n&&e.push(i(e)+"rotate("+n+r)}function u(t,n,e,o){t!==n?o.push({i:e.push(i(e)+"skewX(",null,r)-2,x:D(t,n)}):n&&e.push(i(e)+"skewX("+n+r)}function c(t,n,e,r,o,a){if(t!==e||n!==r){var u=o.push(i(o)+"scale(",null,",",null,")");a.push({i:u-4,x:D(t,e)},{i:u-2,x:D(n,r)})}else 1===e&&1===r||o.push(i(o)+"scale("+e+","+r+")")}return function(n,e){var r=[],i=[];return n=t(n),e=t(e),o(n.translateX,n.translateY,e.translateX,e.translateY,r,i),a(n.rotate,e.rotate,r,i),u(n.skewX,e.skewX,r,i),c(n.scaleX,n.scaleY,e.scaleX,e.scaleY,r,i),n=e=null,function(t){for(var n,e=-1,o=i.length;++e<o;)r[(n=i[e]).i]=n.x(t);return r.join("")}}}function p(t){return((t=Math.exp(t))+1/t)/2}function _(t){return((t=Math.exp(t))-1/t)/2}function y(t){return((t=Math.exp(2*t))-1)/(t+1)}function v(t){return function(e,r){var i=t((e=n.hsl(e)).h,(r=n.hsl(r)).h),o=u(e.s,r.s),a=u(e.l,r.l),c=u(e.opacity,r.opacity);return function(t){return e.h=i(t),e.s=o(t),e.l=a(t),e.opacity=c(t),e+""}}}function g(t,e){var r=u((t=n.lab(t)).l,(e=n.lab(e)).l),i=u(t.a,e.a),o=u(t.b,e.b),a=u(t.opacity,e.opacity);return function(n){return t.l=r(n),t.a=i(n),t.b=o(n),t.opacity=a(n),t+""}}function m(t){return function(e,r){var i=t((e=n.hcl(e)).h,(r=n.hcl(r)).h),o=u(e.c,r.c),a=u(e.l,r.l),c=u(e.opacity,r.opacity);return function(t){return e.h=i(t),e.c=o(t),e.l=a(t),e.opacity=c(t),e+""}}}function b(t){return function e(r){function i(e,i){var o=t((e=n.cubehelix(e)).h,(i=n.cubehelix(i)).h),a=u(e.s,i.s),c=u(e.l,i.l),s=u(e.opacity,i.opacity);return function(t){return e.h=o(t),e.s=a(t),e.l=c(Math.pow(t,r)),e.opacity=s(t),e+""}}return r=+r,i.gamma=e,i}(1)}var x,w,M,T,k=function(t){var n=t.length-1;return function(r){var i=r<=0?r=0:r>=1?(r=1,n-1):Math.floor(r*n),o=t[i],a=t[i+1],u=i>0?t[i-1]:2*o-a,c=i<n-1?t[i+2]:2*a-o;return e((r-i/n)*n,u,o,a,c)}},A=function(t){var n=t.length;return function(r){var i=Math.floor(((r%=1)<0?++r:r)*n),o=t[(i+n-1)%n],a=t[i%n],u=t[(i+1)%n],c=t[(i+2)%n];return e((r-i/n)*n,o,a,u,c)}},N=function(t){return function(){return t}},S=function t(e){function r(t,e){var r=i((t=n.rgb(t)).r,(e=n.rgb(e)).r),o=i(t.g,e.g),a=i(t.b,e.b),c=u(t.opacity,e.opacity);return function(n){return t.r=r(n),t.g=o(n),t.b=a(n),t.opacity=c(n),t+""}}var i=a(e);return r.gamma=t,r}(1),C=c(k),E=c(A),O=function(t,n){var e,r=n?n.length:0,i=t?Math.min(r,t.length):0,o=new Array(r),a=new Array(r);for(e=0;e<i;++e)o[e]=I(t[e],n[e]);for(;e<r;++e)a[e]=n[e];return function(t){for(e=0;e<i;++e)a[e]=o[e](t);return a}},P=function(t,n){var e=new Date;return t=+t,n-=t,function(r){return e.setTime(t+n*r),e}},D=function(t,n){return t=+t,n-=t,function(e){return t+n*e}},q=function(t,n){var e,r={},i={};null!==t&&"object"==typeof t||(t={}),null!==n&&"object"==typeof n||(n={});for(e in n)e in t?r[e]=I(t[e],n[e]):i[e]=n[e];return function(t){for(e in r)i[e]=r[e](t);return i}},U=/[-+]?(?:\d+\.?\d*|\.?\d+)(?:[eE][-+]?\d+)?/g,j=new RegExp(U.source,"g"),F=function(t,n){var e,r,i,o=U.lastIndex=j.lastIndex=0,a=-1,u=[],c=[];for(t+="",n+="";(e=U.exec(t))&&(r=j.exec(n));)(i=r.index)>o&&(i=n.slice(o,i),u[a]?u[a]+=i:u[++a]=i),(e=e[0])===(r=r[0])?u[a]?u[a]+=r:u[++a]=r:(u[++a]=null,c.push({i:a,x:D(e,r)})),o=j.lastIndex;return o<n.length&&(i=n.slice(o),u[a]?u[a]+=i:u[++a]=i),u.length<2?c[0]?f(c[0].x):s(n):(n=c.length,function(t){for(var e,r=0;r<n;++r)u[(e=c[r]).i]=e.x(t);return u.join("")})},I=function(t,e){var r,i=typeof e;return null==e||"boolean"===i?N(e):("number"===i?D:"string"===i?(r=n.color(e))?(e=r,S):F:e instanceof n.color?S:e instanceof Date?P:Array.isArray(e)?O:"function"!=typeof e.valueOf&&"function"!=typeof e.toString||isNaN(e)?q:D)(t,e)},z=function(t,n){return t=+t,n-=t,function(e){return Math.round(t+n*e)}},Y=180/Math.PI,L={translateX:0,translateY:0,rotate:0,skewX:0,scaleX:1,scaleY:1},B=function(t,n,e,r,i,o){var a,u,c;return(a=Math.sqrt(t*t+n*n))&&(t/=a,n/=a),(c=t*e+n*r)&&(e-=t*c,r-=n*c),(u=Math.sqrt(e*e+r*r))&&(e/=u,r/=u,c/=u),t*r<n*e&&(t=-t,n=-n,c=-c,a=-a),{translateX:i,translateY:o,rotate:Math.atan2(n,t)*Y,skewX:Math.atan(c)*Y,scaleX:a,scaleY:u}},R=d(l,"px, ","px)","deg)"),H=d(h,", ",")",")"),X=Math.SQRT2,$=function(t,n){var e,r,i=t[0],o=t[1],a=t[2],u=n[0],c=n[1],s=n[2],f=u-i,l=c-o,h=f*f+l*l;if(h<1e-12)r=Math.log(s/a)/X,e=function(t){return[i+t*f,o+t*l,a*Math.exp(X*t*r)]};else{var d=Math.sqrt(h),v=(s*s-a*a+4*h)/(2*a*2*d),g=(s*s-a*a-4*h)/(2*s*2*d),m=Math.log(Math.sqrt(v*v+1)-v),b=Math.log(Math.sqrt(g*g+1)-g);r=(b-m)/X,e=function(t){var n=t*r,e=p(m),u=a/(2*d)*(e*y(X*n+m)-_(m));return[i+u*f,o+u*l,a*e/p(X*n+m)]}}return e.duration=1e3*r,e},W=v(o),V=v(u),Z=m(o),Q=m(u),J=b(o),G=b(u),K=function(t,n){for(var e=new Array(n),r=0;r<n;++r)e[r]=t(r/(n-1));return e};t.interpolate=I,t.interpolateArray=O,t.interpolateBasis=k,t.interpolateBasisClosed=A,t.interpolateDate=P,t.interpolateNumber=D,t.interpolateObject=q,t.interpolateRound=z,t.interpolateString=F,t.interpolateTransformCss=R,t.interpolateTransformSvg=H,t.interpolateZoom=$,t.interpolateRgb=S,t.interpolateRgbBasis=C,t.interpolateRgbBasisClosed=E,t.interpolateHsl=W,t.interpolateHslLong=V,t.interpolateLab=g,t.interpolateHcl=Z,t.interpolateHclLong=Q,t.interpolateCubehelix=J,t.interpolateCubehelixLong=G,t.quantize=K,Object.defineProperty(t,"__esModule",{value:!0})})},{"d3-color":4}],10:[function(t,n,e){!function(t,r){"object"==typeof e&&void 0!==n?r(e):"function"==typeof define&&define.amd?define(["exports"],r):r(t.d3=t.d3||{})}(this,function(t){"use strict";function n(){this._x0=this._y0=this._x1=this._y1=null,this._=""}function e(){return new n}var r=Math.PI,i=2*r,o=i-1e-6;n.prototype=e.prototype={constructor:n,moveTo:function(t,n){this._+="M"+(this._x0=this._x1=+t)+","+(this._y0=this._y1=+n)},closePath:function(){null!==this._x1&&(this._x1=this._x0,this._y1=this._y0,this._+="Z")},lineTo:function(t,n){this._+="L"+(this._x1=+t)+","+(this._y1=+n)},quadraticCurveTo:function(t,n,e,r){this._+="Q"+ +t+","+ +n+","+(this._x1=+e)+","+(this._y1=+r)},bezierCurveTo:function(t,n,e,r,i,o){this._+="C"+ +t+","+ +n+","+ +e+","+ +r+","+(this._x1=+i)+","+(this._y1=+o)},arcTo:function(t,n,e,i,o){t=+t,n=+n,e=+e,i=+i,o=+o;var a=this._x1,u=this._y1,c=e-t,s=i-n,f=a-t,l=u-n,h=f*f+l*l;if(o<0)throw new Error("negative radius: "+o);if(null===this._x1)this._+="M"+(this._x1=t)+","+(this._y1=n);else if(h>1e-6)if(Math.abs(l*c-s*f)>1e-6&&o){var d=e-a,p=i-u,_=c*c+s*s,y=d*d+p*p,v=Math.sqrt(_),g=Math.sqrt(h),m=o*Math.tan((r-Math.acos((_+h-y)/(2*v*g)))/2),b=m/g,x=m/v;Math.abs(b-1)>1e-6&&(this._+="L"+(t+b*f)+","+(n+b*l)),this._+="A"+o+","+o+",0,0,"+ +(l*d>f*p)+","+(this._x1=t+x*c)+","+(this._y1=n+x*s)}else this._+="L"+(this._x1=t)+","+(this._y1=n);else;},arc:function(t,n,e,a,u,c){t=+t,n=+n;var s=(e=+e)*Math.cos(a),f=e*Math.sin(a),l=t+s,h=n+f,d=1^c,p=c?a-u:u-a;if(e<0)throw new Error("negative radius: "+e);null===this._x1?this._+="M"+l+","+h:(Math.abs(this._x1-l)>1e-6||Math.abs(this._y1-h)>1e-6)&&(this._+="L"+l+","+h),e&&(p<0&&(p=p%i+i),p>o?this._+="A"+e+","+e+",0,1,"+d+","+(t-s)+","+(n-f)+"A"+e+","+e+",0,1,"+d+","+(this._x1=l)+","+(this._y1=h):p>1e-6&&(this._+="A"+e+","+e+",0,"+ +(p>=r)+","+d+","+(this._x1=t+e*Math.cos(u))+","+(this._y1=n+e*Math.sin(u))))},rect:function(t,n,e,r){this._+="M"+(this._x0=this._x1=+t)+","+(this._y0=this._y1=+n)+"h"+ +e+"v"+ +r+"h"+-e+"Z"},toString:function(){return this._}},t.path=e,Object.defineProperty(t,"__esModule",{value:!0})})},{}],11:[function(t,n,e){!function(r,i){"object"==typeof e&&void 0!==n?i(e,t("d3-array"),t("d3-collection"),t("d3-interpolate"),t("d3-format"),t("d3-time"),t("d3-time-format"),t("d3-color")):"function"==typeof define&&define.amd?define(["exports","d3-array","d3-collection","d3-interpolate","d3-format","d3-time","d3-time-format","d3-color"],i):i(r.d3=r.d3||{},r.d3,r.d3,r.d3,r.d3,r.d3,r.d3,r.d3)}(this,function(t,n,e,r,i,o,a,u){"use strict";function c(t){function n(n){var e=n+"",a=r.get(e);if(!a){if(o!==R)return o;r.set(e,a=i.push(n))}return t[(a-1)%t.length]}var r=e.map(),i=[],o=R;return t=null==t?[]:B.call(t),n.domain=function(t){if(!arguments.length)return i.slice();i=[],r=e.map();for(var o,a,u=-1,c=t.length;++u<c;)r.has(a=(o=t[u])+"")||r.set(a,i.push(o));return n},n.range=function(e){return arguments.length?(t=B.call(e),n):t.slice()},n.unknown=function(t){return arguments.length?(o=t,n):o},n.copy=function(){return c().domain(i).range(t).unknown(o)},n}function s(){function t(){var t=o().length,i=u[1]<u[0],c=u[i-0],s=u[1-i];e=(s-c)/Math.max(1,t-l+2*h),f&&(e=Math.floor(e)),c+=(s-c-e*(t-l))*d,r=e*(1-l),f&&(c=Math.round(c),r=Math.round(r));var p=n.range(t).map(function(t){return c+e*t});return a(i?p.reverse():p)}var e,r,i=c().unknown(void 0),o=i.domain,a=i.range,u=[0,1],f=!1,l=0,h=0,d=.5;return delete i.unknown,i.domain=function(n){return arguments.length?(o(n),t()):o()},i.range=function(n){return arguments.length?(u=[+n[0],+n[1]],t()):u.slice()},i.rangeRound=function(n){return u=[+n[0],+n[1]],f=!0,t()},i.bandwidth=function(){return r},i.step=function(){return e},i.round=function(n){return arguments.length?(f=!!n,t()):f},i.padding=function(n){return arguments.length?(l=h=Math.max(0,Math.min(1,n)),t()):l},i.paddingInner=function(n){return arguments.length?(l=Math.max(0,Math.min(1,n)),t()):l},i.paddingOuter=function(n){return arguments.length?(h=Math.max(0,Math.min(1,n)),t()):h},i.align=function(n){return arguments.length?(d=Math.max(0,Math.min(1,n)),t()):d},i.copy=function(){return s().domain(o()).range(u).round(f).paddingInner(l).paddingOuter(h).align(d)},t()}function f(t){var n=t.copy;return t.padding=t.paddingOuter,delete t.paddingInner,delete t.paddingOuter,t.copy=function(){return f(n())},t}function l(){return f(s().paddingInner(1))}function h(t,n){return(n-=t=+t)?function(e){return(e-t)/n}:H(n)}function d(t){return function(n,e){var r=t(n=+n,e=+e);return function(t){return t<=n?0:t>=e?1:r(t)}}}function p(t){return function(n,e){var r=t(n=+n,e=+e);return function(t){return t<=0?n:t>=1?e:r(t)}}}function _(t,n,e,r){var i=t[0],o=t[1],a=n[0],u=n[1];return o<i?(i=e(o,i),a=r(u,a)):(i=e(i,o),a=r(a,u)),function(t){return a(i(t))}}function y(t,e,r,i){var o=Math.min(t.length,e.length)-1,a=new Array(o),u=new Array(o),c=-1;for(t[o]<t[0]&&(t=t.slice().reverse(),e=e.slice().reverse());++c<o;)a[c]=r(t[c],t[c+1]),u[c]=i(e[c],e[c+1]);return function(e){var r=n.bisect(t,e,1,o)-1;return u[r](a[r](e))}}function v(t,n){return n.domain(t.domain()).range(t.range()).interpolate(t.interpolate()).clamp(t.clamp())}function g(t,n){function e(){return o=Math.min(c.length,s.length)>2?y:_,a=u=null,i}function i(n){return(a||(a=o(c,s,l?d(t):t,f)))(+n)}var o,a,u,c=$,s=$,f=r.interpolate,l=!1;return i.invert=function(t){return(u||(u=o(s,c,h,l?p(n):n)))(+t)},i.domain=function(t){return arguments.length?(c=L.call(t,X),e()):c.slice()},i.range=function(t){return arguments.length?(s=B.call(t),e()):s.slice()},i.rangeRound=function(t){return s=B.call(t),f=r.interpolateRound,e()},i.clamp=function(t){return arguments.length?(l=!!t,e()):l},i.interpolate=function(t){return arguments.length?(f=t,e()):f},e()}function m(t){var e=t.domain;return t.ticks=function(t){var r=e();return n.ticks(r[0],r[r.length-1],null==t?10:t)},t.tickFormat=function(t,n){return W(e(),t,n)},t.nice=function(r){null==r&&(r=10);var i,o=e(),a=0,u=o.length-1,c=o[a],s=o[u];return s<c&&(i=c,c=s,s=i,i=a,a=u,u=i),i=n.tickIncrement(c,s,r),i>0?(c=Math.floor(c/i)*i,s=Math.ceil(s/i)*i,i=n.tickIncrement(c,s,r)):i<0&&(c=Math.ceil(c*i)/i,s=Math.floor(s*i)/i,i=n.tickIncrement(c,s,r)),i>0?(o[a]=Math.floor(c/i)*i,o[u]=Math.ceil(s/i)*i,e(o)):i<0&&(o[a]=Math.ceil(c*i)/i,o[u]=Math.floor(s*i)/i,e(o)),t},t}function b(){var t=g(h,r.interpolateNumber);return t.copy=function(){return v(t,b())},m(t)}function x(){function t(t){return+t}var n=[0,1];return t.invert=t,t.domain=t.range=function(e){return arguments.length?(n=L.call(e,X),t):n.slice()},t.copy=function(){return x().domain(n)},m(t)}function w(t,n){return(n=Math.log(n/t))?function(e){return Math.log(e/t)/n}:H(n)}function M(t,n){return t<0?function(e){return-Math.pow(-n,e)*Math.pow(-t,1-e)}:function(e){return Math.pow(n,e)*Math.pow(t,1-e)}}function T(t){return isFinite(t)?+("1e"+t):t<0?0:t}function k(t){return 10===t?T:t===Math.E?Math.exp:function(n){return Math.pow(t,n)}}function A(t){return t===Math.E?Math.log:10===t&&Math.log10||2===t&&Math.log2||(t=Math.log(t),function(n){return Math.log(n)/t})}function N(t){return function(n){return-t(-n)}}function S(){function t(){return a=A(o),u=k(o),r()[0]<0&&(a=N(a),u=N(u)),e}var e=g(w,M).domain([1,10]),r=e.domain,o=10,a=A(10),u=k(10);return e.base=function(n){return arguments.length?(o=+n,t()):o},e.domain=function(n){return arguments.length?(r(n),t()):r()},e.ticks=function(t){var e,i=r(),c=i[0],s=i[i.length-1];(e=s<c)&&(d=c,c=s,s=d);var f,l,h,d=a(c),p=a(s),_=null==t?10:+t,y=[];if(!(o%1)&&p-d<_){if(d=Math.round(d)-1,p=Math.round(p)+1,c>0){for(;d<p;++d)for(l=1,f=u(d);l<o;++l)if(!((h=f*l)<c)){if(h>s)break;y.push(h)}}else for(;d<p;++d)for(l=o-1,f=u(d);l>=1;--l)if(!((h=f*l)<c)){if(h>s)break;y.push(h)}}else y=n.ticks(d,p,Math.min(p-d,_)).map(u);return e?y.reverse():y},e.tickFormat=function(t,n){if(null==n&&(n=10===o?".0e":","),"function"!=typeof n&&(n=i.format(n)),t===1/0)return n;null==t&&(t=10);var r=Math.max(1,o*t/e.ticks().length);return function(t){var e=t/u(Math.round(a(t)));return e*o<o-.5&&(e*=o),e<=r?n(t):""}},e.nice=function(){return r(V(r(),{floor:function(t){return u(Math.floor(a(t)))},ceil:function(t){return u(Math.ceil(a(t)))}}))},e.copy=function(){return v(e,S().base(o))},e}function C(t,n){return t<0?-Math.pow(-t,n):Math.pow(t,n)}function E(){function t(t,n){return(n=C(n,e)-(t=C(t,e)))?function(r){return(C(r,e)-t)/n}:H(n)}function n(t,n){return n=C(n,e)-(t=C(t,e)),function(r){return C(t+n*r,1/e)}}var e=1,r=g(t,n),i=r.domain;return r.exponent=function(t){return arguments.length?(e=+t,i(i())):e},r.copy=function(){return v(r,E().exponent(e))},m(r)}function O(){return E().exponent(.5)}function P(){function t(){var t=0,a=Math.max(1,i.length);for(o=new Array(a-1);++t<a;)o[t-1]=n.quantile(r,t/a);return e}function e(t){if(!isNaN(t=+t))return i[n.bisect(o,t)]}var r=[],i=[],o=[];return e.invertExtent=function(t){var n=i.indexOf(t);return n<0?[NaN,NaN]:[n>0?o[n-1]:r[0],n<o.length?o[n]:r[r.length-1]]},e.domain=function(e){if(!arguments.length)return r.slice();r=[];for(var i,o=0,a=e.length;o<a;++o)null==(i=e[o])||isNaN(i=+i)||r.push(i);return r.sort(n.ascending),t()},e.range=function(n){return arguments.length?(i=B.call(n),t()):i.slice()},e.quantiles=function(){return o.slice()},e.copy=function(){return P().domain(r).range(i)},e}function D(){function t(t){if(t<=t)return u[n.bisect(a,t,0,o)]}function e(){var n=-1;for(a=new Array(o);++n<o;)a[n]=((n+1)*i-(n-o)*r)/(o+1);return t}var r=0,i=1,o=1,a=[.5],u=[0,1];return t.domain=function(t){return arguments.length?(r=+t[0],i=+t[1],e()):[r,i]},t.range=function(t){return arguments.length?(o=(u=B.call(t)).length-1,e()):u.slice()},t.invertExtent=function(t){var n=u.indexOf(t);return n<0?[NaN,NaN]:n<1?[r,a[0]]:n>=o?[a[o-1],i]:[a[n-1],a[n]]},t.copy=function(){return D().domain([r,i]).range(u)},m(t)}function q(){function t(t){if(t<=t)return r[n.bisect(e,t,0,i)]}var e=[.5],r=[0,1],i=1;return t.domain=function(n){return arguments.length?(e=B.call(n),i=Math.min(e.length,r.length-1),t):e.slice()},t.range=function(n){return arguments.length?(r=B.call(n),i=Math.min(e.length,r.length-1),t):r.slice()},t.invertExtent=function(t){var n=r.indexOf(t);return[e[n-1],e[n]]},t.copy=function(){return q().domain(e).range(r)},t}function U(t){return new Date(t)}function j(t){return t instanceof Date?+t:+new Date(+t)}function F(t,e,i,o,a,u,c,s,f){function l(n){return(c(n)<n?m:u(n)<n?b:a(n)<n?x:o(n)<n?w:e(n)<n?i(n)<n?M:T:t(n)<n?k:A)(n)}function d(e,r,i,o){if(null==e&&(e=10),"number"==typeof e){var a=Math.abs(i-r)/e,u=n.bisector(function(t){return t[2]}).right(N,a);u===N.length?(o=n.tickStep(r/nt,i/nt,e),e=t):u?(o=(u=N[a/N[u-1][2]<N[u][2]/a?u-1:u])[1],e=u[0]):(o=n.tickStep(r,i,e),e=s)}return null==o?e:e.every(o)}var p=g(h,r.interpolateNumber),_=p.invert,y=p.domain,m=f(".%L"),b=f(":%S"),x=f("%I:%M"),w=f("%I %p"),M=f("%a %d"),T=f("%b %d"),k=f("%B"),A=f("%Y"),N=[[c,1,Z],[c,5,5*Z],[c,15,15*Z],[c,30,30*Z],[u,1,Q],[u,5,5*Q],[u,15,15*Q],[u,30,30*Q],[a,1,J],[a,3,3*J],[a,6,6*J],[a,12,12*J],[o,1,G],[o,2,2*G],[i,1,K],[e,1,tt],[e,3,3*tt],[t,1,nt]];return p.invert=function(t){return new Date(_(t))},p.domain=function(t){return arguments.length?y(L.call(t,j)):y().map(U)},p.ticks=function(t,n){var e,r=y(),i=r[0],o=r[r.length-1],a=o<i;return a&&(e=i,i=o,o=e),e=d(t,i,o,n),e=e?e.range(i,o+1):[],a?e.reverse():e},p.tickFormat=function(t,n){return null==n?l:f(n)},p.nice=function(t,n){var e=y();return(t=d(t,e[0],e[e.length-1],n))?y(V(e,t)):p},p.copy=function(){return v(p,F(t,e,i,o,a,u,c,s,f))},p}function I(t){var n=t.length;return function(e){return t[Math.max(0,Math.min(n-1,Math.floor(e*n)))]}}function z(t){function n(n){var o=(n-e)/(r-e);return t(i?Math.max(0,Math.min(1,o)):o)}var e=0,r=1,i=!1;return n.domain=function(t){return arguments.length?(e=+t[0],r=+t[1],n):[e,r]},n.clamp=function(t){return arguments.length?(i=!!t,n):i},n.interpolator=function(e){return arguments.length?(t=e,n):t},n.copy=function(){return z(t).domain([e,r]).clamp(i)},m(n)}var Y=Array.prototype,L=Y.map,B=Y.slice,R={name:"implicit"},H=function(t){return function(){return t}},X=function(t){return+t},$=[0,1],W=function(t,e,r){var o,a=t[0],u=t[t.length-1],c=n.tickStep(a,u,null==e?10:e);switch((r=i.formatSpecifier(null==r?",f":r)).type){case"s":var s=Math.max(Math.abs(a),Math.abs(u));return null!=r.precision||isNaN(o=i.precisionPrefix(c,s))||(r.precision=o),i.formatPrefix(r,s);case"":case"e":case"g":case"p":case"r":null!=r.precision||isNaN(o=i.precisionRound(c,Math.max(Math.abs(a),Math.abs(u))))||(r.precision=o-("e"===r.type));break;case"f":case"%":null!=r.precision||isNaN(o=i.precisionFixed(c))||(r.precision=o-2*("%"===r.type))}return i.format(r)},V=function(t,n){var e,r=0,i=(t=t.slice()).length-1,o=t[r],a=t[i];return a<o&&(e=r,r=i,i=e,e=o,o=a,a=e),t[r]=n.floor(o),t[i]=n.ceil(a),t},Z=1e3,Q=60*Z,J=60*Q,G=24*J,K=7*G,tt=30*G,nt=365*G,et=function(){return F(o.timeYear,o.timeMonth,o.timeWeek,o.timeDay,o.timeHour,o.timeMinute,o.timeSecond,o.timeMillisecond,a.timeFormat).domain([new Date(2e3,0,1),new Date(2e3,0,2)])},rt=function(){return F(o.utcYear,o.utcMonth,o.utcWeek,o.utcDay,o.utcHour,o.utcMinute,o.utcSecond,o.utcMillisecond,a.utcFormat).domain([Date.UTC(2e3,0,1),Date.UTC(2e3,0,2)])},it=function(t){return t.match(/.{6}/g).map(function(t){return"#"+t})},ot=it("1f77b4ff7f0e2ca02cd627289467bd8c564be377c27f7f7fbcbd2217becf"),at=it("393b795254a36b6ecf9c9ede6379398ca252b5cf6bcedb9c8c6d31bd9e39e7ba52e7cb94843c39ad494ad6616be7969c7b4173a55194ce6dbdde9ed6"),ut=it("3182bd6baed69ecae1c6dbefe6550dfd8d3cfdae6bfdd0a231a35474c476a1d99bc7e9c0756bb19e9ac8bcbddcdadaeb636363969696bdbdbdd9d9d9"),ct=it("1f77b4aec7e8ff7f0effbb782ca02c98df8ad62728ff98969467bdc5b0d58c564bc49c94e377c2f7b6d27f7f7fc7c7c7bcbd22dbdb8d17becf9edae5"),st=r.interpolateCubehelixLong(u.cubehelix(300,.5,0),u.cubehelix(-240,.5,1)),ft=r.interpolateCubehelixLong(u.cubehelix(-100,.75,.35),u.cubehelix(80,1.5,.8)),lt=r.interpolateCubehelixLong(u.cubehelix(260,.75,.35),u.cubehelix(80,1.5,.8)),ht=u.cubehelix(),dt=function(t){(t<0||t>1)&&(t-=Math.floor(t));var n=Math.abs(t-.5);return ht.h=360*t-100,ht.s=1.5-1.5*n,ht.l=.8-.9*n,ht+""},pt=I(it("44015444025645045745055946075a46085c460a5d460b5e470d60470e6147106347116447136548146748166848176948186a481a6c481b6d481c6e481d6f481f70482071482173482374482475482576482677482878482979472a7a472c7a472d7b472e7c472f7d46307e46327e46337f463480453581453781453882443983443a83443b84433d84433e85423f854240864241864142874144874045884046883f47883f48893e49893e4a893e4c8a3d4d8a3d4e8a3c4f8a3c508b3b518b3b528b3a538b3a548c39558c39568c38588c38598c375a8c375b8d365c8d365d8d355e8d355f8d34608d34618d33628d33638d32648e32658e31668e31678e31688e30698e306a8e2f6b8e2f6c8e2e6d8e2e6e8e2e6f8e2d708e2d718e2c718e2c728e2c738e2b748e2b758e2a768e2a778e2a788e29798e297a8e297b8e287c8e287d8e277e8e277f8e27808e26818e26828e26828e25838e25848e25858e24868e24878e23888e23898e238a8d228b8d228c8d228d8d218e8d218f8d21908d21918c20928c20928c20938c1f948c1f958b1f968b1f978b1f988b1f998a1f9a8a1e9b8a1e9c891e9d891f9e891f9f881fa0881fa1881fa1871fa28720a38620a48621a58521a68522a78522a88423a98324aa8325ab8225ac8226ad8127ad8128ae8029af7f2ab07f2cb17e2db27d2eb37c2fb47c31b57b32b67a34b67935b77937b87838b9773aba763bbb753dbc743fbc7340bd7242be7144bf7046c06f48c16e4ac16d4cc26c4ec36b50c46a52c56954c56856c66758c7655ac8645cc8635ec96260ca6063cb5f65cb5e67cc5c69cd5b6ccd5a6ece5870cf5773d05675d05477d1537ad1517cd2507fd34e81d34d84d44b86d54989d5488bd6468ed64590d74393d74195d84098d83e9bd93c9dd93ba0da39a2da37a5db36a8db34aadc32addc30b0dd2fb2dd2db5de2bb8de29bade28bddf26c0df25c2df23c5e021c8e020cae11fcde11dd0e11cd2e21bd5e21ad8e219dae319dde318dfe318e2e418e5e419e7e419eae51aece51befe51cf1e51df4e61ef6e620f8e621fbe723fde725")),_t=I(it("00000401000501010601010802010902020b02020d03030f03031204041405041606051806051a07061c08071e0907200a08220b09240c09260d0a290e0b2b100b2d110c2f120d31130d34140e36150e38160f3b180f3d19103f1a10421c10441d11471e114920114b21114e22115024125325125527125829115a2a115c2c115f2d11612f116331116533106734106936106b38106c390f6e3b0f703d0f713f0f72400f74420f75440f764510774710784910784a10794c117a4e117b4f127b51127c52137c54137d56147d57157e59157e5a167e5c167f5d177f5f187f601880621980641a80651a80671b80681c816a1c816b1d816d1d816e1e81701f81721f817320817521817621817822817922827b23827c23827e24828025828125818326818426818627818827818928818b29818c29818e2a81902a81912b81932b80942c80962c80982d80992d809b2e7f9c2e7f9e2f7fa02f7fa1307ea3307ea5317ea6317da8327daa337dab337cad347cae347bb0357bb2357bb3367ab5367ab73779b83779ba3878bc3978bd3977bf3a77c03a76c23b75c43c75c53c74c73d73c83e73ca3e72cc3f71cd4071cf4070d0416fd2426fd3436ed5446dd6456cd8456cd9466bdb476adc4869de4968df4a68e04c67e24d66e34e65e44f64e55064e75263e85362e95462ea5661eb5760ec5860ed5a5fee5b5eef5d5ef05f5ef1605df2625df2645cf3655cf4675cf4695cf56b5cf66c5cf66e5cf7705cf7725cf8745cf8765cf9785df9795df97b5dfa7d5efa7f5efa815ffb835ffb8560fb8761fc8961fc8a62fc8c63fc8e64fc9065fd9266fd9467fd9668fd9869fd9a6afd9b6bfe9d6cfe9f6dfea16efea36ffea571fea772fea973feaa74feac76feae77feb078feb27afeb47bfeb67cfeb77efeb97ffebb81febd82febf84fec185fec287fec488fec68afec88cfeca8dfecc8ffecd90fecf92fed194fed395fed597fed799fed89afdda9cfddc9efddea0fde0a1fde2a3fde3a5fde5a7fde7a9fde9aafdebacfcecaefceeb0fcf0b2fcf2b4fcf4b6fcf6b8fcf7b9fcf9bbfcfbbdfcfdbf")),yt=I(it("00000401000501010601010802010a02020c02020e03021004031204031405041706041907051b08051d09061f0a07220b07240c08260d08290e092b10092d110a30120a32140b34150b37160b39180c3c190c3e1b0c411c0c431e0c451f0c48210c4a230c4c240c4f260c51280b53290b552b0b572d0b592f0a5b310a5c320a5e340a5f3609613809623909633b09643d09653e0966400a67420a68440a68450a69470b6a490b6a4a0c6b4c0c6b4d0d6c4f0d6c510e6c520e6d540f6d550f6d57106e59106e5a116e5c126e5d126e5f136e61136e62146e64156e65156e67166e69166e6a176e6c186e6d186e6f196e71196e721a6e741a6e751b6e771c6d781c6d7a1d6d7c1d6d7d1e6d7f1e6c801f6c82206c84206b85216b87216b88226a8a226a8c23698d23698f24699025689225689326679526679727669827669a28659b29649d29649f2a63a02a63a22b62a32c61a52c60a62d60a82e5fa92e5eab2f5ead305dae305cb0315bb1325ab3325ab43359b63458b73557b93556ba3655bc3754bd3853bf3952c03a51c13a50c33b4fc43c4ec63d4dc73e4cc83f4bca404acb4149cc4248ce4347cf4446d04545d24644d34743d44842d54a41d74b3fd84c3ed94d3dda4e3cdb503bdd513ade5238df5337e05536e15635e25734e35933e45a31e55c30e65d2fe75e2ee8602de9612bea632aeb6429eb6628ec6726ed6925ee6a24ef6c23ef6e21f06f20f1711ff1731df2741cf3761bf37819f47918f57b17f57d15f67e14f68013f78212f78410f8850ff8870ef8890cf98b0bf98c0af98e09fa9008fa9207fa9407fb9606fb9706fb9906fb9b06fb9d07fc9f07fca108fca309fca50afca60cfca80dfcaa0ffcac11fcae12fcb014fcb216fcb418fbb61afbb81dfbba1ffbbc21fbbe23fac026fac228fac42afac62df9c72ff9c932f9cb35f8cd37f8cf3af7d13df7d340f6d543f6d746f5d949f5db4cf4dd4ff4df53f4e156f3e35af3e55df2e661f2e865f2ea69f1ec6df1ed71f1ef75f1f179f2f27df2f482f3f586f3f68af4f88ef5f992f6fa96f8fb9af9fc9dfafda1fcffa4")),vt=I(it("0d088710078813078916078a19068c1b068d1d068e20068f2206902406912605912805922a05932c05942e05952f059631059733059735049837049938049a3a049a3c049b3e049c3f049c41049d43039e44039e46039f48039f4903a04b03a14c02a14e02a25002a25102a35302a35502a45601a45801a45901a55b01a55c01a65e01a66001a66100a76300a76400a76600a76700a86900a86a00a86c00a86e00a86f00a87100a87201a87401a87501a87701a87801a87a02a87b02a87d03a87e03a88004a88104a78305a78405a78606a68707a68808a68a09a58b0aa58d0ba58e0ca48f0da4910ea3920fa39410a29511a19613a19814a099159f9a169f9c179e9d189d9e199da01a9ca11b9ba21d9aa31e9aa51f99a62098a72197a82296aa2395ab2494ac2694ad2793ae2892b02991b12a90b22b8fb32c8eb42e8db52f8cb6308bb7318ab83289ba3388bb3488bc3587bd3786be3885bf3984c03a83c13b82c23c81c33d80c43e7fc5407ec6417dc7427cc8437bc9447aca457acb4679cc4778cc4977cd4a76ce4b75cf4c74d04d73d14e72d24f71d35171d45270d5536fd5546ed6556dd7566cd8576bd9586ada5a6ada5b69db5c68dc5d67dd5e66de5f65de6164df6263e06363e16462e26561e26660e3685fe4695ee56a5de56b5de66c5ce76e5be76f5ae87059e97158e97257ea7457eb7556eb7655ec7754ed7953ed7a52ee7b51ef7c51ef7e50f07f4ff0804ef1814df1834cf2844bf3854bf3874af48849f48948f58b47f58c46f68d45f68f44f79044f79143f79342f89441f89540f9973ff9983ef99a3efa9b3dfa9c3cfa9e3bfb9f3afba139fba238fca338fca537fca636fca835fca934fdab33fdac33fdae32fdaf31fdb130fdb22ffdb42ffdb52efeb72dfeb82cfeba2cfebb2bfebd2afebe2afec029fdc229fdc328fdc527fdc627fdc827fdca26fdcb26fccd25fcce25fcd025fcd225fbd324fbd524fbd724fad824fada24f9dc24f9dd25f8df25f8e125f7e225f7e425f6e626f6e826f5e926f5eb27f4ed27f3ee27f3f027f2f227f1f426f1f525f0f724f0f921"));t.scaleBand=s,t.scalePoint=l,t.scaleIdentity=x,t.scaleLinear=b,t.scaleLog=S,t.scaleOrdinal=c,t.scaleImplicit=R,t.scalePow=E,t.scaleSqrt=O,t.scaleQuantile=P,t.scaleQuantize=D,t.scaleThreshold=q,t.scaleTime=et,t.scaleUtc=rt,t.schemeCategory10=ot,t.schemeCategory20b=at,t.schemeCategory20c=ut,t.schemeCategory20=ct,t.interpolateCubehelixDefault=st,t.interpolateRainbow=dt,t.interpolateWarm=ft,t.interpolateCool=lt,t.interpolateViridis=pt,t.interpolateMagma=_t,t.interpolateInferno=yt,t.interpolatePlasma=vt,t.scaleSequential=z,Object.defineProperty(t,"__esModule",{value:!0})})},{"d3-array":2,"d3-collection":3,"d3-color":4,"d3-format":7,"d3-interpolate":9,"d3-time":15,"d3-time-format":14}],12:[function(t,n,e){!function(t,r){"object"==typeof e&&void 0!==n?r(e):"function"==typeof define&&define.amd?define(["exports"],r):r(t.d3=t.d3||{})}(this,function(t){"use strict";function n(t){return function(){var n=this.ownerDocument,e=this.namespaceURI;return e===K&&n.documentElement.namespaceURI===K?n.createElement(t):n.createElementNS(e,t)}}function e(t){return function(){return this.ownerDocument.createElementNS(t.space,t.local)}}function r(){return new i}function i(){this._="@"+(++rt).toString(36)}function o(t,n,e){return t=a(t,n,e),function(n){var e=n.relatedTarget;e&&(e===this||8&e.compareDocumentPosition(this))||t.call(this,n)}}function a(n,e,r){return function(i){var o=t.event;t.event=i;try{n.call(this,this.__data__,e,r)}finally{t.event=o}}}function u(t){return t.trim().split(/^|\s+/).map(function(t){var n="",e=t.indexOf(".");return e>=0&&(n=t.slice(e+1),t=t.slice(0,e)),{type:t,name:n}})}function c(t){return function(){var n=this.__on;if(n){for(var e,r=0,i=-1,o=n.length;r<o;++r)e=n[r],t.type&&e.type!==t.type||e.name!==t.name?n[++i]=e:this.removeEventListener(e.type,e.listener,e.capture);++i?n.length=i:delete this.__on}}}function s(t,n,e){var r=ct.hasOwnProperty(t.type)?o:a;return function(i,o,a){var u,c=this.__on,s=r(n,o,a);if(c)for(var f=0,l=c.length;f<l;++f)if((u=c[f]).type===t.type&&u.name===t.name)return this.removeEventListener(u.type,u.listener,u.capture),this.addEventListener(u.type,u.listener=s,u.capture=e),void(u.value=n);this.addEventListener(t.type,s,e),u={type:t.type,name:t.name,value:n,listener:s,capture:e},c?c.push(u):this.__on=[u]}}function f(n,e,r,i){var o=t.event;n.sourceEvent=t.event,t.event=n;try{return e.apply(r,i)}finally{t.event=o}}function l(){}function h(){return[]}function d(t,n){this.ownerDocument=t.ownerDocument,this.namespaceURI=t.namespaceURI,this._next=null,this._parent=t,this.__data__=n}function p(t,n,e,r,i,o){for(var a,u=0,c=n.length,s=o.length;u<s;++u)(a=n[u])?(a.__data__=o[u],r[u]=a):e[u]=new d(t,o[u]);for(;u<c;++u)(a=n[u])&&(i[u]=a)}function _(t,n,e,r,i,o,a){var u,c,s,f={},l=n.length,h=o.length,p=new Array(l);for(u=0;u<l;++u)(c=n[u])&&(p[u]=s=xt+a.call(c,c.__data__,u,n),s in f?i[u]=c:f[s]=c);for(u=0;u<h;++u)(c=f[s=xt+a.call(t,o[u],u,o)])?(r[u]=c,c.__data__=o[u],f[s]=null):e[u]=new d(t,o[u]);for(u=0;u<l;++u)(c=n[u])&&f[p[u]]===c&&(i[u]=c)}function y(t,n){return t<n?-1:t>n?1:t>=n?0:NaN}function v(t){return function(){this.removeAttribute(t)}}function g(t){return function(){this.removeAttributeNS(t.space,t.local)}}function m(t,n){return function(){this.setAttribute(t,n)}}function b(t,n){return function(){this.setAttributeNS(t.space,t.local,n)}}function x(t,n){return function(){var e=n.apply(this,arguments);null==e?this.removeAttribute(t):this.setAttribute(t,e)}}function w(t,n){return function(){var e=n.apply(this,arguments);null==e?this.removeAttributeNS(t.space,t.local):this.setAttributeNS(t.space,t.local,e)}}function M(t){return function(){this.style.removeProperty(t)}}function T(t,n,e){return function(){this.style.setProperty(t,n,e)}}function k(t,n,e){return function(){var r=n.apply(this,arguments);null==r?this.style.removeProperty(t):this.style.setProperty(t,r,e)}}function A(t,n){return t.style.getPropertyValue(n)||qt(t).getComputedStyle(t,null).getPropertyValue(n)}function N(t){return function(){delete this[t]}}function S(t,n){return function(){this[t]=n}}function C(t,n){return function(){var e=n.apply(this,arguments);null==e?delete this[t]:this[t]=e}}function E(t){return t.trim().split(/^|\s+/)}function O(t){return t.classList||new P(t)}function P(t){this._node=t,this._names=E(t.getAttribute("class")||"")}function D(t,n){for(var e=O(t),r=-1,i=n.length;++r<i;)e.add(n[r])}function q(t,n){for(var e=O(t),r=-1,i=n.length;++r<i;)e.remove(n[r])}function U(t){return function(){D(this,t)}}function j(t){return function(){q(this,t)}}function F(t,n){return function(){(n.apply(this,arguments)?D:q)(this,t)}}function I(){this.textContent=""}function z(t){return function(){this.textContent=t}}function Y(t){return function(){var n=t.apply(this,arguments);this.textContent=null==n?"":n}}function L(){this.innerHTML=""}function B(t){return function(){this.innerHTML=t}}function R(t){return function(){var n=t.apply(this,arguments);this.innerHTML=null==n?"":n}}function H(){this.nextSibling&&this.parentNode.appendChild(this)}function X(){this.previousSibling&&this.parentNode.insertBefore(this,this.parentNode.firstChild)}function $(){return null}function W(){var t=this.parentNode;t&&t.removeChild(this)}function V(t,n,e){var r=qt(t),i=r.CustomEvent;"function"==typeof i?i=new i(n,e):(i=r.document.createEvent("Event"),e?(i.initEvent(n,e.bubbles,e.cancelable),i.detail=e.detail):i.initEvent(n,!1,!1)),t.dispatchEvent(i)}function Z(t,n){return function(){return V(this,t,n)}}function Q(t,n){return function(){return V(this,t,n.apply(this,arguments))}}function J(t,n){this._groups=t,this._parents=n}function G(){return new J([[document.documentElement]],Wt)}var K="http://www.w3.org/1999/xhtml",tt={svg:"http://www.w3.org/2000/svg",xhtml:K,xlink:"http://www.w3.org/1999/xlink",xml:"http://www.w3.org/XML/1998/namespace",xmlns:"http://www.w3.org/2000/xmlns/"},nt=function(t){var n=t+="",e=n.indexOf(":");return e>=0&&"xmlns"!==(n=t.slice(0,e))&&(t=t.slice(e+1)),tt.hasOwnProperty(n)?{space:tt[n],local:t}:t},et=function(t){var r=nt(t);return(r.local?e:n)(r)},rt=0;i.prototype=r.prototype={constructor:i,get:function(t){for(var n=this._;!(n in t);)if(!(t=t.parentNode))return;return t[n]},set:function(t,n){return t[this._]=n},remove:function(t){return this._ in t&&delete t[this._]},toString:function(){return this._}};var it=function(t){return function(){return this.matches(t)}};if("undefined"!=typeof document){var ot=document.documentElement;if(!ot.matches){var at=ot.webkitMatchesSelector||ot.msMatchesSelector||ot.mozMatchesSelector||ot.oMatchesSelector;it=function(t){return function(){return at.call(this,t)}}}}var ut=it,ct={};t.event=null,"undefined"!=typeof document&&("onmouseenter"in document.documentElement||(ct={mouseenter:"mouseover",mouseleave:"mouseout"}));var st=function(t,n,e){var r,i,o=u(t+""),a=o.length;{if(!(arguments.length<2)){for(f=n?s:c,null==e&&(e=!1),r=0;r<a;++r)this.each(f(o[r],n,e));return this}var f=this.node().__on;if(f)for(var l,h=0,d=f.length;h<d;++h)for(r=0,l=f[h];r<a;++r)if((i=o[r]).type===l.type&&i.name===l.name)return l.value}},ft=function(){for(var n,e=t.event;n=e.sourceEvent;)e=n;return e},lt=function(t,n){var e=t.ownerSVGElement||t;if(e.createSVGPoint){var r=e.createSVGPoint();return r.x=n.clientX,r.y=n.clientY,r=r.matrixTransform(t.getScreenCTM().inverse()),[r.x,r.y]}var i=t.getBoundingClientRect();return[n.clientX-i.left-t.clientLeft,n.clientY-i.top-t.clientTop]},ht=function(t){var n=ft();return n.changedTouches&&(n=n.changedTouches[0]),lt(t,n)},dt=function(t){return null==t?l:function(){return this.querySelector(t)}},pt=function(t){"function"!=typeof t&&(t=dt(t));for(var n=this._groups,e=n.length,r=new Array(e),i=0;i<e;++i)for(var o,a,u=n[i],c=u.length,s=r[i]=new Array(c),f=0;f<c;++f)(o=u[f])&&(a=t.call(o,o.__data__,f,u))&&("__data__"in o&&(a.__data__=o.__data__),s[f]=a);return new J(r,this._parents)},_t=function(t){return null==t?h:function(){return this.querySelectorAll(t)}},yt=function(t){"function"!=typeof t&&(t=_t(t));for(var n=this._groups,e=n.length,r=[],i=[],o=0;o<e;++o)for(var a,u=n[o],c=u.length,s=0;s<c;++s)(a=u[s])&&(r.push(t.call(a,a.__data__,s,u)),i.push(a));return new J(r,i)},vt=function(t){"function"!=typeof t&&(t=ut(t));for(var n=this._groups,e=n.length,r=new Array(e),i=0;i<e;++i)for(var o,a=n[i],u=a.length,c=r[i]=[],s=0;s<u;++s)(o=a[s])&&t.call(o,o.__data__,s,a)&&c.push(o);return new J(r,this._parents)},gt=function(t){return new Array(t.length)},mt=function(){return new J(this._enter||this._groups.map(gt),this._parents)};d.prototype={constructor:d,appendChild:function(t){return this._parent.insertBefore(t,this._next)},insertBefore:function(t,n){return this._parent.insertBefore(t,n)},querySelector:function(t){return this._parent.querySelector(t)},querySelectorAll:function(t){return this._parent.querySelectorAll(t)}};var bt=function(t){return function(){return t}},xt="$",wt=function(t,n){if(!t)return d=new Array(this.size()),s=-1,this.each(function(t){d[++s]=t}),d;var e=n?_:p,r=this._parents,i=this._groups;"function"!=typeof t&&(t=bt(t));for(var o=i.length,a=new Array(o),u=new Array(o),c=new Array(o),s=0;s<o;++s){var f=r[s],l=i[s],h=l.length,d=t.call(f,f&&f.__data__,s,r),y=d.length,v=u[s]=new Array(y),g=a[s]=new Array(y);e(f,l,v,g,c[s]=new Array(h),d,n);for(var m,b,x=0,w=0;x<y;++x)if(m=v[x]){for(x>=w&&(w=x+1);!(b=g[w])&&++w<y;);m._next=b||null}}return a=new J(a,r),a._enter=u,a._exit=c,a},Mt=function(){return new J(this._exit||this._groups.map(gt),this._parents)},Tt=function(t){for(var n=this._groups,e=t._groups,r=n.length,i=e.length,o=Math.min(r,i),a=new Array(r),u=0;u<o;++u)for(var c,s=n[u],f=e[u],l=s.length,h=a[u]=new Array(l),d=0;d<l;++d)(c=s[d]||f[d])&&(h[d]=c);for(;u<r;++u)a[u]=n[u];return new J(a,this._parents)},kt=function(){for(var t=this._groups,n=-1,e=t.length;++n<e;)for(var r,i=t[n],o=i.length-1,a=i[o];--o>=0;)(r=i[o])&&(a&&a!==r.nextSibling&&a.parentNode.insertBefore(r,a),a=r);return this},At=function(t){function n(n,e){return n&&e?t(n.__data__,e.__data__):!n-!e}t||(t=y);for(var e=this._groups,r=e.length,i=new Array(r),o=0;o<r;++o){for(var a,u=e[o],c=u.length,s=i[o]=new Array(c),f=0;f<c;++f)(a=u[f])&&(s[f]=a);s.sort(n)}return new J(i,this._parents).order()},Nt=function(){var t=arguments[0];return arguments[0]=this,t.apply(null,arguments),this},St=function(){var t=new Array(this.size()),n=-1;return this.each(function(){t[++n]=this}),t},Ct=function(){for(var t=this._groups,n=0,e=t.length;n<e;++n)for(var r=t[n],i=0,o=r.length;i<o;++i){var a=r[i];if(a)return a}return null},Et=function(){var t=0;return this.each(function(){++t}),t},Ot=function(){return!this.node()},Pt=function(t){for(var n=this._groups,e=0,r=n.length;e<r;++e)for(var i,o=n[e],a=0,u=o.length;a<u;++a)(i=o[a])&&t.call(i,i.__data__,a,o);return this},Dt=function(t,n){var e=nt(t);if(arguments.length<2){var r=this.node();return e.local?r.getAttributeNS(e.space,e.local):r.getAttribute(e)}return this.each((null==n?e.local?g:v:"function"==typeof n?e.local?w:x:e.local?b:m)(e,n))},qt=function(t){return t.ownerDocument&&t.ownerDocument.defaultView||t.document&&t||t.defaultView},Ut=function(t,n,e){return arguments.length>1?this.each((null==n?M:"function"==typeof n?k:T)(t,n,null==e?"":e)):A(this.node(),t)},jt=function(t,n){return arguments.length>1?this.each((null==n?N:"function"==typeof n?C:S)(t,n)):this.node()[t]};P.prototype={add:function(t){this._names.indexOf(t)<0&&(this._names.push(t),this._node.setAttribute("class",this._names.join(" ")))},remove:function(t){var n=this._names.indexOf(t);n>=0&&(this._names.splice(n,1),this._node.setAttribute("class",this._names.join(" ")))},contains:function(t){return this._names.indexOf(t)>=0}};var Ft=function(t,n){var e=E(t+"");if(arguments.length<2){for(var r=O(this.node()),i=-1,o=e.length;++i<o;)if(!r.contains(e[i]))return!1;return!0}return this.each(("function"==typeof n?F:n?U:j)(e,n))},It=function(t){return arguments.length?this.each(null==t?I:("function"==typeof t?Y:z)(t)):this.node().textContent},zt=function(t){return arguments.length?this.each(null==t?L:("function"==typeof t?R:B)(t)):this.node().innerHTML},Yt=function(){return this.each(H)},Lt=function(){return this.each(X)},Bt=function(t){var n="function"==typeof t?t:et(t);return this.select(function(){return this.appendChild(n.apply(this,arguments))})},Rt=function(t,n){var e="function"==typeof t?t:et(t),r=null==n?$:"function"==typeof n?n:dt(n);return this.select(function(){return this.insertBefore(e.apply(this,arguments),r.apply(this,arguments)||null)})},Ht=function(){return this.each(W)},Xt=function(t){return arguments.length?this.property("__data__",t):this.node().__data__},$t=function(t,n){return this.each(("function"==typeof n?Q:Z)(t,n))},Wt=[null];J.prototype=G.prototype={constructor:J,select:pt,selectAll:yt,filter:vt,data:wt,enter:mt,exit:Mt,merge:Tt,order:kt,sort:At,call:Nt,nodes:St,node:Ct,size:Et,empty:Ot,each:Pt,attr:Dt,style:Ut,property:jt,classed:Ft,text:It,html:zt,raise:Yt,lower:Lt,append:Bt,insert:Rt,remove:Ht,datum:Xt,on:st,dispatch:$t};var Vt=function(t){return"string"==typeof t?new J([[document.querySelector(t)]],[document.documentElement]):new J([[t]],Wt)},Zt=function(t){return"string"==typeof t?new J([document.querySelectorAll(t)],[document.documentElement]):new J([null==t?[]:t],Wt)},Qt=function(t,n,e){arguments.length<3&&(e=n,n=ft().changedTouches);for(var r,i=0,o=n?n.length:0;i<o;++i)if((r=n[i]).identifier===e)return lt(t,r);return null},Jt=function(t,n){null==n&&(n=ft().touches);for(var e=0,r=n?n.length:0,i=new Array(r);e<r;++e)i[e]=lt(t,n[e]);return i};t.creator=et,t.local=r,t.matcher=ut,t.mouse=ht,t.namespace=nt,t.namespaces=tt,t.select=Vt,t.selectAll=Zt,t.selection=G,t.selector=dt,t.selectorAll=_t,t.style=A,t.touch=Qt,t.touches=Jt,t.window=qt,t.customEvent=f,Object.defineProperty(t,"__esModule",{value:!0})})},{}],13:[function(t,n,e){!function(r,i){"object"==typeof e&&void 0!==n?i(e,t("d3-path")):"function"==typeof define&&define.amd?define(["exports","d3-path"],i):i(r.d3=r.d3||{},r.d3)}(this,function(t,n){"use strict";function e(t){return t>1?0:t<-1?lt:Math.acos(t)}function r(t){return t>=1?ht:t<=-1?-ht:Math.asin(t)}function i(t){return t.innerRadius}function o(t){return t.outerRadius}function a(t){return t.startAngle}function u(t){return t.endAngle}function c(t){return t&&t.padAngle}function s(t,n,e,r,i,o,a,u){var c=e-t,s=r-n,f=a-i,l=u-o,h=(f*(n-o)-l*(t-i))/(l*c-f*s);return[t+h*c,n+h*s]}function f(t,n,e,r,i,o,a){var u=t-e,c=n-r,s=(a?o:-o)/st(u*u+c*c),f=s*c,l=-s*u,h=t+f,d=n+l,p=e+f,_=r+l,y=(h+p)/2,v=(d+_)/2,g=p-h,m=_-d,b=g*g+m*m,x=i-o,w=h*_-p*d,M=(m<0?-1:1)*st(at(0,x*x*b-w*w)),T=(w*m-g*M)/b,k=(-w*g-m*M)/b,A=(w*m+g*M)/b,N=(-w*g+m*M)/b,S=T-y,C=k-v,E=A-y,O=N-v;return S*S+C*C>E*E+O*O&&(T=A,k=N),{cx:T,cy:k,x01:-f,y01:-l,x11:T*(i/x-1),y11:k*(i/x-1)}}function l(t){this._context=t}function h(t){return t[0]}function d(t){return t[1]}function p(t){this._curve=t}function _(t){function n(n){return new p(t(n))}return n._curve=t,n}function y(t){var n=t.curve;return t.angle=t.x,delete t.x,t.radius=t.y,delete t.y,t.curve=function(t){return arguments.length?n(_(t)):n()._curve},t}function v(t){return t.source}function g(t){return t.target}function m(t){function e(){var e,c=kt.call(arguments),s=r.apply(this,c),f=i.apply(this,c);if(u||(u=e=n.path()),t(u,+o.apply(this,(c[0]=s,c)),+a.apply(this,c),+o.apply(this,(c[0]=f,c)),+a.apply(this,c)),e)return u=null,e+""||null}var r=v,i=g,o=h,a=d,u=null;return e.source=function(t){return arguments.length?(r=t,e):r},e.target=function(t){return arguments.length?(i=t,e):i},e.x=function(t){return arguments.length?(o="function"==typeof t?t:et(+t),e):o},e.y=function(t){return arguments.length?(a="function"==typeof t?t:et(+t),e):a},e.context=function(t){return arguments.length?(u=null==t?null:t,e):u},e}function b(t,n,e,r,i){t.moveTo(n,e),t.bezierCurveTo(n=(n+r)/2,e,n,i,r,i)}function x(t,n,e,r,i){t.moveTo(n,e),t.bezierCurveTo(n,e=(e+i)/2,r,e,r,i)}function w(t,n,e,r,i){var o=Tt(n,e),a=Tt(n,e=(e+i)/2),u=Tt(r,e),c=Tt(r,i);t.moveTo(o[0],o[1]),t.bezierCurveTo(a[0],a[1],u[0],u[1],c[0],c[1])}function M(){return m(b)}function T(){return m(x)}function k(){var t=m(w);return t.angle=t.x,delete t.x,t.radius=t.y,delete t.y,t}function A(t,n,e){t._context.bezierCurveTo((2*t._x0+t._x1)/3,(2*t._y0+t._y1)/3,(t._x0+2*t._x1)/3,(t._y0+2*t._y1)/3,(t._x0+4*t._x1+n)/6,(t._y0+4*t._y1+e)/6)}function N(t){this._context=t}function S(t){this._context=t}function C(t){this._context=t}function E(t,n){this._basis=new N(t),this._beta=n}function O(t,n,e){t._context.bezierCurveTo(t._x1+t._k*(t._x2-t._x0),t._y1+t._k*(t._y2-t._y0),t._x2+t._k*(t._x1-n),t._y2+t._k*(t._y1-e),t._x2,t._y2)}function P(t,n){this._context=t,this._k=(1-n)/6}function D(t,n){this._context=t,this._k=(1-n)/6}function q(t,n){this._context=t,this._k=(1-n)/6}function U(t,n,e){var r=t._x1,i=t._y1,o=t._x2,a=t._y2;if(t._l01_a>ft){var u=2*t._l01_2a+3*t._l01_a*t._l12_a+t._l12_2a,c=3*t._l01_a*(t._l01_a+t._l12_a);r=(r*u-t._x0*t._l12_2a+t._x2*t._l01_2a)/c,i=(i*u-t._y0*t._l12_2a+t._y2*t._l01_2a)/c}if(t._l23_a>ft){var s=2*t._l23_2a+3*t._l23_a*t._l12_a+t._l12_2a,f=3*t._l23_a*(t._l23_a+t._l12_a);o=(o*s+t._x1*t._l23_2a-n*t._l12_2a)/f,a=(a*s+t._y1*t._l23_2a-e*t._l12_2a)/f}t._context.bezierCurveTo(r,i,o,a,t._x2,t._y2)}function j(t,n){this._context=t,this._alpha=n}function F(t,n){this._context=t,this._alpha=n}function I(t,n){this._context=t,this._alpha=n}function z(t){this._context=t}function Y(t){return t<0?-1:1}function L(t,n,e){var r=t._x1-t._x0,i=n-t._x1,o=(t._y1-t._y0)/(r||i<0&&-0),a=(e-t._y1)/(i||r<0&&-0),u=(o*i+a*r)/(r+i);return(Y(o)+Y(a))*Math.min(Math.abs(o),Math.abs(a),.5*Math.abs(u))||0}function B(t,n){var e=t._x1-t._x0;return e?(3*(t._y1-t._y0)/e-n)/2:n}function R(t,n,e){var r=t._x0,i=t._y0,o=t._x1,a=t._y1,u=(o-r)/3;t._context.bezierCurveTo(r+u,i+u*n,o-u,a-u*e,o,a)}function H(t){this._context=t}function X(t){this._context=new $(t)}function $(t){this._context=t}function W(t){return new H(t)}function V(t){return new X(t)}function Z(t){this._context=t}function Q(t){var n,e,r=t.length-1,i=new Array(r),o=new Array(r),a=new Array(r);for(i[0]=0,o[0]=2,a[0]=t[0]+2*t[1],n=1;n<r-1;++n)i[n]=1,o[n]=4,a[n]=4*t[n]+2*t[n+1];for(i[r-1]=2,o[r-1]=7,a[r-1]=8*t[r-1]+t[r],n=1;n<r;++n)e=i[n]/o[n-1],o[n]-=e,a[n]-=e*a[n-1];for(i[r-1]=a[r-1]/o[r-1],n=r-2;n>=0;--n)i[n]=(a[n]-i[n+1])/o[n];for(o[r-1]=(t[r]+i[r-1])/2,n=0;n<r-1;++n)o[n]=2*t[n+1]-i[n+1];return[i,o]}function J(t,n){this._context=t,this._t=n}function G(t){return new J(t,0)}function K(t){return new J(t,1)}function tt(t,n){return t[n]}function nt(t){for(var n,e=0,r=-1,i=t.length;++r<i;)(n=+t[r][1])&&(e+=n);return e}var et=function(t){return function(){return t}},rt=Math.abs,it=Math.atan2,ot=Math.cos,at=Math.max,ut=Math.min,ct=Math.sin,st=Math.sqrt,ft=1e-12,lt=Math.PI,ht=lt/2,dt=2*lt,pt=function(){function t(){var t,i,o=+l.apply(this,arguments),a=+h.apply(this,arguments),u=_.apply(this,arguments)-ht,c=y.apply(this,arguments)-ht,m=rt(c-u),b=c>u;if(g||(g=t=n.path()),a<o&&(i=a,a=o,o=i),a>ft)if(m>dt-ft)g.moveTo(a*ot(u),a*ct(u)),g.arc(0,0,a,u,c,!b),o>ft&&(g.moveTo(o*ot(c),o*ct(c)),g.arc(0,0,o,c,u,b));else{var x,w,M=u,T=c,k=u,A=c,N=m,S=m,C=v.apply(this,arguments)/2,E=C>ft&&(p?+p.apply(this,arguments):st(o*o+a*a)),O=ut(rt(a-o)/2,+d.apply(this,arguments)),P=O,D=O;if(E>ft){var q=r(E/o*ct(C)),U=r(E/a*ct(C));(N-=2*q)>ft?(q*=b?1:-1,k+=q,A-=q):(N=0,k=A=(u+c)/2),(S-=2*U)>ft?(U*=b?1:-1,M+=U,T-=U):(S=0,M=T=(u+c)/2)}var j=a*ot(M),F=a*ct(M),I=o*ot(A),z=o*ct(A);if(O>ft){var Y=a*ot(T),L=a*ct(T),B=o*ot(k),R=o*ct(k);if(m<lt){var H=N>ft?s(j,F,B,R,Y,L,I,z):[I,z],X=j-H[0],$=F-H[1],W=Y-H[0],V=L-H[1],Z=1/ct(e((X*W+$*V)/(st(X*X+$*$)*st(W*W+V*V)))/2),Q=st(H[0]*H[0]+H[1]*H[1]);P=ut(O,(o-Q)/(Z-1)),D=ut(O,(a-Q)/(Z+1))}}S>ft?D>ft?(x=f(B,R,j,F,a,D,b),w=f(Y,L,I,z,a,D,b),g.moveTo(x.cx+x.x01,x.cy+x.y01),D<O?g.arc(x.cx,x.cy,D,it(x.y01,x.x01),it(w.y01,w.x01),!b):(g.arc(x.cx,x.cy,D,it(x.y01,x.x01),it(x.y11,x.x11),!b),g.arc(0,0,a,it(x.cy+x.y11,x.cx+x.x11),it(w.cy+w.y11,w.cx+w.x11),!b),g.arc(w.cx,w.cy,D,it(w.y11,w.x11),it(w.y01,w.x01),!b))):(g.moveTo(j,F),g.arc(0,0,a,M,T,!b)):g.moveTo(j,F),o>ft&&N>ft?P>ft?(x=f(I,z,Y,L,o,-P,b),w=f(j,F,B,R,o,-P,b),g.lineTo(x.cx+x.x01,x.cy+x.y01),P<O?g.arc(x.cx,x.cy,P,it(x.y01,x.x01),it(w.y01,w.x01),!b):(g.arc(x.cx,x.cy,P,it(x.y01,x.x01),it(x.y11,x.x11),!b),g.arc(0,0,o,it(x.cy+x.y11,x.cx+x.x11),it(w.cy+w.y11,w.cx+w.x11),b),g.arc(w.cx,w.cy,P,it(w.y11,w.x11),it(w.y01,w.x01),!b))):g.arc(0,0,o,A,k,b):g.lineTo(I,z)}else g.moveTo(0,0);if(g.closePath(),t)return g=null,t+""||null}var l=i,h=o,d=et(0),p=null,_=a,y=u,v=c,g=null;return t.centroid=function(){var t=(+l.apply(this,arguments)+ +h.apply(this,arguments))/2,n=(+_.apply(this,arguments)+ +y.apply(this,arguments))/2-lt/2;return[ot(n)*t,ct(n)*t]},t.innerRadius=function(n){return arguments.length?(l="function"==typeof n?n:et(+n),t):l},t.outerRadius=function(n){return arguments.length?(h="function"==typeof n?n:et(+n),t):h},t.cornerRadius=function(n){return arguments.length?(d="function"==typeof n?n:et(+n),t):d},t.padRadius=function(n){return arguments.length?(p=null==n?null:"function"==typeof n?n:et(+n),t):p},t.startAngle=function(n){return arguments.length?(_="function"==typeof n?n:et(+n),t):_},t.endAngle=function(n){return arguments.length?(y="function"==typeof n?n:et(+n),t):y},t.padAngle=function(n){return arguments.length?(v="function"==typeof n?n:et(+n),t):v},t.context=function(n){return arguments.length?(g=null==n?null:n,t):g},t};l.prototype={areaStart:function(){this._line=0},areaEnd:function(){this._line=NaN},lineStart:function(){this._point=0},lineEnd:function(){(this._line||0!==this._line&&1===this._point)&&this._context.closePath(),this._line=1-this._line},point:function(t,n){switch(t=+t,n=+n,this._point){case 0:this._point=1,this._line?this._context.lineTo(t,n):this._context.moveTo(t,n);break;case 1:this._point=2;default:this._context.lineTo(t,n)}}};var _t=function(t){return new l(t)},yt=function(){function t(t){var c,s,f,l=t.length,h=!1;for(null==o&&(u=a(f=n.path())),c=0;c<=l;++c)!(c<l&&i(s=t[c],c,t))===h&&((h=!h)?u.lineStart():u.lineEnd()),h&&u.point(+e(s,c,t),+r(s,c,t));if(f)return u=null,f+""||null}var e=h,r=d,i=et(!0),o=null,a=_t,u=null;return t.x=function(n){return arguments.length?(e="function"==typeof n?n:et(+n),t):e},t.y=function(n){return arguments.length?(r="function"==typeof n?n:et(+n),t):r},t.defined=function(n){return arguments.length?(i="function"==typeof n?n:et(!!n),t):i},t.curve=function(n){return arguments.length?(a=n,null!=o&&(u=a(o)),t):a},t.context=function(n){return arguments.length?(null==n?o=u=null:u=a(o=n),t):o},t},vt=function(){function t(t){var e,l,h,d,p,_=t.length,y=!1,v=new Array(_),g=new Array(_);for(null==c&&(f=s(p=n.path())),e=0;e<=_;++e){if(!(e<_&&u(d=t[e],e,t))===y)if(y=!y)l=e,f.areaStart(),f.lineStart();else{for(f.lineEnd(),f.lineStart(),h=e-1;h>=l;--h)f.point(v[h],g[h]);f.lineEnd(),f.areaEnd()}y&&(v[e]=+r(d,e,t),g[e]=+o(d,e,t),f.point(i?+i(d,e,t):v[e],a?+a(d,e,t):g[e]))}if(p)return f=null,p+""||null}function e(){return yt().defined(u).curve(s).context(c)}var r=h,i=null,o=et(0),a=d,u=et(!0),c=null,s=_t,f=null;return t.x=function(n){return arguments.length?(r="function"==typeof n?n:et(+n),i=null,t):r},t.x0=function(n){return arguments.length?(r="function"==typeof n?n:et(+n),t):r},t.x1=function(n){return arguments.length?(i=null==n?null:"function"==typeof n?n:et(+n),t):i},t.y=function(n){return arguments.length?(o="function"==typeof n?n:et(+n),a=null,t):o},t.y0=function(n){return arguments.length?(o="function"==typeof n?n:et(+n),t):o},t.y1=function(n){return arguments.length?(a=null==n?null:"function"==typeof n?n:et(+n),t):a},t.lineX0=t.lineY0=function(){return e().x(r).y(o)},t.lineY1=function(){return e().x(r).y(a)},t.lineX1=function(){return e().x(i).y(o)},t.defined=function(n){return arguments.length?(u="function"==typeof n?n:et(!!n),t):u},t.curve=function(n){return arguments.length?(s=n,null!=c&&(f=s(c)),t):s},t.context=function(n){return arguments.length?(null==n?c=f=null:f=s(c=n),t):c},t},gt=function(t,n){return n<t?-1:n>t?1:n>=t?0:NaN},mt=function(t){return t},bt=function(){function t(t){var u,c,s,f,l,h=t.length,d=0,p=new Array(h),_=new Array(h),y=+i.apply(this,arguments),v=Math.min(dt,Math.max(-dt,o.apply(this,arguments)-y)),g=Math.min(Math.abs(v)/h,a.apply(this,arguments)),m=g*(v<0?-1:1);for(u=0;u<h;++u)(l=_[p[u]=u]=+n(t[u],u,t))>0&&(d+=l);for(null!=e?p.sort(function(t,n){return e(_[t],_[n])}):null!=r&&p.sort(function(n,e){return r(t[n],t[e])}),u=0,s=d?(v-h*m)/d:0;u<h;++u,y=f)c=p[u],f=y+((l=_[c])>0?l*s:0)+m,_[c]={data:t[c],index:u,value:l,startAngle:y,endAngle:f,padAngle:g};return _}var n=mt,e=gt,r=null,i=et(0),o=et(dt),a=et(0);return t.value=function(e){return arguments.length?(n="function"==typeof e?e:et(+e),t):n},t.sortValues=function(n){return arguments.length?(e=n,r=null,t):e},t.sort=function(n){return arguments.length?(r=n,e=null,t):r},t.startAngle=function(n){return arguments.length?(i="function"==typeof n?n:et(+n),t):i},t.endAngle=function(n){return arguments.length?(o="function"==typeof n?n:et(+n),t):o},t.padAngle=function(n){return arguments.length?(a="function"==typeof n?n:et(+n),t):a},t},xt=_(_t);p.prototype={areaStart:function(){this._curve.areaStart()},areaEnd:function(){this._curve.areaEnd()},lineStart:function(){this._curve.lineStart()},lineEnd:function(){this._curve.lineEnd()},point:function(t,n){this._curve.point(n*Math.sin(t),n*-Math.cos(t))}};var wt=function(){return y(yt().curve(xt))},Mt=function(){var t=vt().curve(xt),n=t.curve,e=t.lineX0,r=t.lineX1,i=t.lineY0,o=t.lineY1;return t.angle=t.x,delete t.x,t.startAngle=t.x0,delete t.x0,t.endAngle=t.x1,delete t.x1,t.radius=t.y,delete t.y,t.innerRadius=t.y0,delete t.y0,t.outerRadius=t.y1,delete t.y1,t.lineStartAngle=function(){return y(e())},delete t.lineX0,t.lineEndAngle=function(){return y(r())},delete t.lineX1,t.lineInnerRadius=function(){return y(i())},delete t.lineY0,t.lineOuterRadius=function(){return y(o())},delete t.lineY1,t.curve=function(t){return arguments.length?n(_(t)):n()._curve},t},Tt=function(t,n){return[(n=+n)*Math.cos(t-=Math.PI/2),n*Math.sin(t)]},kt=Array.prototype.slice,At={draw:function(t,n){var e=Math.sqrt(n/lt);t.moveTo(e,0),t.arc(0,0,e,0,dt)}},Nt={draw:function(t,n){var e=Math.sqrt(n/5)/2;t.moveTo(-3*e,-e),t.lineTo(-e,-e),t.lineTo(-e,-3*e),t.lineTo(e,-3*e),t.lineTo(e,-e),t.lineTo(3*e,-e),t.lineTo(3*e,e),t.lineTo(e,e),t.lineTo(e,3*e),t.lineTo(-e,3*e),t.lineTo(-e,e),t.lineTo(-3*e,e),t.closePath()}},St=Math.sqrt(1/3),Ct=2*St,Et={draw:function(t,n){var e=Math.sqrt(n/Ct),r=e*St;t.moveTo(0,-e),t.lineTo(r,0),t.lineTo(0,e),t.lineTo(-r,0),t.closePath()}},Ot=Math.sin(lt/10)/Math.sin(7*lt/10),Pt=Math.sin(dt/10)*Ot,Dt=-Math.cos(dt/10)*Ot,qt={draw:function(t,n){var e=Math.sqrt(.8908130915292852*n),r=Pt*e,i=Dt*e;t.moveTo(0,-e),t.lineTo(r,i);for(var o=1;o<5;++o){var a=dt*o/5,u=Math.cos(a),c=Math.sin(a);t.lineTo(c*e,-u*e),t.lineTo(u*r-c*i,c*r+u*i)}t.closePath()}},Ut={draw:function(t,n){var e=Math.sqrt(n),r=-e/2;t.rect(r,r,e,e)}},jt=Math.sqrt(3),Ft={draw:function(t,n){var e=-Math.sqrt(n/(3*jt));t.moveTo(0,2*e),t.lineTo(-jt*e,-e),t.lineTo(jt*e,-e),t.closePath()}},It=-.5,zt=Math.sqrt(3)/2,Yt=1/Math.sqrt(12),Lt=3*(Yt/2+1),Bt={draw:function(t,n){var e=Math.sqrt(n/Lt),r=e/2,i=e*Yt,o=r,a=e*Yt+e,u=-o,c=a;t.moveTo(r,i),t.lineTo(o,a),t.lineTo(u,c),t.lineTo(It*r-zt*i,zt*r+It*i),t.lineTo(It*o-zt*a,zt*o+It*a),t.lineTo(It*u-zt*c,zt*u+It*c),t.lineTo(It*r+zt*i,It*i-zt*r),t.lineTo(It*o+zt*a,It*a-zt*o),t.lineTo(It*u+zt*c,It*c-zt*u),t.closePath()}},Rt=[At,Nt,Et,Ut,qt,Ft,Bt],Ht=function(){function t(){var t;if(i||(i=t=n.path()),e.apply(this,arguments).draw(i,+r.apply(this,arguments)),t)return i=null,t+""||null}var e=et(At),r=et(64),i=null;return t.type=function(n){return arguments.length?(e="function"==typeof n?n:et(n),t):e},t.size=function(n){return arguments.length?(r="function"==typeof n?n:et(+n),t):r},t.context=function(n){return arguments.length?(i=null==n?null:n,t):i},t},Xt=function(){};N.prototype={areaStart:function(){this._line=0},areaEnd:function(){this._line=NaN},lineStart:function(){this._x0=this._x1=this._y0=this._y1=NaN,this._point=0},lineEnd:function(){switch(this._point){case 3:A(this,this._x1,this._y1);case 2:this._context.lineTo(this._x1,this._y1)}(this._line||0!==this._line&&1===this._point)&&this._context.closePath(),this._line=1-this._line},point:function(t,n){switch(t=+t,n=+n,this._point){case 0:this._point=1,this._line?this._context.lineTo(t,n):this._context.moveTo(t,n);break;case 1:this._point=2;break;case 2:this._point=3,this._context.lineTo((5*this._x0+this._x1)/6,(5*this._y0+this._y1)/6);default:A(this,t,n)}this._x0=this._x1,this._x1=t,this._y0=this._y1,this._y1=n}};var $t=function(t){return new N(t)};S.prototype={areaStart:Xt,areaEnd:Xt,lineStart:function(){this._x0=this._x1=this._x2=this._x3=this._x4=this._y0=this._y1=this._y2=this._y3=this._y4=NaN,this._point=0},lineEnd:function(){switch(this._point){case 1:this._context.moveTo(this._x2,this._y2),this._context.closePath();break;case 2:this._context.moveTo((this._x2+2*this._x3)/3,(this._y2+2*this._y3)/3),this._context.lineTo((this._x3+2*this._x2)/3,(this._y3+2*this._y2)/3),this._context.closePath();break;case 3:this.point(this._x2,this._y2),this.point(this._x3,this._y3),this.point(this._x4,this._y4)}},point:function(t,n){switch(t=+t,n=+n,this._point){case 0:this._point=1,this._x2=t,this._y2=n;break;case 1:this._point=2,this._x3=t,this._y3=n;break;case 2:this._point=3,this._x4=t,this._y4=n,this._context.moveTo((this._x0+4*this._x1+t)/6,(this._y0+4*this._y1+n)/6);break;default:A(this,t,n)}this._x0=this._x1,this._x1=t,this._y0=this._y1,this._y1=n}};var Wt=function(t){return new S(t)};C.prototype={areaStart:function(){this._line=0},areaEnd:function(){this._line=NaN},lineStart:function(){this._x0=this._x1=this._y0=this._y1=NaN,this._point=0},lineEnd:function(){(this._line||0!==this._line&&3===this._point)&&this._context.closePath(),this._line=1-this._line},point:function(t,n){switch(t=+t,n=+n,this._point){case 0:this._point=1;break;case 1:this._point=2;break;case 2:this._point=3;var e=(this._x0+4*this._x1+t)/6,r=(this._y0+4*this._y1+n)/6;this._line?this._context.lineTo(e,r):this._context.moveTo(e,r);break;case 3:this._point=4;default:A(this,t,n)}this._x0=this._x1,this._x1=t,this._y0=this._y1,this._y1=n}};var Vt=function(t){return new C(t)};E.prototype={lineStart:function(){this._x=[],this._y=[],this._basis.lineStart()},lineEnd:function(){var t=this._x,n=this._y,e=t.length-1;if(e>0)for(var r,i=t[0],o=n[0],a=t[e]-i,u=n[e]-o,c=-1;++c<=e;)r=c/e,this._basis.point(this._beta*t[c]+(1-this._beta)*(i+r*a),this._beta*n[c]+(1-this._beta)*(o+r*u));this._x=this._y=null,this._basis.lineEnd()},point:function(t,n){this._x.push(+t),this._y.push(+n)}};var Zt=function t(n){function e(t){return 1===n?new N(t):new E(t,n)}return e.beta=function(n){return t(+n)},e}(.85);P.prototype={areaStart:function(){this._line=0},areaEnd:function(){this._line=NaN},lineStart:function(){this._x0=this._x1=this._x2=this._y0=this._y1=this._y2=NaN,this._point=0},lineEnd:function(){switch(this._point){case 2:this._context.lineTo(this._x2,this._y2);break;case 3:O(this,this._x1,this._y1)}(this._line||0!==this._line&&1===this._point)&&this._context.closePath(),this._line=1-this._line},point:function(t,n){switch(t=+t,n=+n,this._point){case 0:this._point=1,this._line?this._context.lineTo(t,n):this._context.moveTo(t,n);break;case 1:this._point=2,this._x1=t,this._y1=n;break;case 2:this._point=3;default:O(this,t,n)}this._x0=this._x1,this._x1=this._x2,this._x2=t,this._y0=this._y1,this._y1=this._y2,this._y2=n}};var Qt=function t(n){function e(t){return new P(t,n)}return e.tension=function(n){return t(+n)},e}(0);D.prototype={areaStart:Xt,areaEnd:Xt,lineStart:function(){this._x0=this._x1=this._x2=this._x3=this._x4=this._x5=this._y0=this._y1=this._y2=this._y3=this._y4=this._y5=NaN,this._point=0},lineEnd:function(){switch(this._point){case 1:this._context.moveTo(this._x3,this._y3),this._context.closePath();break;case 2:this._context.lineTo(this._x3,this._y3),this._context.closePath();break;case 3:this.point(this._x3,this._y3),this.point(this._x4,this._y4),this.point(this._x5,this._y5)}},point:function(t,n){switch(t=+t,n=+n,this._point){case 0:this._point=1,this._x3=t,this._y3=n;break;case 1:this._point=2,this._context.moveTo(this._x4=t,this._y4=n);break;case 2:this._point=3,this._x5=t,this._y5=n;break;default:O(this,t,n)}this._x0=this._x1,this._x1=this._x2,this._x2=t,this._y0=this._y1,this._y1=this._y2,this._y2=n}};var Jt=function t(n){function e(t){return new D(t,n)}return e.tension=function(n){return t(+n)},e}(0);q.prototype={areaStart:function(){this._line=0},areaEnd:function(){this._line=NaN},lineStart:function(){this._x0=this._x1=this._x2=this._y0=this._y1=this._y2=NaN,this._point=0},lineEnd:function(){(this._line||0!==this._line&&3===this._point)&&this._context.closePath(),this._line=1-this._line},point:function(t,n){switch(t=+t,n=+n,this._point){case 0:this._point=1;break;case 1:this._point=2;break;case 2:this._point=3,this._line?this._context.lineTo(this._x2,this._y2):this._context.moveTo(this._x2,this._y2);break;case 3:this._point=4;default:O(this,t,n)}this._x0=this._x1,this._x1=this._x2,this._x2=t,this._y0=this._y1,this._y1=this._y2,this._y2=n}};var Gt=function t(n){function e(t){return new q(t,n)}return e.tension=function(n){return t(+n)},e}(0);j.prototype={areaStart:function(){this._line=0},areaEnd:function(){this._line=NaN},lineStart:function(){this._x0=this._x1=this._x2=this._y0=this._y1=this._y2=NaN,this._l01_a=this._l12_a=this._l23_a=this._l01_2a=this._l12_2a=this._l23_2a=this._point=0},lineEnd:function(){switch(this._point){case 2:this._context.lineTo(this._x2,this._y2);break;case 3:this.point(this._x2,this._y2)}(this._line||0!==this._line&&1===this._point)&&this._context.closePath(),this._line=1-this._line},point:function(t,n){if(t=+t,n=+n,this._point){var e=this._x2-t,r=this._y2-n;this._l23_a=Math.sqrt(this._l23_2a=Math.pow(e*e+r*r,this._alpha))}switch(this._point){case 0:this._point=1,this._line?this._context.lineTo(t,n):this._context.moveTo(t,n);break;case 1:this._point=2;break;case 2:this._point=3;default:U(this,t,n)}this._l01_a=this._l12_a,this._l12_a=this._l23_a,this._l01_2a=this._l12_2a,this._l12_2a=this._l23_2a,this._x0=this._x1,this._x1=this._x2,this._x2=t,this._y0=this._y1,this._y1=this._y2,this._y2=n}};var Kt=function t(n){function e(t){return n?new j(t,n):new P(t,0)}return e.alpha=function(n){return t(+n)},e}(.5);F.prototype={areaStart:Xt,areaEnd:Xt,lineStart:function(){this._x0=this._x1=this._x2=this._x3=this._x4=this._x5=this._y0=this._y1=this._y2=this._y3=this._y4=this._y5=NaN,this._l01_a=this._l12_a=this._l23_a=this._l01_2a=this._l12_2a=this._l23_2a=this._point=0},lineEnd:function(){switch(this._point){case 1:this._context.moveTo(this._x3,this._y3),this._context.closePath();break;case 2:this._context.lineTo(this._x3,this._y3),this._context.closePath();break;case 3:this.point(this._x3,this._y3),this.point(this._x4,this._y4),this.point(this._x5,this._y5)}},point:function(t,n){if(t=+t,n=+n,this._point){var e=this._x2-t,r=this._y2-n;this._l23_a=Math.sqrt(this._l23_2a=Math.pow(e*e+r*r,this._alpha))}switch(this._point){case 0:this._point=1,this._x3=t,this._y3=n;break;case 1:this._point=2,this._context.moveTo(this._x4=t,this._y4=n);break;case 2:this._point=3,this._x5=t,this._y5=n;break;default:U(this,t,n)}this._l01_a=this._l12_a,this._l12_a=this._l23_a,this._l01_2a=this._l12_2a,this._l12_2a=this._l23_2a,this._x0=this._x1,this._x1=this._x2,this._x2=t,this._y0=this._y1,this._y1=this._y2,this._y2=n}};var tn=function t(n){function e(t){return n?new F(t,n):new D(t,0)}return e.alpha=function(n){return t(+n)},e}(.5);I.prototype={areaStart:function(){this._line=0},areaEnd:function(){this._line=NaN},lineStart:function(){this._x0=this._x1=this._x2=this._y0=this._y1=this._y2=NaN,this._l01_a=this._l12_a=this._l23_a=this._l01_2a=this._l12_2a=this._l23_2a=this._point=0},lineEnd:function(){(this._line||0!==this._line&&3===this._point)&&this._context.closePath(),this._line=1-this._line},point:function(t,n){if(t=+t,n=+n,this._point){var e=this._x2-t,r=this._y2-n;this._l23_a=Math.sqrt(this._l23_2a=Math.pow(e*e+r*r,this._alpha))}switch(this._point){case 0:this._point=1;break;case 1:this._point=2;break;case 2:this._point=3,this._line?this._context.lineTo(this._x2,this._y2):this._context.moveTo(this._x2,this._y2);break;case 3:this._point=4;default:U(this,t,n)}this._l01_a=this._l12_a,this._l12_a=this._l23_a,this._l01_2a=this._l12_2a,this._l12_2a=this._l23_2a,this._x0=this._x1,this._x1=this._x2,this._x2=t,this._y0=this._y1,this._y1=this._y2,this._y2=n}};var nn=function t(n){function e(t){return n?new I(t,n):new q(t,0)}return e.alpha=function(n){return t(+n)},e}(.5);z.prototype={areaStart:Xt,areaEnd:Xt,lineStart:function(){this._point=0},lineEnd:function(){this._point&&this._context.closePath()},point:function(t,n){t=+t,n=+n,this._point?this._context.lineTo(t,n):(this._point=1,this._context.moveTo(t,n))}};var en=function(t){return new z(t)};H.prototype={areaStart:function(){this._line=0},areaEnd:function(){this._line=NaN},lineStart:function(){this._x0=this._x1=this._y0=this._y1=this._t0=NaN,this._point=0},lineEnd:function(){switch(this._point){case 2:this._context.lineTo(this._x1,this._y1);break;case 3:R(this,this._t0,B(this,this._t0))}(this._line||0!==this._line&&1===this._point)&&this._context.closePath(),this._line=1-this._line},point:function(t,n){var e=NaN;if(t=+t,n=+n,t!==this._x1||n!==this._y1){switch(this._point){case 0:this._point=1,this._line?this._context.lineTo(t,n):this._context.moveTo(t,n);break;case 1:this._point=2;break;case 2:this._point=3,R(this,B(this,e=L(this,t,n)),e);break;default:R(this,this._t0,e=L(this,t,n))}this._x0=this._x1,this._x1=t,this._y0=this._y1,this._y1=n,this._t0=e}}},(X.prototype=Object.create(H.prototype)).point=function(t,n){H.prototype.point.call(this,n,t)},$.prototype={moveTo:function(t,n){this._context.moveTo(n,t)},closePath:function(){this._context.closePath()},lineTo:function(t,n){this._context.lineTo(n,t)},bezierCurveTo:function(t,n,e,r,i,o){this._context.bezierCurveTo(n,t,r,e,o,i)}},Z.prototype={areaStart:function(){this._line=0},areaEnd:function(){this._line=NaN},lineStart:function(){this._x=[],this._y=[]},lineEnd:function(){var t=this._x,n=this._y,e=t.length;if(e)if(this._line?this._context.lineTo(t[0],n[0]):this._context.moveTo(t[0],n[0]),2===e)this._context.lineTo(t[1],n[1]);else for(var r=Q(t),i=Q(n),o=0,a=1;a<e;++o,++a)this._context.bezierCurveTo(r[0][o],i[0][o],r[1][o],i[1][o],t[a],n[a]);(this._line||0!==this._line&&1===e)&&this._context.closePath(),this._line=1-this._line,this._x=this._y=null},point:function(t,n){this._x.push(+t),this._y.push(+n)}};var rn=function(t){return new Z(t)};J.prototype={areaStart:function(){this._line=0},areaEnd:function(){this._line=NaN},lineStart:function(){this._x=this._y=NaN,this._point=0},lineEnd:function(){0<this._t&&this._t<1&&2===this._point&&this._context.lineTo(this._x,this._y),(this._line||0!==this._line&&1===this._point)&&this._context.closePath(),this._line>=0&&(this._t=1-this._t,this._line=1-this._line)},point:function(t,n){switch(t=+t,n=+n,this._point){case 0:this._point=1,this._line?this._context.lineTo(t,n):this._context.moveTo(t,n);break;case 1:this._point=2;default:if(this._t<=0)this._context.lineTo(this._x,n),this._context.lineTo(t,n);else{var e=this._x*(1-this._t)+t*this._t;this._context.lineTo(e,this._y),this._context.lineTo(e,n)}}this._x=t,this._y=n}};var on=function(t){return new J(t,.5)},an=function(t,n){if((i=t.length)>1)for(var e,r,i,o=1,a=t[n[0]],u=a.length;o<i;++o)for(r=a,a=t[n[o]],e=0;e<u;++e)a[e][1]+=a[e][0]=isNaN(r[e][1])?r[e][0]:r[e][1]},un=function(t){for(var n=t.length,e=new Array(n);--n>=0;)e[n]=n;return e},cn=function(){function t(t){var o,a,u=n.apply(this,arguments),c=t.length,s=u.length,f=new Array(s);for(o=0;o<s;++o){for(var l,h=u[o],d=f[o]=new Array(c),p=0;p<c;++p)d[p]=l=[0,+i(t[p],h,p,t)],l.data=t[p];d.key=h}for(o=0,a=e(f);o<s;++o)f[a[o]].index=o;return r(f,a),f}var n=et([]),e=un,r=an,i=tt;return t.keys=function(e){return arguments.length?(n="function"==typeof e?e:et(kt.call(e)),t):n},t.value=function(n){return arguments.length?(i="function"==typeof n?n:et(+n),t):i},t.order=function(n){return arguments.length?(e=null==n?un:"function"==typeof n?n:et(kt.call(n)),t):e},t.offset=function(n){return arguments.length?(r=null==n?an:n,t):r},t},sn=function(t,n){if((r=t.length)>0){for(var e,r,i,o=0,a=t[0].length;o<a;++o){for(i=e=0;e<r;++e)i+=t[e][o][1]||0;if(i)for(e=0;e<r;++e)t[e][o][1]/=i}an(t,n)}},fn=function(t,n){if((u=t.length)>1)for(var e,r,i,o,a,u,c=0,s=t[n[0]].length;c<s;++c)for(o=a=0,e=0;e<u;++e)(i=(r=t[n[e]][c])[1]-r[0])>=0?(r[0]=o,r[1]=o+=i):i<0?(r[1]=a,r[0]=a+=i):r[0]=o},ln=function(t,n){if((e=t.length)>0){for(var e,r=0,i=t[n[0]],o=i.length;r<o;++r){for(var a=0,u=0;a<e;++a)u+=t[a][r][1]||0;i[r][1]+=i[r][0]=-u/2}an(t,n)}},hn=function(t,n){if((i=t.length)>0&&(r=(e=t[n[0]]).length)>0){for(var e,r,i,o=0,a=1;a<r;++a){for(var u=0,c=0,s=0;u<i;++u){for(var f=t[n[u]],l=f[a][1]||0,h=(l-(f[a-1][1]||0))/2,d=0;d<u;++d){var p=t[n[d]];h+=(p[a][1]||0)-(p[a-1][1]||0)}c+=l,s+=h*l}e[a-1][1]+=e[a-1][0]=o,c&&(o-=s/c)}e[a-1][1]+=e[a-1][0]=o,an(t,n)}},dn=function(t){var n=t.map(nt);return un(t).sort(function(t,e){return n[t]-n[e]})},pn=function(t){return dn(t).reverse()},_n=function(t){var n,e,r=t.length,i=t.map(nt),o=un(t).sort(function(t,n){return i[n]-i[t]}),a=0,u=0,c=[],s=[];for(n=0;n<r;++n)e=o[n],a<u?(a+=i[e],c.push(e)):(u+=i[e],s.push(e));return s.reverse().concat(c)},yn=function(t){return un(t).reverse()};t.arc=pt,t.area=vt,t.line=yt,t.pie=bt,t.areaRadial=Mt,t.radialArea=Mt,t.lineRadial=wt,t.radialLine=wt,t.pointRadial=Tt,t.linkHorizontal=M,t.linkVertical=T,t.linkRadial=k,t.symbol=Ht,t.symbols=Rt,t.symbolCircle=At,t.symbolCross=Nt,t.symbolDiamond=Et,t.symbolSquare=Ut,t.symbolStar=qt,t.symbolTriangle=Ft,t.symbolWye=Bt,t.curveBasisClosed=Wt,t.curveBasisOpen=Vt,t.curveBasis=$t,t.curveBundle=Zt,t.curveCardinalClosed=Jt,t.curveCardinalOpen=Gt,t.curveCardinal=Qt,t.curveCatmullRomClosed=tn,t.curveCatmullRomOpen=nn,t.curveCatmullRom=Kt,t.curveLinearClosed=en,t.curveLinear=_t,t.curveMonotoneX=W,t.curveMonotoneY=V,t.curveNatural=rn,t.curveStep=on,t.curveStepAfter=K,t.curveStepBefore=G,t.stack=cn,t.stackOffsetExpand=sn,t.stackOffsetDiverging=fn,t.stackOffsetNone=an,t.stackOffsetSilhouette=ln,t.stackOffsetWiggle=hn,t.stackOrderAscending=dn,t.stackOrderDescending=pn,t.stackOrderInsideOut=_n,t.stackOrderNone=un,t.stackOrderReverse=yn,Object.defineProperty(t,"__esModule",{value:!0})})},{"d3-path":10}],14:[function(t,n,e){!function(r,i){"object"==typeof e&&void 0!==n?i(e,t("d3-time")):"function"==typeof define&&define.amd?define(["exports","d3-time"],i):i(r.d3=r.d3||{},r.d3)}(this,function(t,n){"use strict";function e(t){if(0<=t.y&&t.y<100){var n=new Date(-1,t.m,t.d,t.H,t.M,t.S,t.L);return n.setFullYear(t.y),n}return new Date(t.y,t.m,t.d,t.H,t.M,t.S,t.L)}function r(t){if(0<=t.y&&t.y<100){var n=new Date(Date.UTC(-1,t.m,t.d,t.H,t.M,t.S,t.L));return n.setUTCFullYear(t.y),n}return new Date(Date.UTC(t.y,t.m,t.d,t.H,t.M,t.S,t.L))}function i(t){return{y:t,m:0,d:1,H:0,M:0,S:0,L:0}}function o(t){function n(t,n){return function(e){var r,i,o,a=[],u=-1,c=0,s=t.length;for(e instanceof Date||(e=new Date(+e));++u<s;)37===t.charCodeAt(u)&&(a.push(t.slice(c,u)),null!=(i=rt[r=t.charAt(++u)])?r=t.charAt(++u):i="e"===r?" ":"0",(o=n[r])&&(r=o(e,i)),a.push(r),c=u+1);return a.push(t.slice(c,u)),a.join("")}}function o(t,n){return function(e){var o=i(1900);if(a(o,t,e+="",0)!=e.length)return null;if("p"in o&&(o.H=o.H%12+12*o.p),"W"in o||"U"in o){"w"in o||(o.w="W"in o?1:0);var u="Z"in o?r(i(o.y)).getUTCDay():n(i(o.y)).getDay();o.m=0,o.d="W"in o?(o.w+6)%7+7*o.W-(u+5)%7:o.w+7*o.U-(u+6)%7}return"Z"in o?(o.H+=o.Z/100|0,o.M+=o.Z%100,r(o)):n(o)}}function a(t,n,e,r){for(var i,o,a=0,u=n.length,c=e.length;a<u;){if(r>=c)return-1;if(37===(i=n.charCodeAt(a++))){if(i=n.charAt(a++),!(o=Ft[i in rt?n.charAt(a++):i])||(r=o(t,e,r))<0)return-1}else if(i!=e.charCodeAt(r++))return-1}return r}function u(t,n,e){var r=kt.exec(n.slice(e));return r?(t.p=At[r[0].toLowerCase()],e+r[0].length):-1}function K(t,n,e){var r=Ct.exec(n.slice(e));return r?(t.w=Et[r[0].toLowerCase()],e+r[0].length):-1}function tt(t,n,e){var r=Nt.exec(n.slice(e));return r?(t.w=St[r[0].toLowerCase()],e+r[0].length):-1}function nt(t,n,e){var r=Dt.exec(n.slice(e));return r?(t.m=qt[r[0].toLowerCase()],e+r[0].length):-1}function et(t,n,e){var r=Ot.exec(n.slice(e));return r?(t.m=Pt[r[0].toLowerCase()],e+r[0].length):-1}function it(t,n,e){return a(t,vt,n,e)}function ot(t,n,e){return a(t,gt,n,e)}function at(t,n,e){return a(t,mt,n,e)}function ut(t){return wt[t.getDay()]}function ct(t){return xt[t.getDay()]}function st(t){return Tt[t.getMonth()]}function ft(t){return Mt[t.getMonth()]}function lt(t){return bt[+(t.getHours()>=12)]}function ht(t){return wt[t.getUTCDay()]}function dt(t){return xt[t.getUTCDay()]}function pt(t){return Tt[t.getUTCMonth()]}function _t(t){return Mt[t.getUTCMonth()]}function yt(t){return bt[+(t.getUTCHours()>=12)]}var vt=t.dateTime,gt=t.date,mt=t.time,bt=t.periods,xt=t.days,wt=t.shortDays,Mt=t.months,Tt=t.shortMonths,kt=c(bt),At=s(bt),Nt=c(xt),St=s(xt),Ct=c(wt),Et=s(wt),Ot=c(Mt),Pt=s(Mt),Dt=c(Tt),qt=s(Tt),Ut={a:ut,A:ct,b:st,B:ft,c:null,d:T,e:T,H:k,I:A,j:N,L:S,m:C,M:E,p:lt,S:O,U:P,w:D,W:q,x:null,X:null,y:U,Y:j,Z:F,"%":G},jt={a:ht,A:dt,b:pt,B:_t,c:null,d:I,e:I,H:z,I:Y,j:L,L:B,m:R,M:H,p:yt,S:X,U:$,w:W,W:V,x:null,X:null,y:Z,Y:Q,Z:J,"%":G},Ft={a:K,A:tt,b:nt,B:et,c:it,d:v,e:v,H:m,I:m,j:g,L:w,m:y,M:b,p:u,S:x,U:l,w:f,W:h,x:ot,X:at,y:p,Y:d,Z:_,"%":M};return Ut.x=n(gt,Ut),Ut.X=n(mt,Ut),Ut.c=n(vt,Ut),jt.x=n(gt,jt),jt.X=n(mt,jt),jt.c=n(vt,jt),{format:function(t){var e=n(t+="",Ut);return e.toString=function(){return t},e},parse:function(t){var n=o(t+="",e);return n.toString=function(){return t},n},utcFormat:function(t){var e=n(t+="",jt);return e.toString=function(){return t},e},utcParse:function(t){var n=o(t,r);return n.toString=function(){return t},n}}}function a(t,n,e){var r=t<0?"-":"",i=(r?-t:t)+"",o=i.length;return r+(o<e?new Array(e-o+1).join(n)+i:i)}function u(t){return t.replace(at,"\\$&")}function c(t){return new RegExp("^(?:"+t.map(u).join("|")+")","i")}function s(t){for(var n={},e=-1,r=t.length;++e<r;)n[t[e].toLowerCase()]=e;return n}function f(t,n,e){var r=it.exec(n.slice(e,e+1));return r?(t.w=+r[0],e+r[0].length):-1}function l(t,n,e){var r=it.exec(n.slice(e));return r?(t.U=+r[0],e+r[0].length):-1}function h(t,n,e){var r=it.exec(n.slice(e));return r?(t.W=+r[0],e+r[0].length):-1}function d(t,n,e){var r=it.exec(n.slice(e,e+4));return r?(t.y=+r[0],e+r[0].length):-1}function p(t,n,e){var r=it.exec(n.slice(e,e+2));return r?(t.y=+r[0]+(+r[0]>68?1900:2e3),e+r[0].length):-1}function _(t,n,e){var r=/^(Z)|([+-]\d\d)(?:\:?(\d\d))?/.exec(n.slice(e,e+6));return r?(t.Z=r[1]?0:-(r[2]+(r[3]||"00")),e+r[0].length):-1}function y(t,n,e){var r=it.exec(n.slice(e,e+2));return r?(t.m=r[0]-1,e+r[0].length):-1}function v(t,n,e){var r=it.exec(n.slice(e,e+2));return r?(t.d=+r[0],e+r[0].length):-1}function g(t,n,e){var r=it.exec(n.slice(e,e+3));return r?(t.m=0,t.d=+r[0],e+r[0].length):-1}function m(t,n,e){var r=it.exec(n.slice(e,e+2));return r?(t.H=+r[0],e+r[0].length):-1}function b(t,n,e){var r=it.exec(n.slice(e,e+2));return r?(t.M=+r[0],e+r[0].length):-1}function x(t,n,e){var r=it.exec(n.slice(e,e+2));return r?(t.S=+r[0],e+r[0].length):-1}function w(t,n,e){var r=it.exec(n.slice(e,e+3));return r?(t.L=+r[0],e+r[0].length):-1}function M(t,n,e){var r=ot.exec(n.slice(e,e+1));return r?e+r[0].length:-1}function T(t,n){return a(t.getDate(),n,2)}function k(t,n){return a(t.getHours(),n,2)}function A(t,n){return a(t.getHours()%12||12,n,2)}function N(t,e){return a(1+n.timeDay.count(n.timeYear(t),t),e,3)}function S(t,n){return a(t.getMilliseconds(),n,3)}function C(t,n){return a(t.getMonth()+1,n,2)}function E(t,n){return a(t.getMinutes(),n,2)}function O(t,n){return a(t.getSeconds(),n,2)}function P(t,e){return a(n.timeSunday.count(n.timeYear(t),t),e,2)}function D(t){return t.getDay()}function q(t,e){return a(n.timeMonday.count(n.timeYear(t),t),e,2)}function U(t,n){return a(t.getFullYear()%100,n,2)}function j(t,n){return a(t.getFullYear()%1e4,n,4)}function F(t){var n=t.getTimezoneOffset();return(n>0?"-":(n*=-1,"+"))+a(n/60|0,"0",2)+a(n%60,"0",2)}function I(t,n){return a(t.getUTCDate(),n,2)}function z(t,n){return a(t.getUTCHours(),n,2)}function Y(t,n){return a(t.getUTCHours()%12||12,n,2)}function L(t,e){return a(1+n.utcDay.count(n.utcYear(t),t),e,3)}function B(t,n){return a(t.getUTCMilliseconds(),n,3)}function R(t,n){return a(t.getUTCMonth()+1,n,2)}function H(t,n){return a(t.getUTCMinutes(),n,2)}function X(t,n){return a(t.getUTCSeconds(),n,2)}function $(t,e){return a(n.utcSunday.count(n.utcYear(t),t),e,2)}function W(t){return t.getUTCDay()}function V(t,e){return a(n.utcMonday.count(n.utcYear(t),t),e,2)}function Z(t,n){return a(t.getUTCFullYear()%100,n,2)}function Q(t,n){return a(t.getUTCFullYear()%1e4,n,4)}function J(){return"+0000"}function G(){return"%"}function K(n){return et=o(n),t.timeFormat=et.format,t.timeParse=et.parse,t.utcFormat=et.utcFormat,t.utcParse=et.utcParse,et}function tt(t){return t.toISOString()}function nt(t){var n=new Date(t);return isNaN(n)?null:n}var et,rt={"-":"",_:" ",0:"0"},it=/^\s*\d+/,ot=/^%/,at=/[\\\^\$\*\+\?\|\[\]\(\)\.\{\}]/g;K({dateTime:"%x, %X",date:"%-m/%-d/%Y",time:"%-I:%M:%S %p",periods:["AM","PM"],days:["Sunday","Monday","Tuesday","Wednesday","Thursday","Friday","Saturday"],shortDays:["Sun","Mon","Tue","Wed","Thu","Fri","Sat"],months:["January","February","March","April","May","June","July","August","September","October","November","December"],shortMonths:["Jan","Feb","Mar","Apr","May","Jun","Jul","Aug","Sep","Oct","Nov","Dec"]});var ut=Date.prototype.toISOString?tt:t.utcFormat("%Y-%m-%dT%H:%M:%S.%LZ"),ct=+new Date("2000-01-01T00:00:00.000Z")?nt:t.utcParse("%Y-%m-%dT%H:%M:%S.%LZ");t.timeFormatDefaultLocale=K,t.timeFormatLocale=o,t.isoFormat=ut,t.isoParse=ct,Object.defineProperty(t,"__esModule",{value:!0})})},{"d3-time":15}],15:[function(t,n,e){!function(t,r){"object"==typeof e&&void 0!==n?r(e):"function"==typeof define&&define.amd?define(["exports"],r):r(t.d3=t.d3||{})}(this,function(t){"use strict";function n(t,e,r,a){function u(n){return t(n=new Date(+n)),n}return u.floor=u,u.ceil=function(n){return t(n=new Date(n-1)),e(n,1),t(n),n},u.round=function(t){var n=u(t),e=u.ceil(t);return t-n<e-t?n:e},u.offset=function(t,n){return e(t=new Date(+t),null==n?1:Math.floor(n)),t},u.range=function(n,r,i){var o=[];if(n=u.ceil(n),i=null==i?1:Math.floor(i),!(n<r&&i>0))return o;do{o.push(new Date(+n))}while(e(n,i),t(n),n<r);return o},u.filter=function(r){return n(function(n){if(n>=n)for(;t(n),!r(n);)n.setTime(n-1)},function(t,n){if(t>=t)for(;--n>=0;)for(;e(t,1),!r(t););})},r&&(u.count=function(n,e){return i.setTime(+n),o.setTime(+e),t(i),t(o),Math.floor(r(i,o))},u.every=function(t){return t=Math.floor(t),isFinite(t)&&t>0?t>1?u.filter(a?function(n){return a(n)%t==0}:function(n){return u.count(0,n)%t==0}):u:null}),u}function e(t){return n(function(n){n.setDate(n.getDate()-(n.getDay()+7-t)%7),n.setHours(0,0,0,0)},function(t,n){t.setDate(t.getDate()+7*n)},function(t,n){return(n-t-(n.getTimezoneOffset()-t.getTimezoneOffset())*c)/s})}function r(t){return n(function(n){n.setUTCDate(n.getUTCDate()-(n.getUTCDay()+7-t)%7),n.setUTCHours(0,0,0,0)},function(t,n){t.setUTCDate(t.getUTCDate()+7*n)},function(t,n){return(n-t)/s})}var i=new Date,o=new Date,a=n(function(){},function(t,n){t.setTime(+t+n)},function(t,n){return n-t});a.every=function(t){return t=Math.floor(t),isFinite(t)&&t>0?t>1?n(function(n){n.setTime(Math.floor(n/t)*t)},function(n,e){n.setTime(+n+e*t)},function(n,e){return(e-n)/t}):a:null};var u=a.range,c=6e4,s=6048e5,f=n(function(t){t.setTime(1e3*Math.floor(t/1e3))},function(t,n){t.setTime(+t+1e3*n)},function(t,n){return(n-t)/1e3},function(t){return t.getUTCSeconds()}),l=f.range,h=n(function(t){t.setTime(Math.floor(t/c)*c)},function(t,n){t.setTime(+t+n*c)},function(t,n){return(n-t)/c},function(t){return t.getMinutes()}),d=h.range,p=n(function(t){var n=t.getTimezoneOffset()*c%36e5;n<0&&(n+=36e5),t.setTime(36e5*Math.floor((+t-n)/36e5)+n)},function(t,n){t.setTime(+t+36e5*n)},function(t,n){return(n-t)/36e5},function(t){return t.getHours()}),_=p.range,y=n(function(t){t.setHours(0,0,0,0)},function(t,n){t.setDate(t.getDate()+n)},function(t,n){return(n-t-(n.getTimezoneOffset()-t.getTimezoneOffset())*c)/864e5},function(t){return t.getDate()-1}),v=y.range,g=e(0),m=e(1),b=e(2),x=e(3),w=e(4),M=e(5),T=e(6),k=g.range,A=m.range,N=b.range,S=x.range,C=w.range,E=M.range,O=T.range,P=n(function(t){t.setDate(1),t.setHours(0,0,0,0)},function(t,n){t.setMonth(t.getMonth()+n)},function(t,n){return n.getMonth()-t.getMonth()+12*(n.getFullYear()-t.getFullYear())},function(t){return t.getMonth()}),D=P.range,q=n(function(t){t.setMonth(0,1),t.setHours(0,0,0,0)},function(t,n){t.setFullYear(t.getFullYear()+n)},function(t,n){return n.getFullYear()-t.getFullYear()},function(t){return t.getFullYear()});q.every=function(t){return isFinite(t=Math.floor(t))&&t>0?n(function(n){n.setFullYear(Math.floor(n.getFullYear()/t)*t),n.setMonth(0,1),n.setHours(0,0,0,0)},function(n,e){n.setFullYear(n.getFullYear()+e*t)}):null};var U=q.range,j=n(function(t){t.setUTCSeconds(0,0)},function(t,n){t.setTime(+t+n*c)},function(t,n){return(n-t)/c},function(t){return t.getUTCMinutes()}),F=j.range,I=n(function(t){t.setUTCMinutes(0,0,0)},function(t,n){t.setTime(+t+36e5*n)},function(t,n){return(n-t)/36e5},function(t){return t.getUTCHours()}),z=I.range,Y=n(function(t){t.setUTCHours(0,0,0,0)},function(t,n){t.setUTCDate(t.getUTCDate()+n)},function(t,n){return(n-t)/864e5},function(t){return t.getUTCDate()-1}),L=Y.range,B=r(0),R=r(1),H=r(2),X=r(3),$=r(4),W=r(5),V=r(6),Z=B.range,Q=R.range,J=H.range,G=X.range,K=$.range,tt=W.range,nt=V.range,et=n(function(t){t.setUTCDate(1),t.setUTCHours(0,0,0,0)},function(t,n){t.setUTCMonth(t.getUTCMonth()+n)},function(t,n){return n.getUTCMonth()-t.getUTCMonth()+12*(n.getUTCFullYear()-t.getUTCFullYear())},function(t){return t.getUTCMonth()}),rt=et.range,it=n(function(t){t.setUTCMonth(0,1),t.setUTCHours(0,0,0,0)},function(t,n){t.setUTCFullYear(t.getUTCFullYear()+n)},function(t,n){return n.getUTCFullYear()-t.getUTCFullYear()},function(t){return t.getUTCFullYear()});it.every=function(t){return isFinite(t=Math.floor(t))&&t>0?n(function(n){n.setUTCFullYear(Math.floor(n.getUTCFullYear()/t)*t),n.setUTCMonth(0,1),n.setUTCHours(0,0,0,0)},function(n,e){n.setUTCFullYear(n.getUTCFullYear()+e*t)}):null};var ot=it.range;t.timeInterval=n,t.timeMillisecond=a,t.timeMilliseconds=u,t.utcMillisecond=a,t.utcMilliseconds=u,t.timeSecond=f,t.timeSeconds=l,t.utcSecond=f,t.utcSeconds=l,t.timeMinute=h,t.timeMinutes=d,t.timeHour=p,t.timeHours=_,t.timeDay=y,t.timeDays=v,t.timeWeek=g,t.timeWeeks=k,t.timeSunday=g,t.timeSundays=k,t.timeMonday=m,t.timeMondays=A,t.timeTuesday=b,t.timeTuesdays=N,t.timeWednesday=x,t.timeWednesdays=S,t.timeThursday=w,t.timeThursdays=C,t.timeFriday=M,t.timeFridays=E,t.timeSaturday=T,t.timeSaturdays=O,t.timeMonth=P,t.timeMonths=D,t.timeYear=q,t.timeYears=U,t.utcMinute=j,t.utcMinutes=F,t.utcHour=I,t.utcHours=z,t.utcDay=Y,t.utcDays=L,t.utcWeek=B,t.utcWeeks=Z,t.utcSunday=B,t.utcSundays=Z,t.utcMonday=R,t.utcMondays=Q,t.utcTuesday=H,t.utcTuesdays=J,t.utcWednesday=X,t.utcWednesdays=G,t.utcThursday=$,t.utcThursdays=K,t.utcFriday=W,t.utcFridays=tt,t.utcSaturday=V,t.utcSaturdays=nt,t.utcMonth=et,t.utcMonths=rt,t.utcYear=it,t.utcYears=ot,Object.defineProperty(t,"__esModule",{value:!0})})},{}],16:[function(t,n,e){!function(t,r){"object"==typeof e&&void 0!==n?r(e):"function"==typeof define&&define.amd?define(["exports"],r):r(t.d3=t.d3||{})}(this,function(t){"use strict";function n(){return v||(b(e),v=m.now()+g)}function e(){v=0}function r(){this._call=this._time=this._next=null}function i(t,n,e){var i=new r;return i.restart(t,n,e),i}function o(){n(),++h;for(var t,e=f;e;)(t=v-e._time)>=0&&e._call.call(null,t),e=e._next;--h}function a(){v=(y=m.now())+g,h=d=0;try{o()}finally{h=0,c(),v=0}}function u(){var t=m.now(),n=t-y;n>_&&(g-=n,y=t)}function c(){for(var t,n,e=f,r=1/0;e;)e._call?(r>e._time&&(r=e._time),t=e,e=e._next):(n=e._next,e._next=null,e=t?t._next=n:f=n);l=t,s(r)}function s(t){if(!h){d&&(d=clearTimeout(d));var n=t-v;n>24?(t<1/0&&(d=setTimeout(a,n)),p&&(p=clearInterval(p))):(p||(y=v,p=setInterval(u,_)),h=1,b(a))}}var f,l,h=0,d=0,p=0,_=1e3,y=0,v=0,g=0,m="object"==typeof performance&&performance.now?performance:Date,b="object"==typeof window&&window.requestAnimationFrame?window.requestAnimationFrame.bind(window):function(t){setTimeout(t,17)};r.prototype=i.prototype={constructor:r,restart:function(t,e,r){if("function"!=typeof t)throw new TypeError("callback is not a function");r=(null==r?n():+r)+(null==e?0:+e),this._next||l===this||(l?l._next=this:f=this,l=this),this._call=t,this._time=r,s()},stop:function(){this._call&&(this._call=null,this._time=1/0,s())}};var x=function(t,n,e){var i=new r;return n=null==n?0:+n,i.restart(function(e){i.stop(),t(e+n)},n,e),i},w=function(t,e,i){var o=new r,a=e;return null==e?(o.restart(t,e,i),o):(e=+e,i=null==i?n():+i,o.restart(function n(r){r+=a,o.restart(n,a+=e,i),t(r)},e,i),o)};t.now=n,t.timer=i,t.timerFlush=o,t.timeout=x,t.interval=w,Object.defineProperty(t,"__esModule",{value:!0})})},{}],17:[function(t,n,e){!function(r,i){"object"==typeof e&&void 0!==n?i(e,t("d3-selection"),t("d3-dispatch"),t("d3-timer"),t("d3-interpolate"),t("d3-color"),t("d3-ease")):"function"==typeof define&&define.amd?define(["exports","d3-selection","d3-dispatch","d3-timer","d3-interpolate","d3-color","d3-ease"],i):i(r.d3=r.d3||{},r.d3,r.d3,r.d3,r.d3,r.d3,r.d3)}(this,function(t,n,e,r,i,o,a){"use strict";function u(t,n){var e=t.__transition;if(!e||!(e=e[n])||e.state>R)throw new Error("too late");return e}function c(t,n){var e=t.__transition;if(!e||!(e=e[n])||e.state>X)throw new Error("too late");return e}function s(t,n){var e=t.__transition;if(!e||!(e=e[n]))throw new Error("too late");return e}function f(t,n,e){function i(t){e.state=H,e.timer.restart(o,e.delay,e.time),e.delay<=t&&o(t-e.delay)}function o(i){var f,l,h,d;if(e.state!==H)return u();for(f in s)if((d=s[f]).name===e.name){if(d.state===$)return r.timeout(o);d.state===W?(d.state=Z,d.timer.stop(),d.on.call("interrupt",t,t.__data__,d.index,d.group),delete s[f]):+f<n&&(d.state=Z,d.timer.stop(),delete s[f])}if(r.timeout(function(){e.state===$&&(e.state=W,e.timer.restart(a,e.delay,e.time),a(i))}),e.state=X,e.on.call("start",t,t.__data__,e.index,e.group),e.state===X){for(e.state=$,c=new Array(h=e.tween.length),f=0,l=-1;f<h;++f)(d=e.tween[f].value.call(t,t.__data__,e.index,e.group))&&(c[++l]=d);c.length=l+1}}function a(n){for(var r=n<e.duration?e.ease.call(null,n/e.duration):(e.timer.restart(u),e.state=V,1),i=-1,o=c.length;++i<o;)c[i].call(null,r);e.state===V&&(e.on.call("end",t,t.__data__,e.index,e.group),u())}function u(){e.state=Z,e.timer.stop(),delete s[n];for(var r in s)return;delete t.__transition}var c,s=t.__transition;s[n]=e,e.timer=r.timer(i,0,e.time)}function l(t,n){var e,r;return function(){var i=c(this,t),o=i.tween;if(o!==e)for(var a=0,u=(r=e=o).length;a<u;++a)if(r[a].name===n){(r=r.slice()).splice(a,1);break}i.tween=r}}function h(t,n,e){var r,i;if("function"!=typeof e)throw new Error;return function(){var o=c(this,t),a=o.tween;if(a!==r){i=(r=a).slice();for(var u={name:n,value:e},s=0,f=i.length;s<f;++s)if(i[s].name===n){i[s]=u;break}s===f&&i.push(u)}o.tween=i}}function d(t,n,e){var r=t._id;return t.each(function(){var t=c(this,r);(t.value||(t.value={}))[n]=e.apply(this,arguments)}),function(t){return s(t,r).value[n]}}function p(t){return function(){this.removeAttribute(t)}}function _(t){return function(){this.removeAttributeNS(t.space,t.local)}}function y(t,n,e){var r,i;return function(){var o=this.getAttribute(t);return o===e?null:o===r?i:i=n(r=o,e)}}function v(t,n,e){var r,i;return function(){var o=this.getAttributeNS(t.space,t.local);return o===e?null:o===r?i:i=n(r=o,e)}}function g(t,n,e){var r,i,o;return function(){var a,u=e(this);{if(null!=u)return a=this.getAttribute(t),a===u?null:a===r&&u===i?o:o=n(r=a,i=u);this.removeAttribute(t)}}}function m(t,n,e){var r,i,o;return function(){var a,u=e(this);{if(null!=u)return a=this.getAttributeNS(t.space,t.local),a===u?null:a===r&&u===i?o:o=n(r=a,i=u);this.removeAttributeNS(t.space,t.local)}}}function b(t,n){function e(){var e=this,r=n.apply(e,arguments);return r&&function(n){e.setAttributeNS(t.space,t.local,r(n))}}return e._value=n,e}function x(t,n){function e(){var e=this,r=n.apply(e,arguments);return r&&function(n){e.setAttribute(t,r(n))}}return e._value=n,e}function w(t,n){return function(){u(this,t).delay=+n.apply(this,arguments)}}function M(t,n){return n=+n,function(){u(this,t).delay=n}}function T(t,n){return function(){c(this,t).duration=+n.apply(this,arguments)}}function k(t,n){return n=+n,function(){c(this,t).duration=n}}function A(t,n){if("function"!=typeof n)throw new Error;return function(){c(this,t).ease=n}}function N(t){return(t+"").trim().split(/^|\s+/).every(function(t){var n=t.indexOf(".");return n>=0&&(t=t.slice(0,n)),!t||"start"===t})}function S(t,n,e){var r,i,o=N(n)?u:c;return function(){var a=o(this,t),u=a.on;u!==r&&(i=(r=u).copy()).on(n,e),a.on=i}}function C(t){return function(){var n=this.parentNode;for(var e in this.__transition)if(+e!==t)return;n&&n.removeChild(this)}}function E(t,e){var r,i,o;return function(){var a=n.style(this,t),u=(this.style.removeProperty(t),n.style(this,t));return a===u?null:a===r&&u===i?o:o=e(r=a,i=u)}}function O(t){return function(){this.style.removeProperty(t)}}function P(t,e,r){var i,o;return function(){var a=n.style(this,t);return a===r?null:a===i?o:o=e(i=a,r)}}function D(t,e,r){var i,o,a;return function(){var u=n.style(this,t),c=r(this);return null==c&&(this.style.removeProperty(t),c=n.style(this,t)),u===c?null:u===i&&c===o?a:a=e(i=u,o=c)}}function q(t,n,e){function r(){var r=this,i=n.apply(r,arguments);return i&&function(n){r.style.setProperty(t,i(n),e)}}return r._value=n,r}function U(t){return function(){this.textContent=t}}function j(t){return function(){var n=t(this);this.textContent=null==n?"":n}}function F(t,n,e,r){this._groups=t,this._parents=n,this._name=e,this._id=r}function I(t){return n.selection().transition(t)}function z(){return++gt}function Y(t,n){for(var e;!(e=t.__transition)||!(e=e[n]);)if(!(t=t.parentNode))return bt.time=r.now(),bt;return e}var L=e.dispatch("start","end","interrupt"),B=[],R=0,H=1,X=2,$=3,W=4,V=5,Z=6,Q=function(t,n,e,r,i,o){var a=t.__transition;if(a){if(e in a)return}else t.__transition={};f(t,e,{name:n,index:r,group:i,on:L,tween:B,time:o.time,delay:o.delay,duration:o.duration,ease:o.ease,timer:null,state:R})},J=function(t,n){var e,r,i,o=t.__transition,a=!0;if(o){n=null==n?null:n+"";for(i in o)(e=o[i]).name===n?(r=e.state>X&&e.state<V,e.state=Z,e.timer.stop(),r&&e.on.call("interrupt",t,t.__data__,e.index,e.group),delete o[i]):a=!1;a&&delete t.__transition}},G=function(t){return this.each(function(){J(this,t)})},K=function(t,n){var e=this._id;if(t+="",arguments.length<2){for(var r,i=s(this.node(),e).tween,o=0,a=i.length;o<a;++o)if((r=i[o]).name===t)return r.value;return null}return this.each((null==n?l:h)(e,t,n))},tt=function(t,n){var e;return("number"==typeof n?i.interpolateNumber:n instanceof o.color?i.interpolateRgb:(e=o.color(n))?(n=e,i.interpolateRgb):i.interpolateString)(t,n)},nt=function(t,e){var r=n.namespace(t),o="transform"===r?i.interpolateTransformSvg:tt;return this.attrTween(t,"function"==typeof e?(r.local?m:g)(r,o,d(this,"attr."+t,e)):null==e?(r.local?_:p)(r):(r.local?v:y)(r,o,e+""))},et=function(t,e){var r="attr."+t;if(arguments.length<2)return(r=this.tween(r))&&r._value;if(null==e)return this.tween(r,null);if("function"!=typeof e)throw new Error;var i=n.namespace(t);return this.tween(r,(i.local?b:x)(i,e))},rt=function(t){var n=this._id;return arguments.length?this.each(("function"==typeof t?w:M)(n,t)):s(this.node(),n).delay},it=function(t){var n=this._id;return arguments.length?this.each(("function"==typeof t?T:k)(n,t)):s(this.node(),n).duration},ot=function(t){var n=this._id;return arguments.length?this.each(A(n,t)):s(this.node(),n).ease},at=function(t){"function"!=typeof t&&(t=n.matcher(t));for(var e=this._groups,r=e.length,i=new Array(r),o=0;o<r;++o)for(var a,u=e[o],c=u.length,s=i[o]=[],f=0;f<c;++f)(a=u[f])&&t.call(a,a.__data__,f,u)&&s.push(a);return new F(i,this._parents,this._name,this._id)},ut=function(t){if(t._id!==this._id)throw new Error;for(var n=this._groups,e=t._groups,r=n.length,i=e.length,o=Math.min(r,i),a=new Array(r),u=0;u<o;++u)for(var c,s=n[u],f=e[u],l=s.length,h=a[u]=new Array(l),d=0;d<l;++d)(c=s[d]||f[d])&&(h[d]=c);for(;u<r;++u)a[u]=n[u];return new F(a,this._parents,this._name,this._id)},ct=function(t,n){var e=this._id;return arguments.length<2?s(this.node(),e).on.on(t):this.each(S(e,t,n))},st=function(){return this.on("end.remove",C(this._id))},ft=function(t){var e=this._name,r=this._id;"function"!=typeof t&&(t=n.selector(t));for(var i=this._groups,o=i.length,a=new Array(o),u=0;u<o;++u)for(var c,f,l=i[u],h=l.length,d=a[u]=new Array(h),p=0;p<h;++p)(c=l[p])&&(f=t.call(c,c.__data__,p,l))&&("__data__"in c&&(f.__data__=c.__data__),d[p]=f,Q(d[p],e,r,p,d,s(c,r)));return new F(a,this._parents,e,r)},lt=function(t){var e=this._name,r=this._id;"function"!=typeof t&&(t=n.selectorAll(t));for(var i=this._groups,o=i.length,a=[],u=[],c=0;c<o;++c)for(var f,l=i[c],h=l.length,d=0;d<h;++d)if(f=l[d]){for(var p,_=t.call(f,f.__data__,d,l),y=s(f,r),v=0,g=_.length;v<g;++v)(p=_[v])&&Q(p,e,r,v,_,y);a.push(_),u.push(f)}return new F(a,u,e,r)},ht=n.selection.prototype.constructor,dt=function(){return new ht(this._groups,this._parents)},pt=function(t,n,e){var r="transform"==(t+="")?i.interpolateTransformCss:tt;return null==n?this.styleTween(t,E(t,r)).on("end.style."+t,O(t)):this.styleTween(t,"function"==typeof n?D(t,r,d(this,"style."+t,n)):P(t,r,n+""),e)},_t=function(t,n,e){var r="style."+(t+="");if(arguments.length<2)return(r=this.tween(r))&&r._value;if(null==n)return this.tween(r,null);if("function"!=typeof n)throw new Error;return this.tween(r,q(t,n,null==e?"":e))},yt=function(t){return this.tween("text","function"==typeof t?j(d(this,"text",t)):U(null==t?"":t+""))},vt=function(){for(var t=this._name,n=this._id,e=z(),r=this._groups,i=r.length,o=0;o<i;++o)for(var a,u=r[o],c=u.length,f=0;f<c;++f)if(a=u[f]){var l=s(a,n);Q(a,t,e,f,u,{time:l.time+l.delay+l.duration,delay:0,duration:l.duration,ease:l.ease})}return new F(r,this._parents,t,e)},gt=0,mt=n.selection.prototype;F.prototype=I.prototype={constructor:F,select:ft,selectAll:lt,filter:at,merge:ut,selection:dt,transition:vt,call:mt.call,nodes:mt.nodes,node:mt.node,size:mt.size,empty:mt.empty,each:mt.each,on:ct,attr:nt,attrTween:et,style:pt,styleTween:_t,text:yt,remove:st,tween:K,delay:rt,duration:it,ease:ot};var bt={time:null,delay:0,duration:250,ease:a.easeCubicInOut},xt=function(t){var n,e;t instanceof F?(n=t._id,t=t._name):(n=z(),(e=bt).time=r.now(),t=null==t?null:t+"");for(var i=this._groups,o=i.length,a=0;a<o;++a)for(var u,c=i[a],s=c.length,f=0;f<s;++f)(u=c[f])&&Q(u,t,n,f,c,e||Y(u,n));return new F(i,this._parents,t,n)};n.selection.prototype.interrupt=G,n.selection.prototype.transition=xt;var wt=[null],Mt=function(t,n){var e,r,i=t.__transition;if(i){n=null==n?null:n+"";for(r in i)if((e=i[r]).state>H&&e.name===n)return new F([[t]],wt,n,+r)}return null};t.transition=I,t.active=Mt,t.interrupt=J,Object.defineProperty(t,"__esModule",{value:!0})})},{"d3-color":4,"d3-dispatch":5,"d3-ease":6,"d3-interpolate":9,"d3-selection":12,"d3-timer":16}]},{},[1]);
