+++ javascript/.gitignore
@@ -1,0 +1,3 @@
+build/
+node_modules
+npm-debug.log
