+++ javascript/global/htmlwidgets.js
@@ -1,0 +1,1 @@
+export default window.HTMLWidgets;
